### 2025-01-20 #7699 @editorialbot
[PRE REVIEW]: SSAPy - Space Situational Awareness for Python
**Submitting author:** @astrophpeter (<a href="http://orcid.org/0000-0002-1052-6749">Peter McGill</a>)
**Repository:** https://github.com/LLNL/SSAPy
**Branch with paper.md** (empty if default branch):
**Version:** v0.7.0
