### 2025-01-20 #7699 @warrickball
Hi @astrophpeter — Thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!

In the meantime, could you also try changing the affiliation list for Joshua E. Myers from `[1, 2]` to `"1, 2"`, then rerun our paper pipeline by commenting here with `@editorialbot generate pdf`? If that does fix the error, as the pipeline suggests, then there's an error in our documentation that I'll fix.

I also had a brief look at SSAPy's documentation and noticed that the [API reference](https://software.llnl.gov/SSAPy/reference.html) doesn't seem to have any content beyond the top-level modules. I presume that indicates a bug in the documentation pipeline.
