### 2025-01-22 74100f9
Moving plot padding to find_smallest_bounding_cube
Automatic plot limits were still not correct in every situation. Moving the pad calculation into .compute seems to fix.
