+++ ssapy/plotUtils.py
@@ -214,59 +214,30 @@ def check_numpy_array(variable):
 
 
 def orbit_plot(r, t=[], title='', figsize=(7, 7), save_path=False, frame="gcrf", show=False):
-    """
-    Plot 2D and 3D projections of the orbit of one or more objects.
+    input_type = check_numpy_array(r)
+    if input_type == "numpy array":
+        num_orbits = 1
+        r = [r]
 
-    Parameters:
-    ----------
-    r : (n,3) or list of [(n,3), ..., (n,3)] array_like
-        Position of orbiting object(s) in meters. If a list is provided, each element represents the orbit of a different object.
-    t : array_like, optional
-        Time array corresponding to the position vectors. Required for frames "itrf", "lunar", or "lunar fixed".
-    title : str, optional
-        Title of the plot. Default is an empty string.
-    figsize : tuple of int, optional
-        Size of the figure (width, height) in inches. Default is (7, 7).
-    save_path : str, optional
-        Path to save the generated plot. If not provided, the plot will not be saved. Default is False.
-    frame : str, optional
-        The coordinate frame for the plot. Accepted values are "gcrf", "itrf", "lunar", "lunar fixed".
-    show : bool, optional
-        Whether to display the plot. Default is False.
+    if input_type == "list of numpy array":
+        num_orbits = np.shape(r)[0]
 
-    Returns:
-    -------
-    fig : matplotlib.figure.Figure
-        The figure object containing the plot.
-    ax : list of matplotlib.axes.Axes
-        The list of axes objects used in the plot. Includes 2D axes (ax1, ax2, ax3) and 3D axis (ax4).
+    fig = plt.figure(dpi=100, figsize=figsize, facecolor='black')
+    ax1 = fig.add_subplot(2, 2, 1)
+    ax2 = fig.add_subplot(2, 2, 2)
+    ax3 = fig.add_subplot(2, 2, 3)
+    ax4 = fig.add_subplot(2, 2, 4, projection='3d')
+    
+    bounds = {"lower": np.array([np.inf, np.inf, np.inf]), "upper": np.array([-np.inf, -np.inf, -np.inf])}
 
-    Notes:
-    ------
-    - The function creates a set of subplots showing the projection of the orbit data in 2D (x-y, x-z, y-z) and 3D (x-y-z).
-    - For the "lunar" and "lunar fixed" frames, the plot also includes the positions of Lagrange points.
-    - If a list of orbits is provided, each orbit is plotted with a different color.
-    - The plot background is set to black, and axis labels and ticks are set to white.
-
-    Example usage:
-    ----------
-    import numpy as np
-    from your_module import orbit_plot
-
-    # Example data
-    r = np.array([[1e7, 2e7, 3e7], [4e7, 5e7, 6e7]])  # Replace with actual data
-    t = np.linspace(0, 10, len(r))  # Replace with actual time data
-
-    fig, axes = orbit_plot(r, t, title='Orbit Plot', frame='gcrf', show=True)
-    """
-
-    def _make_scatter(fig, ax1, ax2, ax3, ax4, r, t, bounds, title='', orbit_index='', num_orbits=1, frame=False):
+    for orbit_index, xyz in enumerate(r):
         if np.size(t) < 1:
             if frame in ["itrf", "lunar", "lunar_fixed"]:
-                raise("Need to provide t for itrf, lunar or lunar fixed frames")
+                raise ValueError("Need to provide t for itrf, lunar or lunar fixed frames")
[…]
