### 2025-01-22 a3fe213
added pad parameter to find_smallest_bounding_cube
allowing pad to the bounds of a bounding cube.
