### 2025-02-13 820bda0
doc strings and unequal orbits
Added doc strings to the functions and added the ability for orbit_plot to correctly handle plotting multiple unequal length orbits.
