+++ ssapy/plotUtils.py
@@ -204,23 +204,58 @@ def groundTrackVideo(r, time):
     ipv.show()
 
 
-def check_numpy_array(variable):
+def check_numpy_array(variable: Union[np.ndarray, list]) -> str:
+    """
+    Checks if the input variable is a NumPy array, a list of NumPy arrays, or neither.
+
+    Parameters
+    ----------
+    variable : Union[np.ndarray, list]
+        The variable to check. It can either be a NumPy array or a list of NumPy arrays.
+
+    Returns
+    -------
+    str
+        Returns a string indicating the type of the variable:
+        - "numpy array" if the variable is a single NumPy array,
+        - "list of numpy array" if it is a list of NumPy arrays,
+        - "not numpy" if it is neither.
+    """
     if isinstance(variable, np.ndarray):
         return "numpy array"
-    elif isinstance(variable, list) and all(isinstance(item, np.ndarray) for item in variable):
-        return "list of numpy array"
+    elif isinstance(variable, list):
+        if len(variable) == 0:  # Handle empty list explicitly
+            return "not numpy"
+        elif all(isinstance(item, np.ndarray) for item in variable):
+            return "list of numpy array"
+    return "not numpy"
+
+
+def check_type(t):
+    if t is None:
+        return None
+    elif isinstance(t, list):
+        # Check if each element is a list or array
+        if all(isinstance(item, (list, np.ndarray)) for item in t):
+            return "List of arrays"
+        else:
+            return "List of non-arrays"
+    elif isinstance(t, (Time, np.ndarray)):
+        return "Single array or list"
     else:
-        return "not numpy"
+        return "Not a list or array"
+    
 
-
-def orbit_plot(r, t=[], title='', figsize=(7, 7), save_path=False, frame="gcrf", show=False):
+def orbit_plot(r, t=None, title='', figsize=(7, 7), save_path=False, frame="gcrf", show=False):
     input_type = check_numpy_array(r)
+    t_type = check_type(t)
+    
     if input_type == "numpy array":
         num_orbits = 1
         r = [r]
 
     if input_type == "list of numpy array":
-        num_orbits = np.shape(r)[0]
+        num_orbits = len(r)
 
     fig = plt.figure(dpi=100, figsize=figsize, facecolor='black')
     ax1 = fig.add_subplot(2, 2, 1)
@@ -230,13 +265,49 @@ def orbit_plot(r, t=[], title='', figsize=(7, 7), save_path=False, frame="gcrf",
     
     bounds = {"lower": np.array([np.inf, np.inf, np.inf]), "upper": np.array([-np.inf, -np.inf, -np.inf])}
 
-    for orbit_index, xyz in enumerate(r):
-        if np.size(t) < 1:
-            if frame in ["itrf", "lunar", "lunar_fixed"]:
-                raise ValueError("Need to provide t for itrf, lunar or lunar fixed frames")
-            r_moon = np.atleast_2d(get_body("moon").position(Time("2000-1-1")))
+    # Check if all arrays in `r` are the same shape
+    same_shape = all(np.shape(arr)[0] == np.shape(r[0]) for arr in r)
+    for orbit_index in range(num_orbits):
+        xyz = r[orbit_index]
+        
+        if t_type is None:
+            if frame == "gcrf":
+                r_moon = np.atleast_2d(get_body("moon").position(Time("2000-1-1")))
+            else:
+                raise ValueError("Need to provide t or list of t for each orbit in itrf, lunar or lunar fixed frames")
         else:
-            r_moon = get_body("moon").position(t).T
+            if frame == "gcrf":
[…]
