### 2025-02-27 7b849bb
doc string typo
function can take n,3 for all celestial objects in calc_M_v
