+++ .readthedocs.yaml
@@ -8,7 +8,7 @@ version: 2
 build:
   os: ubuntu-24.04
   tools:
-    python: "3.13"
+    python: "3.8"
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
