+++ docs/requirements.txt
@@ -1,4 +1,4 @@
-sphinx==7.2.3
+sphinx
 sphinx-copybutton
 sphinx_rtd_theme
 sphinxcontrib-bibtex
