+++ docs/requirements.txt
@@ -1,5 +1,6 @@
 sphinx
 sphinx-copybutton
+sphinx-tabs
 sphinx_rtd_theme
 sphinxcontrib-bibtex
 nbsphinx
