+++ docs/requirements.txt
@@ -1,5 +1,6 @@
 sphinx
 sphinx-copybutton
+myst-parser
 sphinx-tabs
 sphinx_rtd_theme
 sphinxcontrib-bibtex
