+++ docs/source/conf.py
@@ -31,8 +31,8 @@ extensions = [
     'sphinx.ext.linkcode',
     'sphinx_copybutton',
     'sphinx_tabs.tabs',
-    # 'sphinx_automodapi.automodapi',
-    # 'sphinx_automodapi.smart_resolver',
+    'sphinx_automodapi.automodapi',
+    'sphinx_automodapi.smart_resolver',
     'myst_parser',
     "sphinx_rtd_theme",
     "sphinx.ext.mathjax",
