+++ docs/source/citations.bib
@@ -1,28 +1,0 @@
-@INPROCEEDINGS{2023amos.conf..208Y,
-       author = {{Yeager}, T. and {Pruett}, K. and {Schneider}, M.},
-        title = "{Long-term N-body Stability in Cislunar Space}",
-     keywords = {cislunar, n-body, dynamics, orbit, stability},
-    booktitle = {Proceedings of the Advanced Maui Optical and Space Surveillance (AMOS) Technologies Conference},
-         year = 2023,
-       editor = {{Ryan}, S.},
-        month = sep,
-          eid = {208},
-        pages = {208},
-          url = https://amostech.com/TechnicalPapers/2023/Poster/Yeager.pdf,
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023amos.conf..208Y},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@unpublished{2023amos.conf.poster..Yeager,
-       author = {{Yeager}, Travis and {Pruett}, Kerianne and {Schneider}, Michael},
-        title = "{Long-term N-body Stability in Cislunar Space}",
-         year = 2023,
-         note = {Poster presented at the Advanced Maui Optical and Space Surveillance (AMOS) Technologies Conference},
-}
-
-@unpublished{2022csc.conf.poster..Yeager,
-       author = {{Yeager}, Travis and {Pruett}, Kerianne and {Schneider}, Michael},
-        title = "{Unaided Dynamical Orbit Stability in the Cislunar Regime}",
-         year = 2022,
-         note = {Poster presented at the Cislunar Security Conference},
-}
