+++ docs/source/reference.rst
@@ -1,121 +1,0 @@
-===
-API
-===
-
-Acceleration
----------
-
-.. automodule:: ssapy.accel
-   :members:
-
-
-Body
----------
-
-.. automodule:: ssapy.body
-   :members:
-
-
-Compute
----------
-
-.. automodule:: ssapy.compute
-   :members:
-
-
-Constants
----------
-
-.. automodule:: ssapy.constants
-   :members:
-
-
-Correlate Tracks
----------
-
-.. automodule:: ssapy.correlate_tracks
-   :members:
-
-
-Ellipsoid
----------
-
-.. automodule:: ssapy.ellipsoid
-   :members:
-
-
-Gravity
----------
-
-.. automodule:: ssapy.gravity
-   :members:
-
-
-Io
----------
-
-.. automodule:: ssapy.io
-   :members:
-
-
-Linker
----------
-
-.. automodule:: ssapy.linker
-   :members:
-
-
-Orbit Solver
----------
-
-.. automodule:: ssapy.orbit_solver
-   :members:
-
-
-Orbit
----------
-
-.. automodule:: ssapy.orbit
-   :members:
-
-
-Particles
----------
-
-.. automodule:: ssapy.particles
-   :members:
-
-
-Plot Utilities
----------
-
-.. automodule:: ssapy.plotUtils
-   :members:
-
-
-Propagator
----------
-
-.. automodule:: ssapy.propagator
-   :members:
-
-
-Position and Velocity Sampler
----------
-
-.. automodule:: ssapy.rvsampler
-   :members:
-
-
-Simple
----------
-
-.. automodule:: ssapy.simple
-   :members:
-
-
-Utilities
----------
-
-.. automodule:: ssapy.simple
-   :members:
