+++ docs/source/refs.bib
@@ -17,7 +17,7 @@ year = {2024}
         month = sep,
           eid = {208},
         pages = {208},
-          url = https://amostech.com/TechnicalPapers/2023/Poster/Yeager.pdf,
+          url = {https://amostech.com/TechnicalPapers/2023/Poster/Yeager.pdf},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2023amos.conf..208Y},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
