+++ docs/source/conf.py
@@ -10,7 +10,7 @@ import subprocess
 
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../ssapy/'))
+# sys.path.insert(0, os.path.abspath('../../ssapy/'))
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
