+++ docs/source/refs.bib
@@ -1,10 +1,10 @@
 @inproceedings{ssapyprep,
-title = {{SSAPy} - Space Situational Awareness for Python},
-author = {Meyers, Joshua E. and Schneider, Michael D. and Ebert, Julia T. and Schlafly, Edward F. and Yeager, Travis and Perloff, Alexx and Merl, Daniel and Lifset, Noah and Berstein, Jason and Dawson, William A. and Golovich, Nathan and Higgins, Denvir and McGill, Peter and Miller, Caleb and Pruett, Kerianne},
-booktitle = {The Journal of Open Source Software},
-publisher = {The Open Journal},
-note = {Submitted to},
-year = {2024}
+    title = {{SSAPy} - Space Situational Awareness for Python},
+    author = {Meyers, Joshua E. and Schneider, Michael D. and Ebert, Julia T. and Schlafly, Edward F. and Yeager, Travis and Perloff, Alexx and Merl, Daniel and Lifset, Noah and Berstein, Jason and Dawson, William A. and Golovich, Nathan and Higgins, Denvir and McGill, Peter and Miller, Caleb and Pruett, Kerianne},
+    booktitle = {The Journal of Open Source Software},
+    publisher = {The Open Journal},
+    note = {Submitted to},
+    year = {2024}
 }
 
 @INPROCEEDINGS{2023amos.conf..208Y,
