+++ docs/requirements.txt
@@ -10,11 +10,8 @@ numpy
 astropy
 jupyter
 asdf
-git 
+python-git 
 git-lfs 
-python3 
-python3-distutils 
-python3-venv 
 graphviz
 scipy
 matplotlib
