+++ docs/source/conf.py
@@ -66,23 +66,13 @@ extensions = [
 #         githash, relpath, lineno
 #     )
 
-# def linkcode_resolve(domain, info):
-#     """Create a link to the source code on GitHub."""
-#     if domain != 'py':
-#         return None  # Only resolve links for Python code
-    
-#     # Get the object path from `info`
-#     module_name = info['module']
-#     object_name = info['fullname']
-
-#     # Construct the GitHub URL
-#     github_url = f"http://github.com/LLNL/SSAPy/blob/{module_name}.py"
-    
-#     # If the object is a function or class, add the location in the file
-#     if object_name:
-#         github_url += f"#L{inspect.stack()[0].lineno}"
-    
-#     return github_url
+def linkcode_resolve(domain, info):
+    if domain != 'py':
+        return None
+    if not info['module']:
+        return None
+    filename = info['module'].replace('.', '/')
+    return "https://github.com/LLNL/SSAPy/tree/main/ssapy/%s.py" % filename
 
 autosummary_generate = False
 numpydoc_show_class_members = False
