+++ docs/source/conf.py
@@ -11,7 +11,6 @@ import subprocess
 import os
 import sys
 sys.path.insert(0, os.path.abspath('../../')) 
-autodoc_mock_imports = ['ssapy']
 
 
 # -- Project information -----------------------------------------------------
