+++ docs/source/conf.py
@@ -10,7 +10,31 @@ import subprocess
 
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../')) 
+sys.path.insert(0, os.path.abspath('../../ssapy/')) 
+
+from unittest.mock import Mock
+MOCK_MODULES = [
+    'ssapy.accel',
+    'ssapy.body',
+    'ssapy.compute',
+    'ssapy.constants',
+    'ssapy.correlate_tracks',
+    'ssapy.ellipsoid',
+    'ssapy.gravity',
+    'ssapy.io',
+    'ssapy.linker',
+    'ssapy.orbit_solver',
+    'ssapy.orbit',
+    'ssapy.particles',
+    'ssapy.plotUtils',
+    'ssapy.propagator',
+    'ssapy.rvsampler',
+    'ssapy.simple',
+    'ssapy.utils',
+    'ssapy._ssapy',  # Add this since it was part of the original error
+]
+sys.modules.update((mod_name, Mock()) for mod_name in MOCK_MODULES)
+
 
 
 # -- Project information -----------------------------------------------------
@@ -38,32 +62,9 @@ extensions = [
     'sphinx_rtd_theme',
     'sphinx.ext.mathjax',
     'sphinxcontrib.bibtex',
-    'sphinx.ext.napoleon'
+    'sphinx.ext.napoleon',
+    "numpy": ("https://numpy.org/doc/stable/", None)
 ]
-
-
-# def linkcode_resolve(domain, info):
-#     if domain != "py":
-#         return None
-#     if not info["module"]:
-#         return None
-#     mod = importlib.import_module(info["module"])
-#     modpath = [p for p in sys.path if mod.__file__.startswith(p)]
-#     if len(modpath) < 1:
-#         raise RuntimeError("Cannot deduce module path")
-#     modpath = modpath[0]
-#     obj = reduce(getattr, [mod] + info["fullname"].split("."))
-#     try:
-#         path = inspect.getsourcefile(obj)
-#         relpath = path[len(modpath) + 1:]
-#         relpath = relpath.split(".egg/")[-1]
-#         _, lineno = inspect.getsourcelines(obj)
-#     except TypeError:
-#         # skip property or other type that inspect doesn't like
-#         return None
-#     return "http://github.com/LLNL/SSAPy/blob/{}/{}#L{}".format(
-#         githash, relpath, lineno
-#     )
 
 def linkcode_resolve(domain, info):
     if domain != 'py':
@@ -73,7 +74,7 @@ def linkcode_resolve(domain, info):
     filename = info['module'].replace('.', '/')
     return "https://github.com/LLNL/SSAPy/tree/main/ssapy/%s.py" % filename
 
-autosummary_generate = False
+autosummary_generate = True
 numpydoc_show_class_members = False
 sphinx_tabs_valid_builders = ['linkcheck']
 source_suffix = ['.rst', '.md']
