+++ docs/source/index.rst
@@ -3,7 +3,7 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-.. include:: ../../README.rst
+.. include:: ../README.rst
 
 .. toctree::
    :titlesonly:
