### 2025-03-25 8c99dcc
forcing ssapy git install
