+++ .readthedocs.yaml
@@ -6,9 +6,16 @@ version: 2
 
 # Set the OS, Python version, and other tools you might need
 build:
-  os: ubuntu-24.04
-  tools:
-    python: "3.8"
+   os: ubuntu-24.04
+   tools:
+   python: "3.8"
+   apt_packages:
+      - graphviz
+   commands:
+      - git lfs install
+      - git lfs pull
+      - python3 setup.py build
+      - python3 setup.py install
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
@@ -20,5 +27,7 @@ sphinx:
 python:
    install:
    - requirements: docs/requirements.txt
+
+     
         
  
