+++ .readthedocs.yaml
@@ -14,8 +14,6 @@ build:
    commands:
       - git lfs install
       - git lfs pull
-      - python3 setup.py build
-      - python3 setup.py install
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
