+++ .readthedocs.yaml
@@ -1,3 +1,4 @@
+yaml
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
@@ -6,26 +7,22 @@ version: 2
 
 # Set the OS, Python version, and other tools you might need
 build:
-   os: ubuntu-24.04
-   tools:
-   python: "3.8"
-   apt_packages:
-      - graphviz
-   commands:
-      - git lfs install
-      - git lfs pull
+  os: ubuntu-24.04
+  tools:
+    python: "3.8"
+  apt_packages:
+    - graphviz
+  commands:
+    - git lfs install
+    - git lfs pull
+    - python3 setup.py build
+    - python3 setup.py install
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
-   configuration: docs/source/conf.py
+  configuration: docs/source/conf.py
 
-# Optionally, but recommended,
-# declare the Python requirements required to build your documentation
-# See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
+# Declare the Python requirements required to build your documentation
 python:
-   install:
-   - requirements: docs/requirements.txt
-
-     
-        
- 
+  install:
+    - requirements: docs/requirements.txt
