+++ .readthedocs.yaml
@@ -1,4 +1,3 @@
-yaml
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
@@ -26,3 +25,4 @@ sphinx:
 python:
   install:
     - requirements: docs/requirements.txt
+
