+++ .readthedocs.yaml
@@ -11,16 +11,17 @@ build:
     python: "3.8"
   apt_packages:
     - graphviz
-    - git 
-    - git-lfs 
-    - python3
   commands:
+    - curl -s https://packagecloud.io/install/repositories/github/git-lfs/script.deb.sh | sudo bash
+    - sudo apt-get install git-lfs
+    - git lfs install
+    - git lfs pull
     - git submodule update --init --recursive
     - git lfs install
     - git lfs pull
     - python3 setup.py build
     - python3 setup.py install
-
+  
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/source/conf.py
