### 2025-03-25 9473a84
better install
