+++ .readthedocs.yaml
@@ -11,16 +11,6 @@ build:
     python: "3.8"
   apt_packages:
     - graphviz
-  commands:
-    - wget https://github.com/git-lfs/git-lfs/releases/download/v3.2.0/git-lfs-linux-amd64-v3.2.0.tar.gz
-    - tar xvf git-lfs-linux-amd64-v3.2.0.tar.gz
-    - cd git-lfs-3.2.0/
-    - ./install.sh
-    - git lfs install
-    - git lfs pull
-    - git submodule update --init --recursive
-    - python3 setup.py build
-    - python3 setup.py install
   
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
