+++ .readthedocs.yaml
@@ -11,6 +11,7 @@ build:
     python: "3.8"
   apt_packages:
     - graphviz
+    - cmake
   commands:
     - python3 setup.py build
     - python3 setup.py install
