+++ .readthedocs.yaml
@@ -16,11 +16,13 @@ build:
     - curl -L https://github.com/Kitware/CMake/releases/download/v3.27.0/cmake-3.27.0-linux-x86_64.tar.gz -o cmake.tar.gz
     - tar -xzf cmake.tar.gz
     # Create a local bin directory and move cmake there
-    - mkdir -p ~/bin
-    - mv cmake-3.27.0-linux-x86_64/bin/cmake ~/bin/cmake
+    - mkdir -p $HOME/bin
+    - mv cmake-3.27.0-linux-x86_64/bin/cmake $HOME/bin/cmake
     # Add the local bin directory to PATH
     - export PATH="$HOME/bin:$PATH"
-    # Verify cmake installation
+    # Debug PATH and verify CMake installation
+    - echo $PATH
+    - ls -l $HOME/bin
     - cmake --version
     # Build and install the project
     - python3 setup.py build
