### 2025-03-25 eaa21ee
using conda but at what cost
