+++ .readthedocs.yaml
@@ -12,13 +12,7 @@ build:
   apt_packages:
     - graphviz
   commands:
-    - curl -L https://github.com/Kitware/CMake/releases/download/v3.27.0/cmake-3.27.0-linux-x86_64.tar.gz -o cmake.tar.gz
-    - tar -xzf cmake.tar.gz
-    - mkdir -p $HOME/bin
-    - mv cmake-3.27.0-linux-x86_64/bin/cmake $HOME/bin/cmake
-    - export PATH="$HOME/bin:$PATH"
-    - echo PATH
-    - which cmake
+    - conda install -c conda-forge cmake
     - cmake --version
     - python3 setup.py build
     - python3 setup.py install
