### 2025-03-25 de5df0b
moving cmake install to conf
