+++ .readthedocs.yaml
@@ -11,6 +11,7 @@ build:
     python: "3.8"
   apt_packages:
     - graphviz
+    - cmake
 #   commands:
 #     - conda install -c conda-forge cmake
 #     - cmake --version
