### 2025-03-25 f3c64f3
testing which functions are redundant
