+++ .readthedocs.yaml
@@ -12,12 +12,7 @@ build:
   apt_packages:
     - graphviz
     - cmake
-#   commands:
-#     - conda install -c conda-forge cmake
-#     - cmake --version
-#     - python3 setup.py build
-#     - python3 setup.py install
-  
+
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/source/conf.py
