### 2025-03-25 6162bd8
added git lfs to conf
