### 2025-03-25 38c391f
removed install from yaml
