+++ .readthedocs.yaml
@@ -12,11 +12,7 @@ build:
   apt_packages:
     - graphviz
     - cmake
-  commands:
-    - git submodule update --init --recursive
-    - git lfs install
-    - git lfs pull
-
+    
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/source/conf.py
