+++ .readthedocs.yaml
@@ -12,7 +12,7 @@ build:
   apt_packages:
     - graphviz
     - cmake
-    
+
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/source/conf.py
@@ -22,3 +22,7 @@ python:
   install:
     - requirements: docs/requirements.txt
 
+commands:
+   - apt-get update && apt-get install -y git-lfs
+   - git lfs install
+   - git lfs pull
