+++ .readthedocs.yaml
@@ -12,6 +12,8 @@ build:
   apt_packages:
     - graphviz
     - cmake
+  commands:
+    - ./post-build.sh
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
@@ -21,8 +23,3 @@ sphinx:
 python:
   install:
     - requirements: docs/requirements.txt
-
-   commands:
-    - apt-get update && apt-get install -y git-lfs
-    - git lfs install
-    - git lfs pull
