### 2025-03-25 3db29ad
remove ssapy build in post
