+++ post-build.sh
@@ -1,3 +1,4 @@
 #!/bin/bash
 git submodule update --init --recursive
+git lfs install
 git lfs pull
