+++ .readthedocs.yaml
@@ -12,8 +12,7 @@ build:
   apt_packages:
     - graphviz
     - cmake
-  commands:
-    - ./post-build.sh
+
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
@@ -23,3 +22,7 @@ sphinx:
 python:
   install:
     - requirements: docs/requirements.txt
+   
+build:
+  commands:
+    - ./post-build.sh
