+++ .readthedocs.yaml
@@ -13,7 +13,6 @@ build:
     - graphviz
     - cmake
 
-
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/source/conf.py
