+++ post-build.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
-curl -s https://packagecloud.io/install/repositories/github/git-lfs/script.deb.sh | sudo bash
-sudo apt-get install git-lfs -y
+pip install git-lfs
 git submodule update --init --recursive
 git lfs install
 git lfs pull
