+++ docs/source/conf.py
@@ -8,26 +8,26 @@
 from functools import reduce
 import subprocess
 
-# import os
-# import sys
+import os
+import sys
 
-# on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
-# if on_rtd:
-#     # Initialize Git submodules
-#     subprocess.run(["git", "submodule", "update", "--init", "--recursive"], cwd=os.path.abspath("../.."), check=True)
+on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
+if on_rtd:
+    # Add the ssapy directory to the Python path
+    sys.path.insert(0, os.path.abspath('../../ssapy'))
 
-#     # Install Git LFS
-#     subprocess.run(["git", "lfs", "install"], cwd=os.path.abspath("../.."), check=True)
+    # Initialize Git submodules
+    subprocess.run(["git", "submodule", "update", "--init", "--recursive"], check=True)
 
-#     # Pull files managed by Git LFS
-#     subprocess.run(["git", "lfs", "pull"], cwd=os.path.abspath("../.."), check=True)
+    # Install Git LFS
+    subprocess.run(["git", "lfs", "install"], check=True)
 
-#     # Add the ssapy directory to the Python path
-#     sys.path.insert(0, os.path.abspath('../../ssapy'))
+    # Pull files managed by Git LFS
+    subprocess.run(["git", "lfs", "pull"], check=True)
 
-#     # Build and install the package
-#     subprocess.run(["python3", "setup.py", "build"], cwd=os.path.abspath("../.."), check=True)
-#     subprocess.run(["python3", "setup.py", "install"], cwd=os.path.abspath("../.."), check=True)
+    # Build and install the package
+    subprocess.run(["python3", "setup.py", "build"],  check=True)
+    subprocess.run(["python3", "setup.py", "install"], check=True)
 
 
 # -- Project information -----------------------------------------------------
