+++ .readthedocs.yaml
@@ -17,8 +17,6 @@ build:
   apt_packages:
     - graphviz
     - cmake
-  commands:
-    - ./post-build.sh
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
