+++ docs/source/conf.py
@@ -10,26 +10,26 @@ import os
 import subprocess
 import sys
 
-on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
-if on_rtd:
-    # Add the ssapy directory to the Python path
-    sys.path.insert(0, os.path.abspath('../../ssapy'))
+# on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
+# if on_rtd:
+#     # Add the ssapy directory to the Python path
+#     sys.path.insert(0, os.path.abspath('../../ssapy'))
 
-    # Add the Git LFS binary to PATH
-    os.environ["PATH"] += os.pathsep + os.path.abspath('../../bin')
+#     # Add the Git LFS binary to PATH
+#     os.environ["PATH"] += os.pathsep + os.path.abspath('../../bin')
 
-    # Initialize Git submodules
-    subprocess.run(["git", "submodule", "update", "--init", "--recursive"], check=True)
+#     # Initialize Git submodules
+#     subprocess.run(["git", "submodule", "update", "--init", "--recursive"], check=True)
 
-    # Install Git LFS
-    subprocess.run(["git-lfs", "install"], check=True)
+#     # Install Git LFS
+#     subprocess.run(["git-lfs", "install"], check=True)
 
-    # Pull files managed by Git LFS
-    subprocess.run(["git-lfs", "pull"], check=True)
+#     # Pull files managed by Git LFS
+#     subprocess.run(["git-lfs", "pull"], check=True)
 
-    # Build and install the package
-    subprocess.run(["python3", "setup.py", "build"],  check=True)
-    subprocess.run(["python3", "setup.py", "install"], check=True)
+#     # Build and install the package
+#     subprocess.run(["python3", "setup.py", "build"],  check=True)
+#     subprocess.run(["python3", "setup.py", "install"], check=True)
 
 
 # -- Project information -----------------------------------------------------
