### 2025-03-25 8fff7bb
added commands to yaml file
