+++ .readthedocs.yaml
@@ -17,6 +17,18 @@ build:
   apt_packages:
     - graphviz
     - cmake
+  commands:
+   # Add the Git LFS binary to PATH
+    - export PATH=$PATH:$PWD/bin
+   # Initialize Git submodules
+    - git submodule update --init --recursive
+   # Install Git LFS
+    - git-lfs install
+   # Pull files managed by Git LFS
+    - git-lfs pull
+   # Build and install the package
+    - python3 setup.py build
+    - python3 setup.py install
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
