+++ .readthedocs.yaml
@@ -4,10 +4,6 @@
 # Required
 version: 2
 
-# Declare the Python requirements required to build your documentation
-python:
-  install:
-    - requirements: docs/requirements.txt
 
 # Set the OS, Python version, and other tools you might need
 build:
@@ -17,25 +13,24 @@ build:
   apt_packages:
     - graphviz
     - cmake
-  commands:
-   # Add the Git LFS binary to PATH
-    - export PATH=$PATH:/home/docs/checkouts/readthedocs.org/user_builds/ssapy-test/checkouts/latest/bin
-    # Verify that the binary is accessible
-    - ls /home/docs/checkouts/readthedocs.org/user_builds/ssapy-test/checkouts/latest/bin/git-lfs
-    - which git-lfs
-   # Initialize Git submodules
-    - git submodule update --init --recursive
-   # Install Git LFS
-    - git lfs install
-   # Pull files managed by Git LFS
-    - git lfs pull
-   # Build and install the package
-    - python3 setup.py build
-    - python3 setup.py install
+  extra_requirements:
+    - git-lfs
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/source/conf.py
 
+before_install:
+  - git lfs install
+  - git lfs pull
+
+install:
+  - pip install -r requirements.txt
+
+
+# Declare the Python requirements required to build your documentation
+python:
+  install:
+    - requirements: docs/requirements.txt
 
    
