+++ .readthedocs.yaml
@@ -13,7 +13,29 @@ build:
   apt_packages:
     - graphviz
     - cmake
-
+  jobs:
+    post_checkout:
+      # Taken from: https://docs.readthedocs.io/en/stable/build-customization.html#support-git-lfs-large-file-storage
+      # NOTE(alexmillane): Modified to put lfs files in a sub-directory to stop lfs README.md overwriting ours.
+      - ls -a
+      # Download LFS
+      - wget https://github.com/git-lfs/git-lfs/releases/download/v3.1.4/git-lfs-linux-amd64-v3.1.4.tar.gz
+      # Uncompress
+      - mkdir -p rtd_tools
+      - tar xvfz git-lfs-linux-amd64-v3.1.4.tar.gz --directory rtd_tools
+      # Remove the lfs README.md to stop it being picked up by our docs.
+      - mv ./rtd_tools/git-lfs .
+      - rm -r rtd_tools
+      # Modify LFS config paths to point where git-lfs binary was downloaded
+      - git config filter.lfs.process "`pwd`/git-lfs filter-process"
+      - git config filter.lfs.smudge  "`pwd`/git-lfs smudge -- %f"
+      - git config filter.lfs.clean "`pwd`/git-lfs clean -- %f"
+      # Make LFS available in current repository
+      - ./git-lfs install
+      # Download content from remote
+      - ./git-lfs fetch
+      # Make local files to have the real content on them
+      - ./git-lfs checkout
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
