### 2025-03-25 df069e8
clearing bibtex cache
