+++ .readthedocs.yaml
@@ -9,7 +9,7 @@ version: 2
 build:
   os: ubuntu-24.04
   tools:
-    python: "3.8"
+    python: "3.10"
   apt_packages:
     - graphviz
     - cmake
