+++ docs/requirements.txt
@@ -1,11 +1,10 @@
-sphinx
+sphinx==7.2.3
 sphinx-copybutton
 myst-parser
 sphinx-tabs
 sphinx-automodapi
 sphinx_rtd_theme
 sphinxcontrib-bibtex
-sphinx-rtd-theme
 sphinx-autobuild
 nbsphinx
 numpy
@@ -13,7 +12,6 @@ astropy
 jupyter
 asdf
 python-git 
-git-lfs 
 graphviz
 scipy
 matplotlib
