+++ .readthedocs.yaml
@@ -21,7 +21,7 @@ build:
       # Uncompress
       - mkdir -p rtd_tools
       - tar xvfz git-lfs-linux-amd64-v3.1.4.tar.gz --directory rtd_tools
-      # Remove the lfs README.md to stop it being picked up by our docs.
+      # Remove the lfs README.md
       - mv ./rtd_tools/git-lfs .
       - rm -r rtd_tools
       # Modify LFS config paths to point where git-lfs binary was downloaded
@@ -29,8 +29,9 @@ build:
       - git config filter.lfs.smudge  "`pwd`/git-lfs smudge -- %f"
       - git config filter.lfs.clean "`pwd`/git-lfs clean -- %f"
       # Make LFS available in current repository
-      - ./git-lfs install
-      - ./git-lfs pull
+      # - ./git-lfs install
+      # - ./git-lfs pull
+      - git lfs install
       - git lfs pull
 
  
