### 2025-03-25 89169a3
added git-lfs to apt
