+++ docs/source/index.rst
@@ -2,12 +2,11 @@
    sphinx-quickstart on Fri Aug  4 16:23:59 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
-
 .. include:: ../README.rst
 
 .. toctree::
-   :titlesonly:
    :maxdepth: 2
+   :caption: Basics
 
    installation
    examples
@@ -15,7 +14,12 @@
    concepts
    contribution_guide
    references
-   api
+
+.. toctree::
+   :maxdepth: 2
+   :caption: API DOCS
+
+   API Reference Guide <api>
 
 Indices and tables
 ==================
