+++ docs/source/index.rst
@@ -2,7 +2,7 @@
    sphinx-quickstart on Fri Aug  4 16:23:59 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
-.. include:: ../README.rst
+.. include:: ../../README.rst
 
 .. toctree::
    :maxdepth: 2
