+++ .github/workflows/python-publish.yml
@@ -24,13 +24,16 @@ jobs:
 
       - uses: actions/setup-python@v5
         with:
-          python-version: "3.x"
+          python-version: "3.10"
 
       - name: Build release distributions
         run: |
           # NOTE: put your own distribution build steps here.
-          python -m pip install build
-          python -m build
+          apt update
+          apt install build-essential git git-lfs python3 python3-distutils python3-venv graphviz cmake
+          git submodule update --init --recursive
+          git lfs install
+          git lfs pull
 
       - name: Upload distributions
         uses: actions/upload-artifact@v4
