+++ .github/workflows/ci.yml
@@ -89,7 +89,7 @@ jobs:
       if: matrix.os == 'ubuntu-latest' && matrix.python-version == 3.12
       run: |
         sudo apt-get update
-        sudo apt-get install build-essential git git-lfs python3 python3-distutils python3-venv graphviz cmake
+        sudo apt-get install build-essential git git-lfs python3 python3-setuptools python3-venv graphviz cmake
         git submodule update --init --recursive
         git lfs install
         git lfs pull
