+++ .github/workflows/ci.yml
@@ -93,6 +93,8 @@ jobs:
         git submodule update --init --recursive
         git lfs install
         git lfs pull
+        python3 setup.py build
+        python3 setup.py install
         pip install -r docs/requirements.txt
         cd docs && make html && make html
         touch _build/html/.nojekyll
