+++ docs/requirements.txt
@@ -16,5 +16,4 @@ graphviz
 scipy
 matplotlib
 pandas
-# git+https://github.com/LLNL/SSAPy.git
 scikit-learn
