+++ .readthedocs.yaml
@@ -1,46 +1,0 @@
-# Read the Docs configuration file
-# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-
-# Required
-version: 2
-
-
-# Set the OS, Python version, and other tools you might need
-build:
-  os: ubuntu-24.04
-  tools:
-    python: "3.10"
-  apt_packages:
-    - graphviz
-    - cmake
-    - git-lfs
-  jobs:
-    post_checkout:
-      # Taken from: https://docs.readthedocs.io/en/stable/build-customization.html#support-git-lfs-large-file-storage
-      # Download LFS
-      - wget https://github.com/git-lfs/git-lfs/releases/download/v3.1.4/git-lfs-linux-amd64-v3.1.4.tar.gz
-      # Uncompress
-      - mkdir -p rtd_tools
-      - tar xvfz git-lfs-linux-amd64-v3.1.4.tar.gz --directory rtd_tools
-      # Remove the lfs README.md
-      - mv ./rtd_tools/git-lfs .
-      - rm -r rtd_tools
-      # Modify LFS config paths to point where git-lfs binary was downloaded
-      - git config filter.lfs.process "`pwd`/git-lfs filter-process"
-      - git config filter.lfs.smudge  "`pwd`/git-lfs smudge -- %f"
-      - git config filter.lfs.clean "`pwd`/git-lfs clean -- %f"
-      # Make LFS available in current repository
-      - ./git-lfs install
-      - ./git-lfs pull
-
- 
-# Build documentation in the "docs/" directory with Sphinx
-sphinx:
-  configuration: docs/source/conf.py
-
-# Declare the Python requirements required to build your documentation
-python:
-  install:
-    - requirements: docs/requirements.txt
-
-   
