+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.5)
+cmake_minimum_required(VERSION 2.8.12)
 cmake_policy(SET CMP0048 NEW)  # set VERSION
 cmake_policy(SET CMP0063 NEW)  # Compile the static lib with hidden visibility.
 
