### 2025-04-02 359cb77
adding cmake back in cause it wont compile otherwise
