### 2025-04-02 12d3b30
Fixed docstrings in the documentation (#53)
* wrote api file and edited conf

* added a yaml file

* added .

* changed python version in yaml

* added requirements.txt

* changed sphinx version

* tabs

* testing no automodapi

* myst

* added refs

* maybe we needed this function after all

* test

* test

* bib error

* edited linkcode

* edited toctree

* tested install

* python path

* test

* added requirements and specifying import path

* imports

* circular imports?

* circular imports.

* linkcode

* mocked objects

* still circular

* fixed the circlar import we think

* numpy

* n
[…]
