+++ docs/source/conf.py
@@ -3,14 +3,7 @@
 # For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
-import importlib
-import inspect
-from functools import reduce
-import subprocess
 
-import os
-import sys
-sys.path.insert(0, os.path.abspath('../../ssapy/'))
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
@@ -18,8 +11,6 @@ sys.path.insert(0, os.path.abspath('../../ssapy/'))
 project = 'SSAPy'
 copyright = '2018, Lawrence Livermore National Security, LLC'
 author = 'Michael Schneider, Josh Meyers, Edward Schlafly, Julia Ebert, Travis Yeager, et al.'
-
-githash = subprocess.check_output(["git", "rev-parse", "HEAD"]).strip().decode("ascii")
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -34,34 +25,22 @@ extensions = [
     'sphinx_automodapi.automodapi',
     'sphinx_automodapi.smart_resolver',
     'myst_parser',
+    'sphinx_rtd_theme',
+    'sphinx.ext.mathjax',
+    'sphinxcontrib.bibtex',
+    'sphinx.ext.napoleon',
+    'sphinx.ext.viewcode',
 ]
 
-
 def linkcode_resolve(domain, info):
-    if domain != "py":
+    if domain != 'py':
         return None
-    if not info["module"]:
+    if not info['module']:
         return None
-    mod = importlib.import_module(info["module"])
-    modpath = [p for p in sys.path if mod.__file__.startswith(p)]
-    if len(modpath) < 1:
-        raise RuntimeError("Cannot deduce module path")
-    modpath = modpath[0]
-    obj = reduce(getattr, [mod] + info["fullname"].split("."))
-    try:
-        path = inspect.getsourcefile(obj)
-        relpath = path[len(modpath) + 1:]
-        relpath = relpath.split(".egg/")[-1]
-        _, lineno = inspect.getsourcelines(obj)
-    except TypeError:
-        # skip property or other type that inspect doesn't like
-        return None
-    return "http://github.com/LLNL/SSAPy/blob/{}/{}#L{}".format(
-        githash, relpath, lineno
-    )
+    filename = info['module'].replace('.', '/')
+    return "https://github.com/LLNL/SSAPy/tree/main/%s.py" % filename
 
-
-autosummary_generate = False
+autosummary_generate = True
 numpydoc_show_class_members = False
 sphinx_tabs_valid_builders = ['linkcheck']
 source_suffix = ['.rst', '.md']
@@ -84,3 +63,8 @@ html_theme_options = {
     'logo_only': True,
 }
 html_favicon = '_static/images/logo/ssapy_logo.ico'
+
+bibtex_cache = 'none' 
+bibtex_bibfiles = ["refs.bib"]
+bibtex_reference_style = "author_year"
+bibtex_debug = True
