+++ docs/source/references.rst
@@ -1,0 +1,7 @@
+==========
+References
+==========
+
+.. bibliography:: refs.bib
+    :style: plain
+    :all:
