+++ docs/requirements.txt
@@ -1,0 +1,19 @@
+sphinx==7.2.3
+sphinx-copybutton
+myst-parser
+sphinx-tabs
+sphinx-automodapi
+sphinx_rtd_theme
+sphinxcontrib-bibtex
+sphinx-autobuild
+nbsphinx
+numpy
+astropy
+jupyter
+asdf
+python-git 
+graphviz
+scipy
+matplotlib
+pandas
+scikit-learn
