+++ docs/source/refs.bib
@@ -1,3 +1,12 @@
+@inproceedings{ssapyprep,
+    title = {{SSAPy} - Space Situational Awareness for Python},
+    author = {Meyers, Joshua E. and Schneider, Michael D. and Ebert, Julia T. and Schlafly, Edward F. and Yeager, Travis and Perloff, Alexx and Merl, Daniel and Lifset, Noah and Berstein, Jason and Dawson, William A. and Golovich, Nathan and Higgins, Denvir and McGill, Peter and Miller, Caleb and Pruett, Kerianne},
+    booktitle = {The Journal of Open Source Software},
+    publisher = {The Open Journal},
+    note = {Submitted to},
+    year = {2024}
+}
+
 @INPROCEEDINGS{2023amos.conf..208Y,
        author = {{Yeager}, T. and {Pruett}, K. and {Schneider}, M.},
         title = "{Long-term N-body Stability in Cislunar Space}",
@@ -8,7 +17,7 @@
         month = sep,
           eid = {208},
         pages = {208},
-          url = https://amostech.com/TechnicalPapers/2023/Poster/Yeager.pdf,
+          url = {https://amostech.com/TechnicalPapers/2023/Poster/Yeager.pdf},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2023amos.conf..208Y},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
