+++ docs/source/index.rst
@@ -2,7 +2,6 @@
    sphinx-quickstart on Fri Aug  4 16:23:59 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
-
 .. include:: ../../README.rst
 
 .. toctree::
@@ -14,12 +13,13 @@
    benchmarks
    concepts
    contribution_guide
+   references
 
 .. toctree::
    :maxdepth: 2
    :caption: API DOCS
 
-   API Reference Guide <reference>
+   API Reference Guide <api>
 
 Indices and tables
 ==================
