+++ .github/workflows/ci.yml
@@ -88,6 +88,14 @@ jobs:
     - name: Build documentation
       if: matrix.os == 'ubuntu-latest' && matrix.python-version == 3.12
       run: |
+        sudo apt-get update
+        sudo apt-get install build-essential git git-lfs python3 python3-setuptools python3-venv graphviz cmake
+        git submodule update --init --recursive
+        git lfs install
+        git lfs pull
+        python3 setup.py build
+        python3 setup.py install
+        pip install -r docs/requirements.txt
         cd docs && make html && make html
         touch _build/html/.nojekyll
     - name: Deploy documentation
