### 2025-04-04 #7699 @fruzsinaagocs
:wave: @astrophpeter, thank you for your submission!

If you have any reviewers in mind who could look at your submission, please list them here without @-mentioning them. A list of JOSS reviewers can be found [here](https://reviewers.joss.theoj.org).

I noticed that the documentation pipeline for the API reference has been fixed. While we are looking for reviewers, you may wish to double-check for any gaps and errors in the reference. I specifically noticed
* Some badly formatted code blocks: https://software.llnl.gov/SSAPy/api.html#ssapy.utils.continueClass
* Some remaining TODO-s: https://software.llnl.gov/SSAPy/api.html#ssapy.linker.ModelSelectorParams
* Bad docstring formatting, e.g. https://software.llnl.gov/SSAPy/api.html#ssapy.plotUtils.koe_plot
