+++ docs/source/module_template.rst
@@ -20,7 +20,7 @@
 
    .. autosummary::
       :toctree:
-      :template: custom-class-template.rst
+      :template: class_template.rst
    {% for item in classes %}
       {{ item }}
    {%- endfor %}
@@ -66,7 +66,7 @@
 
 .. autosummary::
    :toctree:
-   :template: custom-module-template.rst
+   :template: module_template.rst
    :recursive:
 {% for item in modules %}
    {{ item }}
