+++ .github/workflows/ci.yml
@@ -75,10 +75,10 @@ jobs:
     #    python3 --version
     #    flake8 --version
     #    flake8 ssapy/ tests/ devel/
-    #- name: Test with pytest
-    #  run: |
-    #    python3 ${CI_PROJECT_DIR}/tests/test_orbit.py
-    #    pytest --cov-report=xml --cov=coffea --deselect=test_taskvine
+    - name: Test with pytest
+      run: |
+       python3 ${CI_PROJECT_DIR}/tests/test_orbit.py
+       pytest --cov-report=xml --cov=coffea --deselect=test_taskvine
     - name: Upload coverage reports to Codecov
       if: matrix.os == 'ubuntu-latest' && matrix.python-version == 3.12
       uses: codecov/codecov-action@v4
