+++ .github/workflows/ci.yml
@@ -75,16 +75,7 @@ jobs:
     #    python3 --version
     #    flake8 --version
     #    flake8 ssapy/ tests/ devel/
-    - name: Test with pytest
-      run: |
-       python3 ${CI_PROJECT_DIR}/tests/test_orbit.py
-       pytest --cov-report=xml --cov=coffea --deselect=test_taskvine
-    - name: Upload coverage reports to Codecov
-      if: matrix.os == 'ubuntu-latest' && matrix.python-version == 3.12
-      uses: codecov/codecov-action@v4
-      with:
-        token: ${{ secrets.CODECOV_TOKEN }}
-        slug: LLNL/SSAPy
+
     - name: Build documentation
       if: matrix.os == 'ubuntu-latest' && matrix.python-version == 3.12
       run: |
@@ -98,6 +89,16 @@ jobs:
         pip install -r docs/requirements.txt
         cd docs && make html && make html
         touch _build/html/.nojekyll
+    - name: Test with pytest
+      run: |
+       python3 ${CI_PROJECT_DIR}/tests/test_orbit.py
+       pytest --cov-report=xml --cov=coffea --deselect=test_taskvine
+    - name: Upload coverage reports to Codecov
+      if: matrix.os == 'ubuntu-latest' && matrix.python-version == 3.12
+      uses: codecov/codecov-action@v4
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        slug: LLNL/SSAPy
     - name: Deploy documentation
       if: (github.event_name == 'push' || github.event_name == 'schedule' || github.event_name == 'release') && matrix.os == 'ubuntu-latest' && matrix.python-version == 3.12
       uses: crazy-max/ghaction-github-pages@v4
