### 2025-04-14 e03c74b
defining project directory
