### 2025-04-14 1a43a74
fixing ssapy install
