### 2025-04-14 7e49479
fixing mac ssapy install
