### 2025-04-14 59c7bc9
fixing ubuntu install
