+++ .github/workflows/ci.yml
@@ -53,8 +53,6 @@ jobs:
         git submodule update --init --recursive
         git lfs install
         git lfs pull
-        python3 setup.py build
-        python3 setup.py install
         python -m pip install --upgrade pip setuptools flake8
         python -m pip install -r requirements.txt
         # install checked out SSAPy
