### 2025-04-14 a72e77a
fixing pytest dependencies
