### 2025-04-15 #7699 @fruzsinaagocs
👋 @steo85it & @sblunt, would any of you be willing to review this submission for JOSS? Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
