+++ .github/workflows/build-containers.yml
@@ -1,0 +1,74 @@
+yaml
+on: 
+  repository_dispatch:
+    types: [trigger-container-build]
+
+jobs:
+  pass:
+    runs-on: ubuntu-latest
+    steps:
+      - run: echo "Initial job to satisfy dependency requirement"
+
+  deploy-images:
+    needs: [pass]
+    if: ${{ github.event_name == 'push' || github.event_name == 'release' }}
+    runs-on: ubuntu-latest
+    permissions:
+      packages: write
+    steps:
+      - name: Checkout SSAPy
+        uses: actions/checkout@v4
+        with:
+          path: SSAPy
+          lfs: true
+          submodules: recursive
+
+      - name: Look around
+        run: |
+          echo ${{ github.workspace }}
+          pwd
+          ls -alh ./
+
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v3
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Setup Metadata
+        uses: docker/metadata-action@v5
+        id: docker_meta
+        with:
+          images: ghcr.io/${{ github.repository_owner }}/ssapy-centos-s9
+          tags: |
+            type=schedule,pattern=nightly
+            type=schedule,pattern=develop
+            type=semver,pattern={{version}}
+            type=semver,pattern={{major}}.{{minor}}
+            type=semver,pattern={{major}}
+            type=ref,event=branch
+            type=ref,event=pr
+            type=ref,event=tag
+
+      - name: Log in to GitHub Container Registry
+        uses: docker/login-action@v3
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Build & Deploy Images
+        uses: docker/build-push-action@v5
+        with:
+          context: .
+          file: SSAPy/.docker/Dockerfile
+          platforms: linux/amd64,linux/arm64
+          push: ${{ github.event_name != 'pull_request' }}
+          tags: ${{ steps.docker_meta.outputs.tags }}
+          labels: ${{ steps.docker_meta.outputs.labels }}
+
+  images-built:
+    needs: [deploy-images]
+    runs-on: ubuntu-latest
+    steps:
+      - run: echo "Container built"
