### 2025-04-15 6ce2ab2
compute docstrings
