### 2025-04-18 #7699 @fruzsinaagocs
Thank you both for the responses and recommendation! @wdesprats would you be willing to review? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
