### 2025-04-18 240a92d
added ssapy/data after pip install functionality
