+++ pyproject.toml
@@ -2,7 +2,7 @@
 name = "test_ssapy_denvirhiggins"
 version = "0.7.0"
 authors = [
-  { name="Denvir Higgins", email="denvir.j.higgins@gmail.com" },
+    { name="Denvir Higgins", email="denvir.j.higgins@gmail.com" },
 ]
 description = "A small example package"
 readme = "README.rst"
@@ -18,9 +18,12 @@ Homepage = "https://github.com/LLNL/SSAPy"
 Issues = "https://github.com/LLNL/SSAPy/issues"
 
 [build-system]
-requires = ["scikit-build-core", "cmake", "setuptools", "wheel"]
-build-backend = "scikit_build_core.build"
+requires = ["setuptools", "wheel", "scikit-build-core"]
+build-backend = "setuptools.build_meta"
 
 [tool.scikit-build]
 build-dir = "build"
 wheel.packages = ["ssapy"]
+
+[tool.setuptools.package-data]
+exclude = ["ssapy/data/**"]
