+++ MANIFEST.in
@@ -1,11 +1,13 @@
-include README.md
+include README.rst
+include LICENSE
 include include/*.h
 include pysrc/*.cpp
 include src/*.cpp
+include pyproject.toml
+include CMakeLists.txt
 recursive-include pybind11/include/ *
 include pybind11/CMakeLists.txt
 include pybind11/LICENSE
 recursive-include pybind11/tools *
-include CMakeLists.txt
 exclude ssapy/data/*
 recursive-exclude ssapy/data *
