### 2025-04-23 4d1f9a7
removing templates
