+++ setup.py
@@ -1,15 +1,14 @@
-from setuptools import setup, Extension
+from setuptools import setup, Extension, find_packages
 from setuptools.command.build_ext import build_ext
+from setuptools.command.install import install
 import subprocess
 import sys
 import os
-
 
 class CMakeExtension(Extension):
     def __init__(self, name, sourcedir=''):
         Extension.__init__(self, name, sources=[])
         self.sourcedir = os.path.abspath(sourcedir)
-
 
 class CMakeBuild(build_ext):
     def run(self):
@@ -42,6 +41,32 @@ class CMakeBuild(build_ext):
         subprocess.check_call(['cmake', ext.sourcedir] + cmake_args, cwd=self.build_temp, env=env)
         subprocess.check_call(['cmake', '--build', '.'] + build_args, cwd=self.build_temp)
 
+class PostInstallCommand(install):
+    """Post-installation to download the ssapy/data directory."""
+    def run(self):
+        install.run(self)
+        # Add logic to download the ssapy/data directory
+        print("Downloading ssapy/data directory from GitHub...")
+        repo_url = "https://github.com/LLNL/SSAPy.git"
+        commit_hash = "0ea3174"  # Specific commit hash
+        temp_repo_dir = "temp_repo"
+        target_dir = os.path.join(os.path.dirname(__file__), "ssapy/data")
+        
+        try:
+            # Clone the repository
+            subprocess.run(["git", "clone", repo_url, temp_repo_dir], check=True)
+            
+            # Checkout the specific commit
+            subprocess.run(["git", "-C", temp_repo_dir, "checkout", commit_hash], check=True)
+            
+            # Copy the ssapy/data directory
+            subprocess.run(["cp", "-r", os.path.join(temp_repo_dir, "ssapy/data"), target_dir], check=True)
+            
+            print(f"Data directory downloaded to {target_dir}")
+        finally:
+            # Cleanup the temporary repository
+            subprocess.run(["rm", "-rf", temp_repo_dir], check=True)
+
 
 setup(
     name='SSAPy',
@@ -50,11 +75,11 @@ setup(
     author='Michael Schneider, Joshua Meyers, Edward Schlafly, Julia Ebert, Travis Yeager',
     author_email='yeager7@llnl.gov',
     url='https://github.com/LLNL/SSAPy',
-    packages=['ssapy'],
+    packages=find_packages(exclude=["ssapy.data", "ssapy.data.*"]),  # Exclude ssapy/data directory
     package_dir={'ssapy': 'ssapy'},
-    package_data={'ssapy': ['ssapy/**/*']},
+    package_data={'ssapy': ['ssapy/**/*']},  # Include other package data
     ext_modules=[CMakeExtension('ssapy._ssapy')],
-    cmdclass=dict(build_ext=CMakeBuild),
+    cmdclass=dict(build_ext=CMakeBuild, install=PostInstallCommand),  # Added PostInstallCommand
     license='MIT',
     tests_require=[
         'pytest',
