+++ pyproject.toml
@@ -1,0 +1,54 @@
+[project]
+name = "test_ssapy_denvirhiggins"
+version = "0.7.11"
+authors = [
+    { name="Denvir Higgins", email="denvir.j.higgins@gmail.com" },
+]
+description = "A small example package"
+readme = "README.rst"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Operating System :: OS Independent",
+]
+license = { file = "LICENSE" }
+dependencies = [
+        'numpy',
+        'scipy',
+        'astropy',
+        'pyerfa',
+        'emcee',
+        'lmfit',
+        'sgp4',
+        'matplotlib',
+        'pandas',
+        'h5py',
+        'pypdf2',
+        'imageio',
+        'ipython',
+        'ipyvolume',
+        'ipython_genutils',
+        'jplephem',
+        'tqdm',
+        'myst-parser',
+        'graphviz',
+    ]
+
+[build-system]
+requires = ["setuptools>=42", "wheel", "scikit-build", "pybind11"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+packages = ["ssapy"]
+include-package-data = true
+
+[tool.setuptools.package-data]
+ssapy = ["*.cpp", "*.h"]
+
+[tool.setuptools.exclude-package-data]
+ssapy = ["data/*"]
+
+[project.urls]
+Homepage = "https://github.com/LLNL/SSAPy"
+Issues = "https://github.com/LLNL/SSAPy/issues"
+Documentation = "https://software.llnl.gov/SSAPy/" 
