+++ docs/source/conf.py
@@ -32,10 +32,6 @@ extensions = [
     'sphinx.ext.viewcode',
 ]
 
-<<<<<<< HEAD
-=======
-
->>>>>>> main
 def linkcode_resolve(domain, info):
     if domain != 'py':
         return None
