+++ ssapy/orbit_solver.py
@@ -3,8 +3,6 @@ Classes to solve for Keplerian orbital parameters from different initial inputs
 
 # TODO: Provide overview of different solvers
 
-TwoPosOrbitSolver:
-
 GaussTwoPosOrbitSolver:
 
 DanchickTwoPosOrbitSolver:
@@ -129,6 +127,29 @@ class TwoPosOrbitSolver(metaclass=abc.ABCMeta):
 
 
 class GaussTwoPosOrbitSolver(TwoPosOrbitSolver):
+    """
+    A class for solving two-position orbit determination problems using 
+    the Gauss method. This class extends the `TwoPosOrbitSolver` base class 
+    and implements a method to compute the orbital parameter `p` based on 
+    Shefer's equations.
+
+    Attributes:
+        eps (float): Convergence tolerance for iterative calculations.
+        maxiter (int): Maximum number of iterations allowed for convergence.
+        m (float): A parameter related to the orbit determination problem.
+        ell (float): A parameter related to the orbit determination problem.
+        kappa (float): A constant used in orbital calculations.
+        sigma (float): A constant used in orbital calculations.
+        tau (float): A constant used in orbital calculations.
+        mu (float): Standard gravitational parameter.
+
+    Methods:
+        _getP():
+            Computes the orbital parameter `p` using an iterative approach 
+            based on Shefer's equations. This method solves for `eta` 
+            iteratively until convergence and then calculates `p` using the 
+            converged value of `eta`.
+    """
     def __init__(self, *args, **kwargs):
         TwoPosOrbitSolver.__init__(self, *args, **kwargs)
 
@@ -157,6 +178,41 @@ class GaussTwoPosOrbitSolver(TwoPosOrbitSolver):
 
 
 class DanchickTwoPosOrbitSolver(TwoPosOrbitSolver):
+    """
+    A class for solving two-position orbit determination problems using 
+    the Danchick method. This class extends the `TwoPosOrbitSolver` base 
+    class and implements methods to compute the orbital parameter `p` 
+    based on Shefer's equations.
+
+    Attributes:
+        eps (float): Convergence tolerance for iterative calculations.
+        maxiter (int): Maximum number of iterations allowed for convergence.
+        m (float): A parameter related to the orbit determination problem.
+        ell (float): A parameter related to the orbit determination problem.
+        kappa (float): A constant used in orbital calculations.
+        sigma (float): A constant used in orbital calculations.
+        tau (float): A constant used in orbital calculations.
+        mu (float): Standard gravitational parameter.
+        cos2f (float): Cosine of twice the true anomaly, used to determine 
+            the branch of the solution.
+
+    Methods:
+        X(g):
+            Compute the function X(g) as described in Shefer's equation (11).
+            This is used in the iterative solution process.
+
+        dXdg(g):
+            Compute the derivative of X(g) with respect to g, based on 
+            Shefer's equation (12). This is used to refine the solution 
+            during iteration.
+
+        _getP():
+            Computes the orbital parameter `p` using iterative methods 
+            based on Shefer's equations. Depending on the value of `cos2f`, 
+            the method selects the appropriate branch of the solution and 
+            iteratively solves for `eta` or `x` until convergence. The final 
+            value of `p` is calculated using Shefer's equation (2).
+    """
     def __init__(self, *args, **kwargs):
         TwoPosOrbitSolver.__init__(self, *args, **kwargs)
 
@@ -214,6 +270,101 @@ class DanchickTwoPosOrbitSolver(TwoPosOrbitSolver):
 
 
 class SheferTwoPosOrbitSolver(TwoPosOrbitSolver):
+    """
+    A class for solving two-position orbit determination problems using 
[…]
