### 2025-04-25 47dd5eb
module docstring
