### 2025-04-25 3f67d76
plot docstrings
