### 2025-04-28 cc20e44
Refactor tests to simplify sampling LEO and GEO orbits (#40)
Co-authored-by: Jason Bernstein <bernstein8@llnl.gov>
