+++ tests/test_orbit.py
@@ -15,7 +15,7 @@ from ssapy.orbit import _hyperbolicEccentricToTrueLongitude, _hyperbolicTrueToEc
 from ssapy.orbit import _ellipticalEccentricToMeanLongitude, _ellipticalMeanToEccentricLongitude
 from ssapy.orbit import _hyperbolicEccentricToMeanLongitude, _hyperbolicMeanToEccentricLongitude
 from ssapy.utils import normed, norm, teme_to_gcrf
-from ssapy_test_helpers import timer, checkAngle, checkSphere
+from ssapy_test_helpers import timer, checkAngle, checkSphere, sample_orbit, sample_LEO_orbit, sample_GEO_orbit
 
 
 try:
@@ -142,28 +142,10 @@ def test_orbit_ctor():
     eqElts = []
     kElts = []
     for _ in range(1000):
-        # Pick a distance near GEO
-        r = np.random.uniform(4e7, 5e7)
-        # Pick a random direction (not uniform on sphere)
-        theta = np.random.uniform(0, 2*np.pi)
-        phi = np.random.uniform(0, np.pi)
-        x = r * np.cos(theta) * np.sin(phi)
-        y = r * np.sin(theta) * np.sin(phi)
-        z = r * np.cos(phi)
-        r = np.array([x, y, z])
-        rs.append(r)
-        # Pick a velocity near VGEO
-        v = np.random.uniform(2.7e3, 3.3e3)
-        # and a randomish direction
-        theta = np.random.uniform(0, 2*np.pi)
-        phi = np.random.uniform(0, np.pi)
-        vx = v * np.cos(theta) * np.sin(phi)
-        vy = v * np.sin(theta) * np.sin(phi)
-        vz = v * np.cos(phi)
-        v = np.array([vx, vy, vz])
-        vs.append(v)
+        orbit = sample_GEO_orbit(t=0.0)
+        rs.append(orbit.r)
+        vs.append(orbit.v)
 
-        orbit = ssapy.Orbit(r, v, 0.0)
         eqElts.append(orbit.equinoctialElements)
         kElts.append(orbit.keplerianElements)
         orbit2 = ssapy.Orbit.fromEquinoctialElements(*orbit.equinoctialElements, orbit.t)
@@ -182,32 +164,14 @@ def test_orbit_ctor():
             np.testing.assert_allclose(orb.hx, np.tan(orb.i/2)*np.cos(orb.raan))
             np.testing.assert_allclose(orb.hy, np.tan(orb.i/2)*np.sin(orb.raan))
             checkAngle(orb.lv, orb.trueAnomaly + lonPa, atol=1e-13)
-            np.testing.assert_allclose(r, orb.r, atol=1e-4, rtol=0)
-            np.testing.assert_allclose(v, orb.v, atol=1e-8, rtol=0)
+            np.testing.assert_allclose(orbit.r, orb.r, atol=1e-4, rtol=0)
+            np.testing.assert_allclose(orbit.v, orb.v, atol=1e-8, rtol=0)
             np.testing.assert_equal(0, orb.t)
 
-        # Repeat near LEO
-        r = np.random.uniform(7e6, 1e7)
-        # Pick a random direction (not uniform on sphere)
-        theta = np.random.uniform(0, 2*np.pi)
-        phi = np.random.uniform(0, np.pi)
-        x = r * np.cos(theta) * np.sin(phi)
-        y = r * np.sin(theta) * np.sin(phi)
-        z = r * np.cos(phi)
-        r = np.array([x, y, z])
-        rs.append(r)
-        # Pick a velocity near VLEO
-        v = np.random.uniform(7.7e3, 7.9e3)
-        # and a randomish direction
-        theta = np.random.uniform(0, 2*np.pi)
[…]
