### 2025-04-28 01ae96c
Added docstrings to compute.py (#56)
* testing autosum

* testing automodule

* testing template

* fixing templates

* fixing api

* module test

* module test again

* api test

* edits

* codecov test

* testing order

* defining project directory

* fixing ssapy install

* fixing mac ssapy install

* fixing ubuntu install

* fixing pytest dependencies

* continue on error

* reverting ci

* testing container build

* .

* reverting ci

* compute docstrings

* module doc string

* alexx suggestions

* removing templates

* fixing dir

* alexx questions

* still works

* merge conflicts

* s
[…]
