### 2025-04-28 fa002a0
removing containerization
