### 2025-04-28 34ce655
Fixing codecov and README (#55)
* testing autosum

* testing automodule

* testing template

* fixing templates

* fixing api

* module test

* module test again

* api test

* edits

* codecov test

* testing order

* defining project directory

* fixing ssapy install

* fixing mac ssapy install

* fixing ubuntu install

* fixing pytest dependencies

* continue on error

* reverting ci

* testing container build

* .

* reverting ci

* testing container build

* editing container file

* pytest and trigger updates

* removing coffea

* removing containerization and acknowledging coffea

* REA
[…]
