+++ .codecov.yml
@@ -1,3 +1,17 @@
+# Attempt to fix "Missing base commit" messages in the codecov UI.
+# Because we do not run full tests on package PRs, package PRs' merge
+# commits on `main` don't have coverage info.  It appears that
+# codecov will give you an error if the pseudo-base's coverage data
+# doesn't all apply properly to the real PR base.
+#
+# See here for docs:
+#   https://docs.codecov.com/docs/comparing-commits#pseudo-comparison
+# See here for another potential solution:
+#   https://community.codecov.com/t/2480/15
+
+codecov:
+  allow_coverage_offsets: true
+
 coverage:
   precision: 2
   round: nearest
@@ -10,22 +24,17 @@ coverage:
 #ignore:
 #  - test/.*
 
-comment: off
+comment:                
+  layout: " diff, flags, files"
+  behavior: default
+  require_changes: false  # learn more in the Requiring Changes section below
+  require_base: false        # [true :: must have a base report to post]
+  require_head: true       # [true :: must have a head report to post]
+  hide_project_coverage: false # [true :: only show coverage on the git diff]
 
 # Inline codecov annotations make the code hard to read, and they add
 # annotations in files that seemingly have nothing to do with the PR.
 github_checks:
-    annotations: false
+  annotations: false
 
-# Attempt to fix "Missing base commit" messages in the codecov UI.
-# Because we do not run full tests on package PRs, package PRs' merge
-# commits on `main` don't have coverage info.  It appears that
-# codecov will give you an error if the pseudo-base's coverage data
-# doesn't all apply properly to the real PR base.
-#
-# See here for docs:
-#   https://docs.codecov.com/docs/comparing-commits#pseudo-comparison
-# See here for another potential solution:
-#   https://community.codecov.com/t/2480/15
-codecov:
-  allow_coverage_offsets: true
+
