+++ CITATION.cff
@@ -72,7 +72,7 @@ authors:
 - family-names: "Ebert"
   given-names: "Julia"
   affiliation: "Fleet Robotics"
-  orcid: ""
+  orcid: "https://orcid.org/0000-0002-1975-772X"
   email: "julia@juliaebert.com"
 - family-names: "Merl"
   given-names: "Daniel"
