+++ ssapy/rvsampler.py
@@ -543,12 +543,66 @@ def circular_guess(arc, indices=[0, 1]):
     # circular orbit.
 
     def vr2(r):
+        """
+        Calculate the radial velocity difference based on the provided position vector.
+
+        Parameters:
+        -----------
+        r : numpy.ndarray
+            A position vector representing the relative location along the line of sight.
+
+        Returns:
+        --------
+        float
+            The calculated radial velocity difference based on the Earth's gravitational constant,
+            position, and velocity components.
+
+        Example:
+        --------
+        >>> r = np.array([1.0, 2.0, 3.0])
+        >>> vr2(r)
+        -12345.678  # Example output (depends on global variable values)
+        """
+
         pos = rStation + r*lineOfSight
         R = np.sqrt(np.sum(pos**2))
         return WGS84_EARTH_MU/R - normSq(muAlpha*rAlpha*r + muDelta*rDelta*r +
                                          vGroundSky)
 
     def vR(r, ub=np.inf, verbose=False, square=False):
+        """
+        Compute the radial velocity at a given position along the line of sight, 
+        with optional bounds and verbosity.
+
+        Parameters:
+        -----------
+        r : float
+            The position along the line of sight.
+        ub : float, optional
+            Upper bound for the position `r`. If `r` exceeds this bound, it is reflected 
+            back within the bounds. Default is `np.inf` (no bound).
+        verbose : bool, optional
+            If `True`, additional information may be printed during execution. 
+            Default is `False`.
+        square : bool, optional
+            If `True`, the squared radial velocity (`vrsquared`) is returned instead 
+            of the computed radial velocity. Default is `False`.
+
+        Returns:
+        --------
+        float
+            The computed radial velocity at the given position `r`. If `square=True`, 
+            the squared radial velocity is returned instead.
+
+        Example:
+        --------
+        >>> r = 5.0
+        >>> vR(r, ub=10.0, verbose=True, square=False)
+        123.456  # Example output (depends on global variable values)
+
+        >>> vR(r, ub=10.0, square=True)
+        15234.789  # Example squared radial velocity output
+        """
         if r > ub:
             sgn = -1
             r = 2*ub - r
@@ -1333,6 +1387,99 @@ class ParamOrbitRV(ParamOrbitTranslator):
 
 
 class ParamOrbitEquinoctial(ParamOrbitTranslator):
+    """
+    A class for translating orbital parameters in equinoctial form, 
+    including input/output transformations and conversions between 
+    parameters and orbit objects.
+
+    Inherits:
+    ---------
+    ParamOrbitTranslator : Base class providing foundational functionality 
+    for orbital parameter translation.
+
+    Methods:
+    --------
+    __init__(*args, **kwargs):
+        Initialize the `ParamOrbitEquinoctial` object, inheriting behavior 
+        from the `ParamOrbitTranslator` base class.
+
+    input_param_translation(p):
+        Transform input orbital parameters by scaling the first parameter 
+        (typically semi-major axis) for internal use.
+
+        Parameters:
+        -----------
+        p : array-like
+            Input orbital parameters.
+
+        Returns:
+        --------
+        numpy.ndarray
+            Transformed orbital parameters with the first element scaled.
+
+    output_covar_translation(covar):
+        Transform covariance matrix for output by scaling the first row 
+        and column (typically related to semi-major axis).
+
+        Parameters:
+        -----------
+        covar : array-like
+            Input covariance matrix.
+
+        Returns:
+        --------
+        numpy.ndarray
[…]
