### 2025-04-28 51c733f
more docstrings
