+++ ssapy/rvsampler.py
@@ -1368,6 +1368,72 @@ class ParamOrbitTranslator():
 
 
 class ParamOrbitRV(ParamOrbitTranslator):
+    """
+    A class for translating orbital parameters in position-velocity (RV) form, 
+    including conversions between parameters and orbit objects.
+
+    Inherits:
+    ---------
+    ParamOrbitTranslator : Base class providing foundational functionality 
+    for orbital parameter translation.
+
+    Methods:
+    --------
+    __init__(*args, **kwargs):
+        Initialize the `ParamOrbitRV` object, inheriting behavior from the 
+        `ParamOrbitTranslator` base class.
+
+    param_to_orbit(p):
+        Convert position-velocity (RV) parameters to an `Orbit` object.
+
+        Parameters:
+        -----------
+        p : array-like
+            Orbital parameters in position-velocity form. The first three 
+            elements represent the position vector (`r`), and the next three 
+            elements represent the velocity vector (`v`).
+
+        Returns:
+        --------
+        Orbit
+            An `Orbit` object created from the position (`r`) and velocity (`v`) 
+            vectors, along with propagation keywords.
+
+        Notes:
+        ------
+        - Additional propagation keywords are extracted from the full parameter set.
+        - The `Orbit` class is expected to support initialization with `r` and `v`.
+
+    orbit_to_param(orbit):
+        Convert an `Orbit` object to a list of position-velocity (RV) parameters 
+        and associated propagation keywords.
+
+        Parameters:
+        -----------
+        orbit : Orbit
+            An `Orbit` object containing position (`r`) and velocity (`v`) vectors.
+
+        Returns:
+        --------
+        numpy.ndarray
+            A concatenated array of position (`r`), velocity (`v`), and propagation 
+            keywords extracted from the orbit.
+
+    Notes:
+    ------
+    - This class assumes orbital parameters are represented in position-velocity form.
+    - The `Orbit` class is expected to provide attributes `r` and `v` for position 
+      and velocity vectors, respectively.
+
+    Example:
+    --------
+    >>> translator = ParamOrbitRV()
+    >>> params = [7000, 0, 0, 0, 7.5, 0]  # Example position-velocity parameters
+    >>> orbit = translator.param_to_orbit(params)
+    >>> new_params = translator.orbit_to_param(orbit)
+    >>> print(new_params)
+    [7000, 0, 0, 0, 7.5, 0]  # Example output matching input
+    """
     def __init__(self, *args, **kwargs):
         super(ParamOrbitRV, self).__init__(*args, **kwargs)
 
@@ -1516,6 +1582,96 @@ class ParamOrbitEquinoctial(ParamOrbitTranslator):
 
 
 class ParamOrbitAngle(ParamOrbitTranslator):
+    """
+    A class for translating orbital parameters in angle-based form, 
+    including conversions between angular parameters and orbit objects.
+
+    Inherits:
+    ---------
+    ParamOrbitTranslator : Base class providing foundational functionality 
+    for orbital parameter translation.
+
+    Attributes:
+    -----------
+    initObsPos : array-like
+        Initial observer position vector used for angle-based calculations.
+    initObsVel : array-like
+        Initial observer velocity vector used for angle-based calculations.
+
+    Methods:
+    --------
+    __init__(initparam, epoch, initObsPos, initObsVel, **kwargs):
+        Initialize the `ParamOrbitAngle` object with initial parameters, 
+        epoch, observer position, and observer velocity.
+
+        Parameters:
+        -----------
+        initparam : array-like
+            Initial orbital parameters.
[…]
