### 2025-04-28 d4594d2
Added docstrings to utils.py (#67)
* testing autosum

* testing automodule

* testing template

* fixing templates

* fixing api

* module test

* module test again

* api test

* edits

* codecov test

* testing order

* defining project directory

* fixing ssapy install

* fixing mac ssapy install

* fixing ubuntu install

* fixing pytest dependencies

* continue on error

* reverting ci

* testing container build

* .

* reverting ci

* alexx suggestions

* removing templates

* fixing dir

* alexx questions

* still works

* one billion docstrings

---------

Co-authored-by: dehiggin <higg
[…]
