+++ ssapy/utils.py
@@ -31,6 +31,28 @@ def find_file(filename, ext=None):
 
 
 def _wrapToPi(angle):
+    """
+    Wrap an angle to the range [-π, π].
+
+    Parameters:
+    -----------
+    angle : float or array-like
+        The input angle(s) in radians. Can be a single value or an array of values.
+
+    Returns:
+    --------
+    wrapped_angle : float or array-like
+        The angle(s) wrapped to the range [-π, π].
+
+    Example:
+    --------
+    >>> _wrapToPi(4)
+    -2.283185307179586
+    >>> _wrapToPi(-4)
+    2.283185307179586
+    >>> _wrapToPi([3.5, -3.5, 6.5])
+    array([-2.78318531,  2.78318531,  0.28318531])
+    """
     return (angle + np.pi) % (2 * np.pi) - np.pi
 
 
@@ -172,20 +194,129 @@ def find_all_zeros(f, low, high, n=100):
 # faster than np.linalg.norm().
 # These compute the norm over the last axis and maintain the other leading axes.
 def normSq(arr):
+    """
+    Compute the squared norm of an array over the last axis while preserving leading axes.
+
+    This function calculates the squared norm of vectors along the last axis of the input array 
+    using Einstein summation notation (`np.einsum`). It is designed to be portable and efficient, 
+    offering better performance than `np.linalg.norm()` for this specific use case.
+
+    Parameters:
+    -----------
+    arr : array-like
+        Input array where the squared norm is computed along the last axis. 
+        The array can have any shape, with the last axis representing the vector components.
+
+    Returns:
+    --------
+    norm_squared : array-like
+        The squared norm of the input array computed along the last axis. 
+        The output shape matches the input shape, excluding the last axis.
+
+    Notes:
+    ------
+    - This implementation is more portable and faster than using `np.linalg.norm()` 
+      for computing norms over the last axis, especially for large arrays.
+    - It is slightly less specialized compared to methods optimized for fixed shapes 
+      (e.g., shape `(3,)`), but it generalizes well to arrays of arbitrary shape.
+    - The computation uses `np.einsum("...i,...i", arr, arr)`, which efficiently performs 
+      the summation of squared components along the last axis.
+
+    Example:
+    --------
+    >>> import numpy as np
+    >>> arr = np.array([[1, 2, 3], [4, 5, 6]])
+    >>> normSq(arr)
+    array([14, 77])
+    """
     return np.einsum("...i,...i", arr, arr)
 
 
 # Faster to not dispatch back to normSq
 def norm(arr):
+    """
+    Compute the Euclidean norm of an array over the last axis while preserving leading axes.
+
+    Parameters:
+    -----------
+    arr : array-like
+        Input array where the Euclidean norm is computed along the last axis. 
+        The array can have any shape, with the last axis representing the vector components.
+
+    Returns:
+    --------
+    norm : array-like
+        The Euclidean norm of the input array computed along the last axis. 
+        The output shape matches the input shape, excluding the last axis.
+
+    Notes:
+    ------
+    - This implementation directly computes the norm using `np.sqrt` and `np.einsum`, 
+      avoiding the intermediate step of calling `normSq`. This improves performance.
+    - The computation uses `np.einsum("...i,...i", arr, arr)` to efficiently sum the squared 
+      components along the last axis, followed by taking the square root.
+    - Faster than using `np.linalg.norm()` for this specific use case and generalizes well 
+      to arrays of arbitrary shape.
+
+    Example:
+    --------
+    >>> import numpy as np
[…]
