### 2025-04-28 b33c071
Creating pip install capabilities (#60)
* testing autosum

* testing automodule

* testing template

* fixing templates

* fixing api

* module test

* module test again

* api test

* edits

* codecov test

* testing order

* defining project directory

* fixing ssapy install

* fixing mac ssapy install

* fixing ubuntu install

* fixing pytest dependencies

* continue on error

* reverting ci

* testing container build

* .

* reverting ci

* alexx suggestions

* removing templates

* fixing dir

* redoing pypi branch

* version num

* alexx questions

* still works

* merge conflicts

* mer
[…]
