### 2025-04-29 0ffa3f1
fixing setup.py and toml bugs
