+++ CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.5)
 cmake_policy(SET CMP0048 NEW)
 cmake_policy(SET CMP0063 NEW)
 
-project(_ssapy VERSION 1.0 LANGUAGES CXX)
+project(ssapy VERSION 1.0 LANGUAGES CXX)
 
 # Compiler settings
 set(CMAKE_CXX_STANDARD 11)
@@ -12,7 +12,7 @@ set(CMAKE_CXX_VISIBILITY_PRESET hidden)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 set(CMAKE_MACOSX_RPATH 1)
 
-# Output directories
+# Output directory
 if(NOT CMAKE_LIBRARY_OUTPUT_DIRECTORY)
     set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 endif()
@@ -20,7 +20,7 @@ endif()
 # Find Python
 find_package(Python3 REQUIRED COMPONENTS Interpreter Development)
 
-# Include pybind11 via FetchContent
+# Bring in pybind11
 include(FetchContent)
 FetchContent_Declare(
   pybind11
@@ -29,19 +29,18 @@ FetchContent_Declare(
 )
 FetchContent_MakeAvailable(pybind11)
 
-# Include dirs
+# Include your headers
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
 
-pybind11_add_module(_ssapy pysrc/ssapy.cpp)
+# Build static C++ logic (optional)
 add_library(ssapy STATIC src/ssapy.cpp)
+set_target_properties(ssapy PROPERTIES POSITION_INDEPENDENT_CODE ON)
+
+# Build Python extension module
+pybind11_add_module(_ssapy pysrc/ssapy.cpp)
 target_link_libraries(_ssapy PUBLIC ssapy)
 
+# Install into the ssapy/ Python package directory
 install(TARGETS _ssapy
     LIBRARY DESTINATION ssapy
-)
-
-install(TARGETS _ssapy
-    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
-    ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
-    RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
 )
