### 2025-04-29 d6873d0
uncommenting _ssapy
