+++ pyproject.toml
@@ -11,7 +11,7 @@ classifiers = [
     "Programming Language :: Python :: 3",
     "Operating System :: OS Independent",
 ]
-license = { file = "LICENSE" }
+license = "MIT"
 dependencies = [
     'numpy',
     'scipy',
