+++ requirements.txt
@@ -25,3 +25,4 @@ ipython
 pytest
 imageio
 tqdm
+build
