+++ .github/workflows/ci.yml
@@ -65,7 +65,7 @@ jobs:
 
         # Build and install the project
         python3 -m build
-        pip install SSAPy/dist/ssapy-1.0.0-cp38-cp38-linux_x86_64.whl
+        pip install dist/ssapy-1.0.0-cp38-cp38-linux_x86_64.whl
 
         # List installed Python packages
         python -m pip list
@@ -89,7 +89,7 @@ jobs:
     
         # Build and install the project
         python3 -m build
-        pip install SSAPy/dist/ssapy-1.0.0-cp38-cp38-linux_x86_64.whl
+        pip install dist/ssapy-1.0.0-cp38-cp38-linux_x86_64.whl
     
         # List installed Python packages
         python3 -m pip list
@@ -144,7 +144,7 @@ jobs:
         git lfs install
         git lfs pull
         python3 -m build
-        pip install SSAPy/dist/ssapy-1.0.0-cp38-cp38-linux_x86_64.whl
+        pip install dist/ssapy-1.0.0-cp38-cp38-linux_x86_64.whl
         pip install -r docs/requirements.txt
         cd docs && sphinx-autogen source/api.rst -t source/_templates -o source/modules && cd ../
         cd docs && make html && make html
