+++ requirements.txt
@@ -26,3 +26,4 @@ pytest
 imageio
 tqdm
 build
+cibuildwheel
