+++ .github/workflows/ci.yml
@@ -51,22 +51,25 @@ jobs:
         # Install dependencies
         sudo apt-get update
         sudo apt-get install build-essential git git-lfs python3 python3-setuptools python3-venv graphviz cmake
-
+    
         # Initialize and update submodules
         git submodule update --init --recursive
-
+    
         # Install Git LFS and pull LFS objects
         git lfs install
         git lfs pull
-
+    
         # Upgrade pip and install Python dependencies
         python -m pip install --upgrade pip setuptools flake8
         python -m pip install -r requirements.txt
-
+    
+        # Get Python version dynamically
+        PYTHON_VERSION=$(python -c "import sys; print(f'cp{sys.version_info.major}{sys.version_info.minor}')")
+    
         # Build and install the project
         python3 -m build
-        pip install dist/ssapy-1.0.0-cp38-cp38-linux_x86_64.whl
-
+        pip install dist/ssapy-1.0.0-${PYTHON_VERSION}-${PYTHON_VERSION}-linux_x86_64.whl
+    
         # List installed Python packages
         python -m pip list
     - name: Install dependencies (MacOS)
@@ -86,10 +89,13 @@ jobs:
         # Upgrade pip and install Python dependencies
         python3 -m pip install --upgrade pip setuptools flake8
         python3 -m pip install -r requirements.txt
+
+        # Get Python version dynamically
+        PYTHON_VERSION=$(python -c "import sys; print(f'cp{sys.version_info.major}{sys.version_info.minor}')")
     
         # Build and install the project
         cibuildwheel --platform macos
-        pip install wheelhouse/ssapy-1.0.0-cp38-cp38-macosx_10_9_x86_64.whl
+        pip install wheelhouse/ssapy-1.0.0-${PYTHON_VERSION}-${PYTHON_VERSION}-macosx_10_9_x86_64.whl
 
         # List installed Python packages
         python3 -m pip list
@@ -144,7 +150,8 @@ jobs:
         git lfs install
         git lfs pull
         python3 -m build
-        pip install dist/ssapy-1.0.0-cp38-cp38-linux_x86_64.whl
+        pip install dist/ssapy-1.0.0-cp312-cp312-linux_x86_64.whl
+        
         pip install -r docs/requirements.txt
         cd docs && sphinx-autogen source/api.rst -t source/_templates -o source/modules && cd ../
         cd docs && make html && make html
