+++ docs/source/conf.py
@@ -3,7 +3,9 @@
 # For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
-
+import os
+import sys
+sys.path.insert(0, os.path.abspath('../../'))  # Adjust based on your project layout
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
