+++ docs/source/conf.py
@@ -5,7 +5,7 @@
 
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../'))  # Adjust based on your project layout
+sys.path.insert(0, os.path.abspath('../../ssapy/'))
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
