+++ docs/source/conf.py
@@ -34,7 +34,6 @@ extensions = [
     'sphinx.ext.viewcode',
 ]
 
-
 def linkcode_resolve(domain, info):
     if domain != 'py':
         return None
