+++ CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.5)
 cmake_policy(SET CMP0048 NEW)
 cmake_policy(SET CMP0063 NEW)
 
-project(ssapy VERSION 1.0 LANGUAGES CXX)
+project(_ssapy VERSION 1.0 LANGUAGES CXX)
 
 # Compiler settings
 set(CMAKE_CXX_STANDARD 11)
@@ -12,7 +12,7 @@ set(CMAKE_CXX_VISIBILITY_PRESET hidden)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 set(CMAKE_MACOSX_RPATH 1)
 
-# Output directory
+# Output directories
 if(NOT CMAKE_LIBRARY_OUTPUT_DIRECTORY)
     set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 endif()
@@ -20,7 +20,7 @@ endif()
 # Find Python
 find_package(Python3 REQUIRED COMPONENTS Interpreter Development)
 
-# Bring in pybind11
+# Include pybind11 via FetchContent
 include(FetchContent)
 FetchContent_Declare(
   pybind11
@@ -29,18 +29,19 @@ FetchContent_Declare(
 )
 FetchContent_MakeAvailable(pybind11)
 
-# Include your headers
+# Include dirs
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
 
-# Build static C++ logic (optional)
-add_library(ssapy STATIC src/ssapy.cpp)
-set_target_properties(ssapy PROPERTIES POSITION_INDEPENDENT_CODE ON)
-
-# Build Python extension module
 pybind11_add_module(_ssapy pysrc/ssapy.cpp)
+add_library(ssapy STATIC src/ssapy.cpp)
 target_link_libraries(_ssapy PUBLIC ssapy)
 
-# Install into the ssapy/ Python package directory
 install(TARGETS _ssapy
     LIBRARY DESTINATION ssapy
+)
+
+install(TARGETS _ssapy
+    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
 )
