+++ docs/source/conf.py
@@ -6,7 +6,6 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('../../ssapy'))
-autodoc_mock_imports = ['ssapy']
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
