+++ docs/requirements.txt
@@ -6,6 +6,7 @@ sphinx-automodapi
 sphinx_rtd_theme
 sphinxcontrib-bibtex
 sphinx-autobuild
+docutils<0.21
 nbsphinx
 numpy
 astropy
@@ -16,3 +17,5 @@ scipy
 matplotlib
 pandas
 scikit-learn
+sys
+os
