+++ requirements.txt
@@ -20,4 +20,3 @@ imageio
 tqdm
 build
 cibuildwheel
-os-sys
