+++ docs/requirements.txt
@@ -8,4 +8,3 @@ sphinxcontrib-bibtex
 sphinx-autobuild
 docutils<0.21
 nbsphinx
-os-sys
