### 2025-04-29 d410875
fixing setup.py
