### 2025-04-29 5a8e378
cmake edits
