+++ MANIFEST.in
@@ -1,13 +1,15 @@
 include README.rst
 include LICENSE
+include pyproject.toml
+include CMakeLists.txt
+
 include include/*.h
 include pysrc/*.cpp
 include src/*.cpp
-include pyproject.toml
-include CMakeLists.txt
+
 recursive-include pybind11/include/ *
 include pybind11/CMakeLists.txt
 include pybind11/LICENSE
 recursive-include pybind11/tools *
-exclude ssapy/data/*
+
 recursive-exclude ssapy/data *
