+++ ssapy/__init__.py
@@ -5,6 +5,7 @@ import os
 datadir = os.path.join(os.path.dirname(__file__), "data")
 
 from . import _ssapy
+from _ssapy import accel
 
 from .orbit import Orbit, EarthObserver, OrbitalObserver
 from .propagator import (
