+++ docs/source/api.rst
@@ -14,7 +14,6 @@ and/or heavy dependencies.  Below lists the packages available in the ``ssapy`` 
       :toctree: modules
       :template: automodapi_templ.rst
 
-      ssapy.accel
       ssapy.body
       ssapy.compute
       ssapy.constants
@@ -31,3 +30,4 @@ and/or heavy dependencies.  Below lists the packages available in the ``ssapy`` 
       ssapy.rvsampler
       ssapy.simple
       ssapy.utils
+      ssapy.accel
