+++ docs/source/conf.py
@@ -6,6 +6,8 @@
 import os
 import sys
 import ssapy
+sys.path.insert(0, os.path.abspath(os.path.join('..', '..', 'src')))
+
 # sys.path.insert(0, os.path.abspath('../../ssapy'))
 # print(sys.path)
 
