+++ docs/source/conf.py
@@ -6,10 +6,8 @@
 import os
 import sys
 import ssapy
-sys.path.insert(0, os.path.abspath(os.path.join('..', '..', 'src')))
-
-# sys.path.insert(0, os.path.abspath('../../ssapy'))
-# print(sys.path)
+sys.path.insert(0, os.path.abspath(os.path.join('..', '..', 'ssapy')))
+print(sys.path)
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
