+++ docs/source/api.rst
@@ -13,7 +13,8 @@ and/or heavy dependencies.  Below lists the packages available in the ``ssapy`` 
    .. autosummary::
       :toctree: modules
       :template: automodapi_templ.rst
-
+      
+      ssapy.accel
       ssapy.body
       ssapy.compute
       ssapy.constants
@@ -30,4 +31,3 @@ and/or heavy dependencies.  Below lists the packages available in the ``ssapy`` 
       ssapy.rvsampler
       ssapy.simple
       ssapy.utils
-      ssapy.accel
