### 2025-04-29 0d6dad1
reworking cmake
