+++ ssapy/accel.py
@@ -1,6 +1,7 @@
 """
 Classes for modeling accelerations.
 """
+from . import _ssapy
 
 import numpy as np
 
@@ -357,7 +358,6 @@ class AccelDrag(Accel):
         (area, mass, CR)
     """
     def __init__(self, recalc_threshold=86400 * 30, **defaultkw):
-        from . import _ssapy
 
         self.recalc_threshold = recalc_threshold
         self._t = None
