+++ ssapy/ellipsoid.py
@@ -7,9 +7,9 @@ parameter f, but that's good enough for simple Earth models.
 """
 
 import numpy as np
-# from . import _ssapy
+import ssapy._ssapy as _ssapy
 
-from ._ssapy import Ellipsoid
+from _ssapy import Ellipsoid
 from .utils import continueClass
 
 
