### 2025-04-30 8617b19
i hate cmake
