+++ CMakeLists.txt
@@ -25,7 +25,7 @@ endif()
 
 
 # Find Python
-find_package(Python3 REQUIRED COMPONENTS Interpreter Development)
+find_package(Python3 REQUIRED COMPONENTS Interpreter Development.Module)
 set(PYTHON_MODULE_INSTALL_DIR "${Python3_SITEARCH}/ssapy")
 
 # Include pybind11 using FetchContent
@@ -46,9 +46,8 @@ pybind11_add_module(_ssapy pysrc/ssapy.cpp)
 # Ensure the Python module is installed in the correct location
 set(CMAKE_POSITION_INDEPENDENT_CODE True)
 install(TARGETS _ssapy
-    LIBRARY DESTINATION ${PYTHON_MODULE_INSTALL_DIR}  # Install the compiled module in Python's site-packages
+    LIBRARY DESTINATION ssapy # Install the compiled module in Python's site-packages
 )
-
 
 # Define a static library target (ssapy)
 add_library(ssapy STATIC src/ssapy.cpp)
