+++ setup.py
@@ -40,7 +40,6 @@ class CMakeBuild(build_ext):
             cmake_args += ['-DCMAKE_VERBOSE_MAKEFILE=1']
         subprocess.check_call(['cmake', ext.sourcedir] + cmake_args, cwd=self.build_temp, env=env)
         subprocess.check_call(['cmake', '--build', '.'] + build_args, cwd=self.build_temp)
-        subprocess.check_call(['cmake', '--install', '.'], cwd=self.build_temp)
 
 class PostInstallCommand(install):
     """Post-installation to download the ssapy/data directory."""
