+++ MANIFEST.in
@@ -11,5 +11,7 @@ recursive-include pybind11/include/ *
 include pybind11/CMakeLists.txt
 include pybind11/LICENSE
 recursive-include pybind11/tools *
+recursive-include ssapy *.so
+
 
 recursive-exclude ssapy/data *
