+++ pyproject.toml
@@ -54,7 +54,7 @@ packages = [
 ]
 
 [tool.poetry.include]
-files = ["ssapy/_ssapy.cpython-38-x86_64-linux-gnu.so"
+files = ["ssapy/_ssapy.cpython-38-x86_64-linux-gnu.so"]
 
 [project.urls]
 Homepage = "https://github.com/LLNL/SSAPy"
