+++ CMakeLists.txt
@@ -48,6 +48,9 @@ add_library(ssapy STATIC src/ssapy.cpp)
 
 # Link the static library to the Python module
 target_link_libraries(_ssapy PUBLIC ssapy)
+set_target_properties(_ssapy PROPERTIES
+    LIBRARY_OUTPUT_DIRECTORY "${PROJECT_SOURCE_DIR}/src/ssapy"
+)
 
 set(CMAKE_POSITION_INDEPENDENT_CODE True)
 install(TARGETS _ssapy DESTINATION ssapy) 
