### 2025-04-30 3eee2f5
remove postinstallcommand
