+++ setup.py
@@ -73,8 +73,8 @@ setup(
     version='1.0.0',
     ext_modules=[CMakeExtension("_ssapy", sourcedir=".")],
     cmdclass={"build_ext": CMakeBuild},
-    packages=find_packages(where="src"),
-    package_dir={"": "src"},
+    packages=find_packages(),
+    package_dir={"": "."},
     license='MIT',
     zip_safe=False,
     include_package_data=True,
