+++ CMakeLists.txt
@@ -1,56 +1,45 @@
-# Minimum required CMake version
 cmake_minimum_required(VERSION 3.15)
 
-# Set policies for compatibility
-cmake_policy(SET CMP0048 NEW)  # Project version policy
-cmake_policy(SET CMP0063 NEW)  # Visibility preset policy
-
-# Project settings
+# Set project
 project(ssapy VERSION 1.0 LANGUAGES CXX)
 
-# Compiler settings
-set(CMAKE_CXX_STANDARD 17)                # Use C++17 standard
-set(CMAKE_CXX_STANDARD_REQUIRED ON)       # Require C++17
-set(CMAKE_CXX_EXTENSIONS OFF)             # Disable compiler-specific extensions
-set(CMAKE_CXX_VISIBILITY_PRESET hidden)   # Hide symbols by default
-set(CMAKE_POSITION_INDEPENDENT_CODE ON)   # Enable position-independent code (PIC)
+# Use C++17
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_CXX_EXTENSIONS OFF)
 
-# macOS-specific settings
+# Optional: Hide unused symbols unless explicitly exported
+set(CMAKE_CXX_VISIBILITY_PRESET hidden)
+set(CMAKE_POSITION_INDEPENDENT_CODE ON)
+
+# Required for linking static libs into shared modules
 set(CMAKE_MACOSX_RPATH 1)
 
-# Output directories
-if(NOT CMAKE_LIBRARY_OUTPUT_DIRECTORY)
-    set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)  # Set default output directory for libraries
-endif()
-
-
-# Find Python
+# Find Python with development headers and module support
 find_package(Python3 REQUIRED COMPONENTS Interpreter Development.Module)
-set(PYTHON_MODULE_INSTALL_DIR "${Python3_SITEARCH}/ssapy")
 
-# Include pybind11 using FetchContent
-include(FetchContent)
-FetchContent_Declare(
-    pybind11
-    GIT_REPOSITORY https://github.com/pybind/pybind11.git
-    GIT_TAG        v2.11.1
+# Find pybind11 (assumes you vendored or added as submodule)
+add_subdirectory(pybind11)
+
+# Optional: Add your own library to link
+add_library(ssapy STATIC
+    src/ssapy.cpp
+    include/ssapy.h
 )
-FetchContent_MakeAvailable(pybind11)
+target_include_directories(ssapy PUBLIC include)
 
-# Include directories
-include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
+# Python extension target
+pybind11_add_module(_ssapy
+    pysrc/ssapy.cpp        # your pybind11 bindings file
+)
 
-# Define the Python module target (_ssapy)
-pybind11_add_module(_ssapy pysrc/ssapy.cpp)
+# Link your custom library to the Python module
+target_link_libraries(_ssapy PRIVATE ssapy)
 
-# Define a static library target (ssapy)
-add_library(ssapy STATIC src/ssapy.cpp)
-
-# Link the static library to the Python module
-target_link_libraries(_ssapy PUBLIC ssapy)
+# Place the shared object directly into the ssapy Python package
 set_target_properties(_ssapy PROPERTIES
     LIBRARY_OUTPUT_DIRECTORY "${PROJECT_SOURCE_DIR}/ssapy"
 )
 
-set(CMAKE_POSITION_INDEPENDENT_CODE True)
-install(TARGETS _ssapy DESTINATION ssapy) 
[…]
