+++ .github/workflows/ci.yml
@@ -99,6 +99,11 @@ jobs:
 
         # List installed Python packages
         python3 -m pip list
+    - name: Debug installed ssapy
+      run: |
+        python -c "import ssapy; print('ssapy is at:', ssapy.__file__)"
+        ls -l $(python -c "import ssapy; import os; print(os.path.dirname(ssapy.__file__))")
+
     #- name: Install dependencies (Windows)
     #  if: matrix.os == 'windows-latest'
     #  run: |
