+++ ssapy/__init__.py
@@ -4,8 +4,6 @@ ensure_data_downloaded()
 import os
 datadir = os.path.join(os.path.dirname(__file__), "data")
 
-from . import _ssapy
-
 from .orbit import Orbit, EarthObserver, OrbitalObserver
 from .propagator import (
     KeplerianPropagator, SeriesPropagator, RK4Propagator, SGP4Propagator,
