+++ .github/workflows/ci.yml
@@ -67,6 +67,7 @@ jobs:
         PYTHON_VERSION=$(python -c "import sys; print(f'cp{sys.version_info.major}{sys.version_info.minor}')")
     
         # Build and install the project
+        rm -rf build dist *.egg-info
         python3 -m build
         pip install dist/ssapy-1.0.0-${PYTHON_VERSION}-${PYTHON_VERSION}-linux_x86_64.whl
     
@@ -94,6 +95,7 @@ jobs:
         PYTHON_VERSION=$(python -c "import sys; print(f'cp{sys.version_info.major}{sys.version_info.minor}')")
     
         # Build and install the project
+        rm -rf build dist *.egg-info
         python3 -m cibuildwheel --platform macos
         pip install wheelhouse/ssapy-1.0.0-${PYTHON_VERSION}-${PYTHON_VERSION}-macosx_10_9_x86_64.whl
 
