### 2025-04-30 #7699 @andrea-sesta
Hi @fruzsinaagocs. Thanks for the invitation! I’d be happy to serve as a reviewer for this JOSS submission.
