+++ .github/workflows/ci.yml
@@ -64,11 +64,8 @@ jobs:
         python -m pip install -r requirements.txt
     
         # Build and install the project
-        rm -rf build dist *.egg-info
-        python3 -m build
-        pip install dist/ssapy-1.0.0-${PYTHON_VERSION}-${PYTHON_VERSION}-linux_x86_64.whl
-    
-        # List installed Python packages
+        python setup.py build
+        python setup.py install
         python -m pip list
     - name: Install dependencies (MacOS)
       if: matrix.os == 'macOS-latest'
@@ -89,12 +86,9 @@ jobs:
         python3 -m pip install -r requirements.txt
 
         # Build and install the project
-        rm -rf build dist *.egg-info
-        python3 -m cibuildwheel --platform macos
-        pip install wheelhouse/ssapy-1.0.0-${PYTHON_VERSION}-${PYTHON_VERSION}-macosx_10_9_x86_64.whl
-
-        # List installed Python packages
-        python3 -m pip list
+        python setup.py build
+        python setup.py install
+        python -m pip list
     - name: Debug installed ssapy
       run: |
         python -c "import ssapy; print('ssapy is at:', ssapy.__file__)"
