+++ setup.py
@@ -52,10 +52,11 @@ class CMakeBuild(build_ext):
 setup(
     name='ssapy',
     version='1.0.0',
-    ext_modules=[CMakeExtension("_ssapy", sourcedir=".")],
+    ext_modules=[CMakeExtension("_ssapy")],
     cmdclass={"build_ext": CMakeBuild},
     packages=find_packages(),
-    package_dir={"": "."},
+    # package_dir={"": "."},
+    package_data={'ssapy': ['*.so']},
     license='MIT',
     zip_safe=False,
     include_package_data=True,
