+++ .github/workflows/ci.yml
@@ -62,6 +62,9 @@ jobs:
         # Upgrade pip and install Python dependencies
         python -m pip install --upgrade pip setuptools flake8
         python -m pip install -r requirements.txt
+
+        # Get Python version dynamically
+        PYTHON_VERSION=$(python -c "import sys; print(f'cp{sys.version_info.major}{sys.version_info.minor}')")
     
         # Build and install the project
         rm -rf build dist *.egg-info
@@ -87,6 +90,9 @@ jobs:
         # Upgrade pip and install Python dependencies
         python3 -m pip install --upgrade pip setuptools flake8
         python3 -m pip install -r requirements.txt
+
+        # Get Python version dynamically
+        PYTHON_VERSION=$(python -c "import sys; print(f'cp{sys.version_info.major}{sys.version_info.minor}')")
 
         # Build and install the project
         rm -rf build dist *.egg-info
