+++ ssapy/ellipsoid.py
@@ -7,15 +7,10 @@ parameter f, but that's good enough for simple Earth models.
 """
 
 import numpy as np
- 
 from .utils import continueClass
+from . import _ssapy
+from ._ssapy import Ellipsoid   
 
-def myfunc():
-    from . import _ssapy
-    from ._ssapy import Ellipsoid   
-    return _ssapy.call(...)
-
-myfunc()
 
 @continueClass
 class Ellipsoid:
