+++ .github/workflows/ci.yml
@@ -62,14 +62,11 @@ jobs:
         # Upgrade pip and install Python dependencies
         python -m pip install --upgrade pip setuptools flake8
         python -m pip install -r requirements.txt
-
-        # Get Python version dynamically
-        PYTHON_VERSION=$(python -c "import sys; print(f'cp{sys.version_info.major}{sys.version_info.minor}')")
     
         # Build and install the project
         rm -rf build dist *.egg-info
         python3 -m build
-        pip install dist/ssapy-1.0.0-${PYTHON_VERSION}-${PYTHON_VERSION}-linux_x86_64.whl
+        pip install dist/*.whl
     
         # List installed Python packages
         python -m pip list
@@ -78,26 +75,23 @@ jobs:
       run: |
         # Update Homebrew and install dependencies
         brew update
-        brew install git git-lfs python3 graphviz cmake
-    
+        brew install git python3 cmake
+        
         # Initialize and update submodules
         git submodule update --init --recursive
-    
-        # Install Git LFS and pull LFS objects
-        git lfs install
-        git lfs pull
-    
+        
         # Upgrade pip and install Python dependencies
         python3 -m pip install --upgrade pip setuptools flake8
         python3 -m pip install -r requirements.txt
-
-        # Get Python version dynamically
-        PYTHON_VERSION=$(python -c "import sys; print(f'cp{sys.version_info.major}{sys.version_info.minor}')")
-
+    
         # Build and install the project
         rm -rf build dist *.egg-info
+        export CIBW_PLATFORM="macos"
+        export CIBW_ARCHS="x86_64 arm64"
         python3 -m cibuildwheel --platform macos
-        pip install wheelhouse/ssapy-1.0.0-${PYTHON_VERSION}-${PYTHON_VERSION}-macosx_10_9_x86_64.whl
+    
+        # Install the generated wheel
+        pip install wheelhouse/*.whl
 
         # List installed Python packages
         python3 -m pip list
