+++ CMakeLists.txt
@@ -1,32 +1,47 @@
-cmake_minimum_required(VERSION 3.15)
+cmake_minimum_required(VERSION 3.5)
+cmake_policy(SET CMP0048 NEW)
+cmake_policy(SET CMP0063 NEW)
 
-# Set project
-project(ssapy VERSION 1.0 LANGUAGES CXX)
+project(_ssapy VERSION 1.0 LANGUAGES CXX)
 
-# Use C++17
-set(CMAKE_CXX_STANDARD 17)
+# Compiler settings
+set(CMAKE_CXX_STANDARD 11)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
-
-# Optional: Hide unused symbols unless explicitly exported
 set(CMAKE_CXX_VISIBILITY_PRESET hidden)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
-
-# Required for linking static libs into shared modules
 set(CMAKE_MACOSX_RPATH 1)
 
-# Find Python with development headers and module support
-find_package(Python3 REQUIRED COMPONENTS Interpreter Development.Module)
+# Output directories
+if(NOT CMAKE_LIBRARY_OUTPUT_DIRECTORY)
+    set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
+endif()
 
-# Find pybind11 (assumes you vendored or added as submodule)
-add_subdirectory(pybind11)
+# Find Python
+find_package(Python3 REQUIRED COMPONENTS Interpreter Development)
 
-# Optional: Add your own library to link
-add_library(ssapy STATIC
-    src/ssapy.cpp
-    include/ssapy.h
+# Include pybind11 via FetchContent
+include(FetchContent)
+FetchContent_Declare(
+  pybind11
+  GIT_REPOSITORY https://github.com/pybind/pybind11.git
+  GIT_TAG        v2.11.1
 )
-target_include_directories(ssapy PUBLIC include)
+FetchContent_MakeAvailable(pybind11)
 
-# Include the Python-specific logic
-include(python.cmake)
+# Include dirs
+include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
+
+pybind11_add_module(_ssapy pysrc/ssapy.cpp)
+add_library(ssapy STATIC src/ssapy.cpp)
+target_link_libraries(_ssapy PUBLIC ssapy)
+
+install(TARGETS _ssapy
+    LIBRARY DESTINATION ssapy
+)
+
+install(TARGETS _ssapy
+    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
+)
