### 2025-04-30 edb1a01
utils edits
