### 2025-04-30 fde527a
cmake dir
