+++ .github/workflows/ci.yml
@@ -86,16 +86,11 @@ jobs:
         python3 -m pip install -r requirements.txt
     
         # Ensure delocate is installed
-        python3 -m pip install --upgrade cibuildwheel delocate
+        python3 -m pip install --upgrade cibuildwheel delocate        
+        python -m cibuildwheel --platform macos
 
-        # Set environment variables
-        export MACOSX_DEPLOYMENT_TARGET=10.9
-        export CIBW_PLATFORM="macos"
-        export CIBW_ARCHS="x86_64"
-        export CIBW_REPAIR_WHEEL_COMMAND="delocate-wheel --require-archs {delocate_archs} -w {dest_dir} -v {wheel}"
-
-        # Build wheels
-        python3 -m cibuildwheel --platform macos
+        # Check the wheel for missing shared libraries
+        find wheelhouse -name '*.whl' -exec delocate-listdeps {} \;
             
         # Install the generated wheel
         pip install wheelhouse/*.whl
