+++ ssapy/ellipsoid.py
@@ -7,9 +7,9 @@ parameter f, but that's good enough for simple Earth models.
 """
 
 import numpy as np
-
-from ._ssapy import Ellipsoid
 from .utils import continueClass
+from . import _ssapy
+from ._ssapy import Ellipsoid   
 
 
 @continueClass
