+++ pyproject.toml
@@ -35,7 +35,7 @@ dependencies = [
     ]
 
 [build-system]
-requires = ["setuptools>=42", "wheel", "scikit-build", "pybind11"]
+requires = ["setuptools>=42", "wheel", "scikit-build", "pybind11", "cmake"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
@@ -47,6 +47,14 @@ ssapy = ["*.cpp", "*.h"]
 
 [tool.setuptools.exclude-package-data]
 ssapy = ["data/*"]
+
+[tool.poetry]
+packages = [
+    { include = "ssapy" },
+]
+
+[tool.poetry.include]
+files = ["ssapy/_ssapy.cpython-38-x86_64-linux-gnu.so"]
 
 [project.urls]
 Homepage = "https://github.com/LLNL/SSAPy"
