### 2025-05-01 fb01b36
test and cmake updates
