+++ tests/__init__.py
@@ -1,0 +1,2 @@
+from . import ssapy_test_helpers
+from . import orekit
