+++ tests/__init__.py
@@ -1,2 +1,1 @@
 from . import ssapy_test_helpers
-from . import orekit
