+++ requirements.txt
@@ -20,3 +20,4 @@ tqdm
 build
 cibuildwheel
 delocate
+orekit-factory
