+++ tests/__init__.py
@@ -1,1 +1,0 @@
-from . import ssapy_test_helpers
