+++ tests/test_frame.py
@@ -4,6 +4,7 @@ import numpy as np
 from astropy.time import Time
 
 import ssapy
+from . import ssapy_test_helpers
 from ssapy_test_helpers import timer
 
 
