+++ tests/test_accel.py
@@ -4,6 +4,7 @@ import astropy.units as u
 
 import ssapy
 from ssapy.utils import norm, iers_interp
+from . import ssapy_test_helpers
 from ssapy_test_helpers import timer, sample_LEO_orbit, sample_GEO_orbit
 
 iers_interp(0.0)  # Prime the IERS interpolant cache
