+++ .github/workflows/ci.yml
@@ -138,7 +138,7 @@ jobs:
     - name: Test with pytest
       run: |
         # Run the specific test file first
-        python3 ${CI_PROJECT_DIR}/tests/test_orbit.py || true
+        pytest ${CI_PROJECT_DIR}/tests/test_orbit.py || true
         
         # Run pytest with coverage, excluding the specified test
         pytest --cov-report=xml --cov=ssapy --deselect=tests/test_taskvine.py || true
