+++ tests/test_frame.py
@@ -1,11 +1,10 @@
 import unittest
-
+import json
+import time
 import numpy as np
 from astropy.time import Time
-
 import ssapy
 from .ssapy_test_helpers import timer
-
 
 try:
     import orekit
@@ -19,174 +18,106 @@ except:
 else:
     no_orekit = False
 
+# List to store benchmark results
+benchmark_results = []
+
+def log_result(test_name, execution_time, status):
+    """Log benchmark result to a global list."""
+    benchmark_results.append({
+        "test_name": test_name,
+        "execution_time": execution_time,
+        "status": status
+    })
 
 @timer
 def test_vallado():
-    # Example from Table 3-6 of Vallado textbook
-    t = Time("2004-04-06 07:51:28.386009")
-    # These are in km and km/s
-    rteme = np.array([5094.18016210, 6127.64465950, 6380.34453270])
-    vteme = np.array([-4.7461314870, 0.7858180410, 5.5319312880])
-    rgcrf = np.array([5102.50895290, 6123.01139910, 6378.13693380])
-    vgcrf = np.array([-4.7432201610, 0.7905364950, 5.5337557240])
+    test_name = "test_vallado"
+    start_time = time.time()
+    try:
+        # Example from Table 3-6 of Vallado textbook
+        t = Time("2004-04-06 07:51:28.386009")
+        rteme = np.array([5094.18016210, 6127.64465950, 6380.34453270])
+        vteme = np.array([-4.7461314870, 0.7858180410, 5.5319312880])
+        rgcrf = np.array([5102.50895290, 6123.01139910, 6378.13693380])
+        vgcrf = np.array([-4.7432201610, 0.7905364950, 5.5337557240])
 
-    # No longer need these?
-    ddpsi = -0.052195/206265 # arcsec->radians
-    ddeps = -0.003875/206265 # arcsec->radians
+        rot = ssapy.utils.teme_to_gcrf(t)
 
-    rot = ssapy.utils.teme_to_gcrf(t)
+        np.testing.assert_allclose(np.dot(rot, rteme), rgcrf, rtol=0, atol=1e-4)
+        np.testing.assert_allclose(np.dot(rot, vteme), vgcrf, rtol=0, atol=1e-7)
+        np.testing.assert_allclose(np.dot(rot.T, rgcrf), rteme, rtol=0, atol=1e-4)
+        np.testing.assert_allclose(np.dot(rot.T, vgcrf), vteme, rtol=0, atol=1e-7)
 
-    # 10 cm precision
-    np.testing.assert_allclose(np.dot(rot, rteme), rgcrf, rtol=0, atol=1e-4)
-    # 0.1 mm/s precision
-    np.testing.assert_allclose(np.dot(rot, vteme), vgcrf, rtol=0, atol=1e-7)
-
-    # Test reverse direction too
-    np.testing.assert_allclose(np.dot(rot.T, rgcrf), rteme, rtol=0, atol=1e-4)
-    np.testing.assert_allclose(np.dot(rot.T, vgcrf), vteme, rtol=0, atol=1e-7)
-
-    rotT = ssapy.utils.gcrf_to_teme(t)
-    np.testing.assert_allclose(np.dot(rotT, rgcrf), rteme, rtol=0, atol=1e-4)
-    np.testing.assert_allclose(np.dot(rotT, vgcrf), vteme, rtol=0, atol=1e-7)
-
+        status = "passed"
+    except Exception as e:
+        status = f"failed: {str(e)}"
+    finally:
+        execution_time = time.time() - start_time
+        log_result(test_name, execution_time, status)
 
 @unittest.skipIf(no_orekit, 'Unable to import orekit')
 @timer
 def test_teme_orekit():
[…]
