+++ .github/workflows/ci.yml
@@ -137,11 +137,8 @@ jobs:
         pip install pytest pytest-cov pytest-xdist
     - name: Test with pytest
       run: |
-        # Run the specific test file first
-        pytest ${CI_PROJECT_DIR}/tests/test_orbit.py || true
-        
         # Run pytest with coverage, excluding the specified test
-        pytest --cov-report=xml --cov=ssapy --deselect=tests/test_taskvine.py || true
+        pytest --cov-report=xml --cov=ssapy || true
     - name: Verify coverage files
       run: |
         if [ -f coverage.xml ]; then
