+++ tests/test_frame.py
@@ -34,19 +34,31 @@ def test_vallado():
     test_name = "test_vallado"
     start_time = time.time()
     try:
-        # Example from Table 3-6 of Vallado textbook
         t = Time("2004-04-06 07:51:28.386009")
+        # These are in km and km/s
         rteme = np.array([5094.18016210, 6127.64465950, 6380.34453270])
         vteme = np.array([-4.7461314870, 0.7858180410, 5.5319312880])
         rgcrf = np.array([5102.50895290, 6123.01139910, 6378.13693380])
         vgcrf = np.array([-4.7432201610, 0.7905364950, 5.5337557240])
 
+        # No longer need these?
+        ddpsi = -0.052195/206265 # arcsec->radians
+        ddeps = -0.003875/206265 # arcsec->radians
+
         rot = ssapy.utils.teme_to_gcrf(t)
 
+        # 10 cm precision
         np.testing.assert_allclose(np.dot(rot, rteme), rgcrf, rtol=0, atol=1e-4)
+        # 0.1 mm/s precision
         np.testing.assert_allclose(np.dot(rot, vteme), vgcrf, rtol=0, atol=1e-7)
+
+        # Test reverse direction too
         np.testing.assert_allclose(np.dot(rot.T, rgcrf), rteme, rtol=0, atol=1e-4)
         np.testing.assert_allclose(np.dot(rot.T, vgcrf), vteme, rtol=0, atol=1e-7)
+
+        rotT = ssapy.utils.gcrf_to_teme(t)
+        np.testing.assert_allclose(np.dot(rotT, rgcrf), rteme, rtol=0, atol=1e-4)
+        np.testing.assert_allclose(np.dot(rotT, vgcrf), vteme, rtol=0, atol=1e-7)
 
         status = "passed"
     except Exception as e:
@@ -66,6 +78,15 @@ def test_teme_orekit():
         from org.orekit.time import AbsoluteDate
         from org.orekit.utils import PVCoordinates
 
+        # Some cheap utilities to convert between java PVCoordinates and numpy vectors
+        def v3toarr(v3):
+            p3 = v3.getPosition()
+            return np.array([p3.x, p3.y, p3.z])
+
+        def arrtov3(arr):
+            return PVCoordinates(Vector3D(float(arr[0]), float(arr[1]), float(arr[2])))
+
+
         TEME = FramesFactory.getTEME()
         GCRF = FramesFactory.getGCRF()
 
@@ -79,8 +100,8 @@ def test_teme_orekit():
             transform = TEME.getTransformTo(GCRF, oreDate)
 
             for rteme_, rgcrf_ in zip(rteme, rgcrf):
-                rtest = transform.transformPVCoordinates(PVCoordinates(Vector3D(*rteme_))).getPosition()
-                np.testing.assert_allclose(rtest, rgcrf_, rtol=0, atol=1e-3)
+                rtest = v3toarr(transform.transformPVCoordinates(arrtov3(rteme_)))
+                np.testing.assert_allclose(rtest, rgcrf_, rtol=0, atol=1e-3)  # 1m precision
 
         status = "passed"
     except Exception as e:
@@ -94,14 +115,100 @@ def test_MG_5_1():
     test_name = "test_MG_5_1"
     start_time = time.time()
     try:
-        # Exercise 5.1 from Montenbruck and Gill
+        """Exercise 5.1 from Montenbruck and Gill"""
         import astropy.units as u
         t0 = Time("1999-03-04T00:00:00", scale='utc')
+        mjd_tt_j2000 = 51544.5
         mjd_tt = ssapy.utils._gpsToTT(t0.gps)
         dut1_mjd = t0.ut1.mjd - t0.tt.mjd
[…]
