+++ .github/workflows/ci.yml
@@ -163,7 +163,7 @@ jobs:
       run: |
         pip show ssapy
         pip install -r docs/requirements.txt
-        python /ssapy/update_benchmarks.py
+        python ./ssapy/update_benchmarks.py
 
         cd docs
         # Ensure necessary directories exist
