+++ .github/workflows/ci.yml
@@ -138,7 +138,7 @@ jobs:
     - name: Test with pytest
       run: |
         # Run pytest with coverage, excluding the specified test
-        pytest --cov-report=xml --cov=ssapy || true
+        pytest --cov-report=xml --cov=ssapy --dist=loadscope || true
     - name: Verify coverage files
       run: |
         if [ -f coverage.xml ]; then
@@ -164,10 +164,7 @@ jobs:
         pip show ssapy
         pip install -r docs/requirements.txt
 
-        cd ssapy
-        python update_benchmarks.py
-
-        cd ../docs
+        cd docs
         # Ensure necessary directories exist
         mkdir -p source/modules source/_templates
         
