+++ setup.cfg
@@ -6,3 +6,5 @@ addopts = -n 8 --junitxml=pytest.xml
 filterwarnings =
   ignore::DeprecationWarning:importlib:
   ignore::DeprecationWarning:emcee:
+[coverage:run]
+source = ssapy
