+++ .github/workflows/ci.yml
@@ -138,8 +138,8 @@ jobs:
     - name: Test with pytest
       run: |
         # Run the specific test file first
-        python3 ${CI_PROJECT_DIR}/tests/test_orbit.py || true
-        
+        pytest ${CI_PROJECT_DIR}/tests/test_orbit.py || true
+
         # Run pytest with coverage, excluding the specified test
         pytest --cov-report=xml --cov=ssapy --dist=loadscope || true
     - name: Verify coverage files
