+++ .github/workflows/ci.yml
@@ -138,7 +138,7 @@ jobs:
     - name: Test with pytest
       run: |
         # Run the specific test file first
-        pytest ${CI_PROJECT_DIR}/tests/test_orbit.py || true
+        python3 ${CI_PROJECT_DIR}/tests/test_orbit.py || true
 
         # Run pytest with coverage, excluding the specified test
         pytest --cov-report=xml --cov=ssapy --dist=loadscope || true
@@ -167,6 +167,7 @@ jobs:
         pip show ssapy
         pip install -r docs/requirements.txt
 
+        python3 ${CI_PROJECT_DIR}/ssapy/update_benchmarks.py
         cd docs
         # Ensure necessary directories exist
         mkdir -p source/modules source/_templates
