### 2025-05-02 #8147 @editorialbot
[REVIEW]: SSAPy - Space Situational Awareness for Python
**Submitting author:** @astrophpeter (<a href="http://orcid.org/0000-0002-1052-6749">Peter McGill</a>)
**Repository:** https://github.com/LLNL/SSAPy
**Branch with paper.md** (empty if default branch):
**Version:** v1.1
