+++ .github/workflows/ci.yml
@@ -167,7 +167,7 @@ jobs:
         pip show ssapy
         pip install -r docs/requirements.txt
 
-        python3 ${CI_PROJECT_DIR}/ssapy/update_benchmarks.py
+        # python3 ${CI_PROJECT_DIR}/ssapy/update_benchmarks.py
         cd docs
         # Ensure necessary directories exist
         mkdir -p source/modules source/_templates
