+++ .github/workflows/ci.yml
@@ -138,7 +138,7 @@ jobs:
     - name: Test with pytest
       run: |
         # Run the specific test file first
-        python3 ${CI_PROJECT_DIR}/tests/test_orbit.py 
+        python3 ${CI_PROJECT_DIR}/tests/test_plots.py 
 
         # Run pytest with coverage, excluding the specified test
         pytest --cov-report=xml --cov=ssapy || true
