### 2025-05-02 ea982ac
maybe the cfg was right
