+++ setup.cfg
@@ -6,5 +6,3 @@ addopts = -n 8 --junitxml=pytest.xml
 filterwarnings =
   ignore::DeprecationWarning:importlib:
   ignore::DeprecationWarning:emcee:
-[coverage:run]
-source = ssapy
