+++ tests/test_linker.py
@@ -7,7 +7,7 @@ from astropy.coordinates import Longitude, Latitude
 from astropy.table import QTable
 
 import ssapy
-from ssapy_test_helpers import timer
+from .ssapy_test_helpers import timer
 from ssapy.constants import RGEO, VGEO
 
 def _mock_chain_from_rv(r, v, nSteps=10):
