+++ .github/workflows/ci.yml
@@ -137,11 +137,8 @@ jobs:
         pip install pytest pytest-cov pytest-xdist
     - name: Test with pytest
       run: |
-        # Run the specific test file first
-        python3 ${CI_PROJECT_DIR}/tests/test_orbit.py || true
-        
         # Run pytest with coverage, excluding the specified test
-        pytest --cov-report=xml --cov=ssapy --deselect=tests/test_taskvine.py || true
+        pytest --cov-report=xml --cov=ssapy || true
     - name: Verify coverage files
       run: |
         if [ -f coverage.xml ]; then
@@ -167,8 +164,8 @@ jobs:
         pip show ssapy
         pip install -r docs/requirements.txt
 
+        # python3 ${CI_PROJECT_DIR}/ssapy/update_benchmarks.py
         cd docs
-
         # Ensure necessary directories exist
         mkdir -p source/modules source/_templates
         
