+++ README.rst
@@ -1,7 +1,7 @@
 SSAPy - Space Situational Awareness for Python
 ==============================================
 
-|ci_badge|  |docs_badge| |codecov_badge| 
+|ci_badge|  |docs_badge| |codecov_badge| |joss_badge|
 
 .. |ci_badge| image:: https://github.com/LLNL/SSAPy/actions/workflows/ci.yml/badge.svg
     :target: https://github.com/LLNL/SSAPy/actions/workflows/ci.yml
@@ -11,6 +11,9 @@ SSAPy - Space Situational Awareness for Python
 
 .. |codecov_badge| image:: https://codecov.io/gh/LLNL/SSAPy/branch/main/graph/badge.svg
     :target: https://codecov.io/gh/LLNL/SSAPy
+
+.. |joss_badge| .. image:: https://joss.theoj.org/papers/a629353cbdd8d64a861bb807e12c5d06/status.svg
+   :target: https://joss.theoj.org/papers/a629353cbdd8d64a861bb807e12c5d06
 
 `SSAPy <https://github.com/LLNL/SSAPy>`_ is a fast, flexible, high-fidelity orbital modeling and analysis tool for orbits spanning from low-Earth orbit into the cislunar regime, and includes the following:
 
