### 2025-05-05 9bb3107
remove pip show
