+++ tests/test_tracks.py
@@ -1,0 +1,309 @@
+import pytest
+import numpy as np
+from unittest.mock import MagicMock
+import ssapy
+from ssapy.correlate_tracks import CircVelocityPrior, ZeroRadialVelocityPrior, GaussPrior, VolumeDistancePrior, orbit_to_param, partial, make_param_guess
+from ssapy.constants import EARTH_MU, RGEO
+from .ssapy_test_helpers import sample_LEO_orbit 
+from ssapy.utils import normed
+
+####################
+## TESTING PRIORS ##
+####################
+def test_circ_velocity_prior_init():
+    sigma = np.log(2) / 3.5
+    prior = CircVelocityPrior(sigma=sigma)
+    assert prior.sigma == sigma
+
+def test_circ_velocity_prior_call():
+    t = 0  # Example time
+    orbit = sample_LEO_orbit(t)  # Sample a LEO orbit
+    distance = np.linalg.norm(orbit.r)  # Calculate distance from position vector
+    prior = CircVelocityPrior(sigma=np.log(2) / 3.5)
+
+    # Calculate expected values
+    loggmorv2 = np.log(EARTH_MU / np.linalg.norm(orbit.r) / np.dot(orbit.v, orbit.v))
+    chi0 = loggmorv2 / prior.sigma
+    expected_logprior = -0.5 * chi0**2
+
+    result = prior(orbit, distance)
+    assert pytest.approx(result[0]) == expected_logprior
+
+def test_circ_velocity_prior_call_chi():
+    t = 0  # Example time
+    orbit = sample_LEO_orbit(t)  # Sample a LEO orbit
+    distance = np.linalg.norm(orbit.r)  # Calculate distance from position vector
+    prior = CircVelocityPrior(sigma=np.log(2) / 3.5)
+
+    # Calculate expected values
+    loggmorv2 = np.log(EARTH_MU / np.linalg.norm(orbit.r) / np.dot(orbit.v, orbit.v))
+    chi0 = loggmorv2 / prior.sigma
+    expected_chi = chi0
+
+    # Call the method with chi=True and verify output
+    result = prior(orbit, distance, chi=True)
+    assert pytest.approx(result[0]) == expected_chi
+
+# Test ZeroRadialVelocityPrior initialization
+def test_zero_radial_velocity_prior_init():
+    sigma = 0.25
+    prior = ZeroRadialVelocityPrior(sigma=sigma)
+    assert prior.sigma == sigma
+
+# Test __call__ method with valid inputs
+def test_zero_radial_velocity_prior_call():
+    # Use utility function to sample an orbit
+    t = 0  # Example time
+    orbit = sample_LEO_orbit(t)  # Sample a LEO orbit
+    distance = np.linalg.norm(orbit.r)  # Calculate distance from position vector
+    prior = ZeroRadialVelocityPrior(sigma=0.25)
+
+    # Calculate expected values
+    vdotr = np.dot(normed(orbit.r), normed(orbit.v))
+    chi0 = vdotr / prior.sigma
+    expected_logprior = -0.5 * chi0**2
+
+    # Call the method and verify output
+    result = prior(orbit, distance)
+    assert pytest.approx(result[0]) == expected_logprior
+
+# Test __call__ method with chi=True
+def test_zero_radial_velocity_prior_call_chi():
+    # Use utility function to sample an orbit
+    t = 0  # Example time
+    orbit = sample_LEO_orbit(t)  # Sample a LEO orbit
+    distance = np.linalg.norm(orbit.r)  # Calculate distance from position vector
+    prior = ZeroRadialVelocityPrior(sigma=0.25)
+
+    # Calculate expected values
+    vdotr = np.dot(normed(orbit.r), normed(orbit.v))
+    chi0 = vdotr / prior.sigma
+    expected_chi = chi0
+
[…]
