+++ tests/test_compute.py
@@ -1,0 +1,880 @@
+import pytest
+import numpy as np
+from astropy.time import Time
+
+import ssapy
+from ssapy.utils import get_angle
+from ssapy.compute import (groundTrack, radecRate, radecRateObsToRV, lb_to_unit, nby3shape, calculate_orbital_elements, moon_shine, earth_shine, sun_shine, calc_M_v, M_v_lambertian,
+                           calc_gamma, moon_normal_vector, lunar_lagrange_points, lunar_lagrange_points_circular, lagrange_points_lunar_frame, gcrf_to_itrf, angle_between_vectors, get_body, rotation_matrix_from_vectors)
+from ssapy.constants import EARTH_MU,  MOON_RADIUS, EARTH_RADIUS
+from .ssapy_test_helpers import sample_GEO_orbit
+from ssapy.propagator import KeplerianPropagator
+from ssapy.orbit import EarthObserver
+
+# Mock implementations for dependencies
+def mock_rv(orbit, time, propagator=None):
+    """Mock implementation of rv function."""
+    r = np.tile(orbit.r, (len(time), 1, 1))  # Repeat position for all times
+    v = np.tile(orbit.v, (len(time), 1, 1))  # Repeat velocity for all times
+    return r, v
+
+def mock_ellipsoid_cartToSphere(x, y, z):
+    """Mock implementation of Ellipsoid.cartToSphere."""
+    lon = np.arctan2(y, x)  # Mock longitude
+    lat = np.arctan2(z, np.sqrt(x**2 + y**2))  # Mock latitude
+    height = np.sqrt(x**2 + y**2 + z**2) - 6371e3  # Mock height (subtract Earth's radius)
+    return lon, lat, height
+
+# Replace actual functions with mocks for testing
+@pytest.fixture(autouse=True)
+def mock_dependencies(monkeypatch):
+    monkeypatch.setattr("your_module.rv", mock_rv)
+    monkeypatch.setattr("your_module.Ellipsoid.cartToSphere", mock_ellipsoid_cartToSphere)
+
+def test_groundTrack_geodetic():
+    # Use sample_GEO_orbit and create time
+    orbit = sample_GEO_orbit
+    time = Time([0, 100, 200], format='gps')
+
+    # Run groundTrack function
+    lon, lat, height = groundTrack(orbit, time, format='geodetic')
+
+    # Verify outputs
+    assert lon.shape == (len(time),)
+    assert lat.shape == (len(time),)
+    assert height.shape == (len(time),)
+    assert np.all(height >= 0)  # Height should be non-negative
+
+def test_groundTrack_cartesian():
+    # Use sample_GEO_orbit and create time
+    orbit = sample_GEO_orbit
+    time = Time([0, 100, 200], format='gps')
+
+    # Run groundTrack function
+    x, y, z = groundTrack(orbit, time, format='cartesian')
+
+    # Verify outputs
+    assert x.shape == (len(time),)
+    assert y.shape == (len(time),)
+    assert z.shape == (len(time),)
+    assert np.allclose(x, orbit.r[0])  # Mock position should match input
+
+def test_groundTrack_invalid_format():
+    # Use sample_GEO_orbit and create time
+    orbit = sample_GEO_orbit
+    time = Time([0, 100, 200], format='gps')
+
+    # Run groundTrack with invalid format
+    with pytest.raises(ValueError, match="Format must be either 'cartesian' or 'geodetic'"):
+        groundTrack(orbit, time, format='invalid')
+
+def test_groundTrack_scalar_time():
+    # Use sample_GEO_orbit and scalar time
+    orbit = sample_GEO_orbit
[…]
