+++ tests/test_tracks.py
@@ -1,11 +1,20 @@
 import pytest
 import numpy as np
-from unittest.mock import MagicMock
+from astropy.time import Time
+from astropy import units as u
+import math
+
 import ssapy
-from ssapy.correlate_tracks import CircVelocityPrior, ZeroRadialVelocityPrior, GaussPrior, VolumeDistancePrior, orbit_to_param, partial, make_param_guess
+from ssapy.correlate_tracks import (CircVelocityPrior, ZeroRadialVelocityPrior, GaussPrior, VolumeDistancePrior, orbit_to_param, partial, 
+                                    make_param_guess, make_optimizer,fit_arc_blind, fit_arc, fit_arc_with_gaussian_prior, data_for_satellite,
+                                    wrap_angle_difference, radeczn, param_to_orbit, orbit_to_param, TrackBase, Track, TrackGauss, MHT,
+                                    data_for_satellite, Hypothesis, time_ordered_satIDs, summarize_tracklet, summarize_tracklets, iterate_mht)
 from ssapy.constants import EARTH_MU, RGEO
-from .ssapy_test_helpers import sample_LEO_orbit 
+from .ssapy_test_helpers import sample_GEO_orbit, sample_LEO_orbit, checkAngle, checkSphere 
 from ssapy.utils import normed
+from ssapy.orbit import Orbit
+from ssapy import propagator
+from ssapy import rvsampler
 
 ####################
 ## TESTING PRIORS ##
@@ -218,7 +227,7 @@ def test_volume_distance_prior_negative_distance():
 
 
 #######################
-## testing functions ##
+## TESTING FUNCTIONS ##
 #######################
 
 # Mock data for testing
@@ -307,3 +316,1121 @@ def test_make_param_guess_invalid_mode():
 
     with pytest.raises(ValueError, match='unrecognized mode'):
         make_param_guess(rvguess, arc, mode)
+
+def test_make_optimizer_valid_modes():
+    """Test make_optimizer with valid modes."""
+    param = [1, 2, 3, 4, 5, 6, 7, 8, 9]  # Example parameter list
+
+    # Test 'rv' mode
+    optimizer = make_optimizer(mode='rv', param=param, lsq=False)
+    assert optimizer == rvsampler.LMOptimizer, "Expected LMOptimizer for 'rv' mode"
+
+    # Test 'angle' mode
+    optimizer = make_optimizer(mode='angle', param=param, lsq=False)
+    assert isinstance(optimizer, partial), "Expected a partial function for 'angle' mode"
+    assert optimizer.func == rvsampler.LMOptimizerAngular, "Expected LMOptimizerAngular for 'angle' mode"
+    assert optimizer.keywords['initObsPos'] == param[-6:-3], "Incorrect initObsPos for 'angle' mode"
+    assert optimizer.keywords['initObsVel'] == param[-3:], "Incorrect initObsVel for 'angle' mode"
+
+    # Test 'equinoctial' mode
+    optimizer = make_optimizer(mode='equinoctial', param=param, lsq=False)
+    assert optimizer == rvsampler.EquinoctialLMOptimizer, "Expected EquinoctialLMOptimizer for 'equinoctial' mode"
+
+def test_make_optimizer_valid_modes_lsq():
+    """Test make_optimizer with valid modes and lsq=True."""
+    param = [1, 2, 3, 4, 5, 6, 7, 8, 9]  # Example parameter list
+
+    # Test 'rv' mode
+    optimizer = make_optimizer(mode='rv', param=param, lsq=True)
[…]
