+++ tests/test_particles.py
@@ -12,133 +12,82 @@ from .ssapy_test_helpers import timer
 from ssapy.constants import RGEO, VGEO
 
 
+def random_vector(min_mag, max_mag):
+    """Generate a random 3D vector with magnitude between min_mag and max_mag."""
+    vec = np.random.normal(size=3)
+    vec /= np.linalg.norm(vec)
+    magnitude = np.random.uniform(min_mag, max_mag)
+    return vec * magnitude
+
+
 @pytest.mark.slow
 @timer
 def test_particles():
     np.random.seed(562)
 
     time0 = Time(2458316., format='jd')
+    lon, lat, elevation = 100.0, 33.0, 1300.0
+    observer = ssapy.EarthObserver(lon, lat, elevation)
 
-    # Setup observer location
-    lon = 100.0
-    lat = 33.0
-    elevation = 1300.0
-    earthObserver = ssapy.EarthObserver(lon, lat, elevation)
-
-    #
-    # Create orbit for the target
-    #
-    # a = RGEO * 0.98
-    # e = 0.01
-    # i = 0.001
-    # pa = 0.0
-    # raan = 1.2
-    # trueAnomaly = 2.03
-    # orbit = Orbit.fromKeplerianElements(a, e, i, raan, pa, trueAnomaly, time0)
-
-    r = np.random.uniform(RGEO*0.9, RGEO*1.1)
-    theta = np.random.uniform(0, 2*np.pi)
-    phi = np.random.uniform(0, np.pi)
-    x = r * np.cos(theta) * np.sin(phi)
-    y = r * np.sin(theta) * np.sin(phi)
-    z = r * np.cos(phi)
-    r = np.array([x, y, z])
-    # Pick a velocity near VGEO
-    v = np.random.uniform(VGEO*0.9, VGEO*1.1)
-    # Pick a random direction
-    theta = np.random.uniform(0, 2*np.pi)
-    phi = np.random.uniform(0, np.pi)
-    vx = v * np.cos(theta) * np.sin(phi)
-    vy = v * np.sin(theta) * np.sin(phi)
-    vz = v * np.cos(phi)
-    v = np.array([vx, vy, vz])
+    # True orbit
+    r = random_vector(RGEO * 0.9, RGEO * 1.1)
+    v = random_vector(VGEO * 0.9, VGEO * 1.1)
     orbit = ssapy.Orbit(r, v, time0)
 
-    num_epochs = 2
     particles = []
     epoch_times = []
     lnprobs = []
-    ch_rmean = []
-    ch_vmean = []
     time = time0
-    for _ in range(num_epochs):
+    for _ in range(2):  # two epochs
         epoch_times.append(time)
-
-        # Build observation times
-        earthTimes = time + np.linspace(0, 10, 5)*u.h
-        earthRaDec = ssapy.radec(orbit, earthTimes, observer=earthObserver)
-        earthRStation, earthVStation = earthObserver.getRV(earthTimes)
+        times = time + np.linspace(0, 10, 5) * u.h
+        ra, dec = ssapy.radec(orbit, times, observer=observer)
+        r_station, v_station = observer.getRV(times)
 
         arc = QTable()
-        arc['ra'] = Longitude(earthRaDec[0]*u.rad)
-        arc['dec'] = Latitude(earthRaDec[1]*u.rad)
-        arc['rStation_GCRF'] = earthRStation*u.m
-        arc['vStation_GCRF'] = earthVStation*u.m/u.s
-        arc['time'] = Time(earthTimes)
-        arc['sigma'] = np.ones(5, dtype=float)*u.arcsec
+        arc['ra'] = Longitude(ra * u.rad)
+        arc['dec'] = Latitude(dec * u.rad)
+        arc['rStation_GCRF'] = r_station * u.m
[…]
