+++ tests/test_linker.py
@@ -1,172 +1,125 @@
-import pytest
 import numpy as np
-
+import pytest
 from astropy.time import Time
 import astropy.units as u
 from astropy.coordinates import Longitude, Latitude
 from astropy.table import QTable
 
 import ssapy
-from .ssapy_test_helpers import timer
 from ssapy.constants import RGEO, VGEO
 
-def _mock_chain_from_rv(r, v, nSteps=10):
-    """ Make a mock MCMC chain output from an orbit model
-    """
-    params = np.concatenate((r, v))
-    chain = np.zeros((nSteps, 6))
-    for istep in range(nSteps):
-        chain[istep, :] = params
-    return chain
 
-
-def _create_iods(num_epochs=10, num_obs_per_track=3, nBurn=500, nStep=100):
+# Helper: simplified particle generation
+def _create_iods_small(num_epochs=2, num_obs_per_track=3, nBurn=10, nStep=10):
     from ssapy.rvsampler import RPrior, APrior
-    np.random.seed(27)
+    np.random.seed(42)
 
     sigma_arcsec = 1.0
-    sigma_rad = sigma_arcsec * np.pi / (180.*3600.)
-
+    sigma_rad = sigma_arcsec * np.pi / (180. * 3600.)
     time0 = Time(2458316., format='jd')
 
-    # Setup observer location
-    lon = 100.0
-    lat = 33.0
-    elevation = 1300.0
-    earthObserver = ssapy.EarthObserver(lon, lat, elevation)
+    lon, lat, elevation = 100.0, 33.0, 1300.0
+    observer = ssapy.EarthObserver(lon, lat, elevation)
+    orbit = ssapy.Orbit.fromKeplerianElements(RGEO * 0.98, 0.01, 0.001, 0.0, 1.2, 1.03, time0)
+    times = time0 + np.linspace(0, 4, num_obs_per_track) * u.h
+    ra, dec = ssapy.radec(orbit, times, observer=observer)
+    rstation, vstation = observer.getRV(times)
 
-    # Create orbit for target
-    orbit = ssapy.Orbit.fromKeplerianElements(RGEO*0.98, 0.01, 0.001, 0.0, 1.2, 1.03, time0)
-
-    # Build observation times
-    earthTimes = time0 + np.linspace(0, 4, num_obs_per_track)*u.h
-    earthRaDec = ssapy.radec(orbit, earthTimes, observer=earthObserver)
-    earthRStation, earthVStation = earthObserver.getRV(earthTimes)
-
-    # Create multiple tracks for a single object
     iods = []
-    time = time0
-    for iepoch in range(num_epochs):
-        print("IOD epoch {:d}".format(iepoch+1))
-
+    for _ in range(num_epochs):
         arc = QTable()
-        err_ra = np.random.randn(len(earthRaDec[0])) * sigma_rad
-        err_dec = np.random.randn(len(earthRaDec[1])) * sigma_rad
-        arc['ra'] = Longitude((earthRaDec[0] + err_ra)*u.rad)
-        arc['dec'] = Latitude((earthRaDec[1] + err_dec)*u.rad)
-        arc['rStation_GCRF'] = earthRStation*u.m
-        arc['vStation_GCRF'] = earthVStation*u.m/u.s
-        arc['time'] = Time(earthTimes)
-        arc['sigma'] = sigma_arcsec * np.ones(num_obs_per_track, dtype=float)*u.arcsec
+        arc['ra'] = Longitude((ra[0] + np.random.randn(len(ra[0])) * sigma_rad) * u.rad)
+        arc['dec'] = Latitude((dec[1] + np.random.randn(len(dec[1])) * sigma_rad) * u.rad)
+        arc['rStation_GCRF'] = rstation * u.m
[…]
