+++ tests/test_tracks.py
@@ -3,391 +3,33 @@ import numpy as np
 from astropy.time import Time
 from astropy import units as u
 import math
-import unittest
+from functools import partial
 
 import ssapy
-from ssapy.correlate_tracks import (CircVelocityPrior, ZeroRadialVelocityPrior, GaussPrior, VolumeDistancePrior, orbit_to_param, partial, 
-                                    make_param_guess, make_optimizer,fit_arc_blind, fit_arc, fit_arc_with_gaussian_prior, data_for_satellite,
-                                    wrap_angle_difference, radeczn, param_to_orbit, orbit_to_param, TrackBase, Track, TrackGauss, MHT,
-                                    data_for_satellite, Hypothesis, time_ordered_satIDs, summarize_tracklet, summarize_tracklets, iterate_mht,
-                                    fit_arc_blind_via_track)
 from ssapy.constants import EARTH_MU, RGEO
-from .ssapy_test_helpers import sample_GEO_orbit, sample_LEO_orbit, checkAngle, checkSphere 
-from ssapy.utils import normed
+from ssapy.correlate_tracks import (
+    CircVelocityPrior, ZeroRadialVelocityPrior, GaussPrior, VolumeDistancePrior,
+    orbit_to_param, make_param_guess, make_optimizer, fit_arc_blind, fit_arc,
+    fit_arc_with_gaussian_prior, data_for_satellite, wrap_angle_difference,
+    radeczn, param_to_orbit, Track, TrackGauss, MHT, summarize_tracklet,
+    summarize_tracklets, iterate_mht, fit_arc_blind_via_track, Hypothesis
+)
 from ssapy.orbit import Orbit
-from ssapy import propagator
-from ssapy import rvsampler
+from ssapy import propagator, rvsampler
+from .ssapy_test_helpers import sample_GEO_orbit, sample_LEO_orbit, checkAngle, checkSphere
 
-####################
-## TESTING PRIORS ##
-####################
-def test_circ_velocity_prior_init():
-    sigma = np.log(2) / 3.5
-    prior = CircVelocityPrior(sigma=sigma)
-    assert prior.sigma == sigma
-
-def test_circ_velocity_prior_call():
-    t = 0  # Example time
-    orbit = sample_LEO_orbit(t)  # Sample a LEO orbit
-    distance = np.linalg.norm(orbit.r)  # Calculate distance from position vector
-    prior = CircVelocityPrior(sigma=np.log(2) / 3.5)
-
-    # Calculate expected values
-    loggmorv2 = np.log(EARTH_MU / np.linalg.norm(orbit.r) / np.dot(orbit.v, orbit.v))
-    chi0 = loggmorv2 / prior.sigma
-    expected_logprior = -0.5 * chi0**2
-
-    result = prior(orbit, distance)
-    assert pytest.approx(result[0]) == expected_logprior
-
-def test_circ_velocity_prior_call_chi():
-    t = 0  # Example time
-    orbit = sample_LEO_orbit(t)  # Sample a LEO orbit
-    distance = np.linalg.norm(orbit.r)  # Calculate distance from position vector
-    prior = CircVelocityPrior(sigma=np.log(2) / 3.5)
-
-    # Calculate expected values
-    loggmorv2 = np.log(EARTH_MU / np.linalg.norm(orbit.r) / np.dot(orbit.v, orbit.v))
-    chi0 = loggmorv2 / prior.sigma
-    expected_chi = chi0
-
-    # Call the method with chi=True and verify output
-    result = prior(orbit, distance, chi=True)
-    assert pytest.approx(result[0]) == expected_chi
-
-# Test ZeroRadialVelocityPrior initialization
-def test_zero_radial_velocity_prior_init():
[…]
