### 2025-05-12 eaf0260
better plotutils and particles
