+++ tests/test_compute.py
@@ -1,4 +1,3 @@
-import pytest
 import numpy as np
 from astropy.time import Time
 
