+++ tests/test_ellipsoid.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pytest
 from ssapy.ellipsoid import Ellipsoid
 
 @pytest.fixture
