+++ tests/test_tracks.py
@@ -1,4 +1,3 @@
-import pytest
 import numpy as np
 from astropy.time import Time
 from astropy import units as u
