+++ tests/test_plotutils.py
@@ -1,4 +1,3 @@
-import pytest
 import numpy as np
 import os
 import io
@@ -25,10 +24,10 @@ from ssapy.plotUtils import (
 
 def test_load_earth_file():
     fake_file_path = "fake_path/earth.png"
-    with patch("find_file", return_value=fake_file_path) as mock_find_file:
+    with patch("ssapy.plotUtils.find_file", return_value=fake_file_path) as mock_find_file:
         mock_image = MagicMock(spec=PILImage.Image)
         mock_image.size = (5400, 2700)  # Original size of the image
-        with patch("PILImage.open", return_value=mock_image) as mock_open:
+        with patch("ssapy.plotUtils.PILImage.open", return_value=mock_image) as mock_open:
             # Mock the resize method to return a resized image
             resized_image = MagicMock(spec=PILImage.Image)
             resized_image.size = (1080, 540)  # Resized dimensions
@@ -46,12 +45,12 @@ def test_load_earth_file():
 
 def test_draw_earth():
     fake_texture = MagicMock()
-    with patch("load_earth_file", return_value=fake_texture) as mock_load_earth_file:
+    with patch("ssapy.plotUtils.load_earth_file", return_value=fake_texture) as mock_load_earth_file:
         # Mock the ipv.plot_mesh function
         fake_plot_mesh = MagicMock()
-        with patch("ipv.plot_mesh", return_value=fake_plot_mesh) as mock_plot_mesh:
+        with patch("ssapy.plotUtils.ipv.plot_mesh", return_value=fake_plot_mesh) as mock_plot_mesh:
             fake_gst = np.array([0.1, 0.2, 0.3])  # Fake GST values
-            with patch("gst94", return_value=fake_gst) as mock_gst94:
+            with patch("ssapy.plotUtils.gst94", return_value=fake_gst) as mock_gst94:
                 # Define test inputs
                 test_time = np.array([100000, 200000, 300000])  # GPS seconds
                 test_ngrid = 50
@@ -88,11 +87,11 @@ def test_draw_earth():
 
 def test_load_moon_file():
     fake_file_path = "fake_path/moon.png"
-    with patch("find_file", return_value=fake_file_path) as mock_find_file:
+    with patch("ssapy.plotUtils.find_file", return_value=fake_file_path) as mock_find_file:
         # Mock the PILImage.open function to return a mock image
         mock_image = MagicMock(spec=PILImage.Image)
         mock_image.size = (5400, 2700)  # Original size of the image
-        with patch("PILImage.open", return_value=mock_image) as mock_open:
+        with patch("ssapy.plotUtils.PILImage.open", return_value=mock_image) as mock_open:
             # Mock the resize method to return a resized image
             resized_image = MagicMock(spec=PILImage.Image)
             resized_image.size = (1080, 540)  # Resized dimensions
@@ -111,13 +110,13 @@ def test_load_moon_file():
 def test_draw_moon():
     # Mock the load_moon_file function to return a fake texture
     fake_texture = MagicMock()
-    with patch("load_moon_file", return_value=fake_texture) as mock_load_moon_file:
+    with patch("ssapy.plotUtils.load_moon_file", return_value=fake_texture) as mock_load_moon_file:
         # Mock the ipv.plot_mesh function
         fake_plot_mesh = MagicMock()
[…]
