+++ tests/test_io.py
@@ -1,4 +1,3 @@
-import pytest
 import os
 import numpy as np
 from io import StringIO
