+++ _sources/contribution_guide.rst.txt
@@ -1,0 +1,528 @@
+.. Copyright 2013-2024 Lawrence Livermore National Security, LLC. See the top-level LICENSE file for details.
+
+   SPDX-License-Identifier: MIT
+
+.. _contribution-guide:
+
+==================
+Contribution Guide
+==================
+
+This guide is intended for developers or administrators who want to
+contribute a new feature or bugfix to SSAPy.
+It assumes that you have at least some familiarity with Git and GitHub.
+The guide will show a few examples of contributing workflows and discuss
+the granularity of pull-requests (PRs). It will also discuss the tests your
+PR must pass in order to be accepted into SSAPy.
+
+First, what is a PR? Quoting `Bitbucket's tutorials <https://www.atlassian.com/git/tutorials/making-a-pull-request/>`_:
+
+  Pull requests are a mechanism for a developer to notify team members that
+  they have **completed a feature**. The pull request is more than just a
+  notification—it’s a dedicated forum for discussing the proposed feature.
+
+The emphasis is on a **completed feature**. The changes one proposes in a PR should
+correspond to one feature/bugfix/extension/etc. One can create PRs with
+changes relevant to different ideas, however reviewing such PRs becomes tedious
+and error prone. If possible, try to follow the **one-PR-one-feature** rule.
+
+--------
+Branches
+--------
+
+SSAPy's ``main`` branch has the latest contributions. Nearly all pull
+requests should start from ``main`` and target ``main``.
+
+There is a branch for each major release series. Release branches
+originate from ``main`` and have tags for each point release in the
+series. For example, ``releases/v0.14`` might have tags for ``0.14.0``,
+``0.14.1``, ``0.14.2``, etc. versions of SSAPy. We backport important bug
+fixes to these branches, but we do not advance the package versions or
+make other changes that would change the way SSAPy concretizes
+dependencies. Currently, the maintainers manage these branches by
+cherry-picking from ``main``. See :ref:`releases` for more
+information.
+
+----------------------
+Continuous Integration
+----------------------
+
+SSAPy uses `Github Actions <https://docs.github.com/en/actions>`_ for Continuous Integration
+testing. This means that every time you submit a pull request, a series of tests will
+be run to make sure you didn't accidentally introduce any bugs into SSAPy. **Your PR
+will not be accepted until it passes all of these tests.** While you can certainly wait
+for the results of these tests after submitting a PR, we recommend that you run them
+locally to speed up the review process.
+
+.. note::
+
+   Oftentimes, CI will fail for reasons other than a problem with your PR.
+   For example, apt-get, pip, or homebrew will fail to download one of the
+   dependencies for the test suite, or a transient bug will cause the unit tests
+   to timeout. If any job fails, click the "Details" link and click on the test(s)
+   that is failing. If it doesn't look like it is failing for reasons related to
+   your PR, you have two options. If you have write permissions for the SSAPy
+   repository, you should see a "Restart workflow" button on the right-hand side. If
+   not, you can close and reopen your PR to rerun all of the tests. If the same
+   test keeps failing, there may be a problem with your PR. If you notice that
+   every recent PR is failing with the same error message, it may be that an issue
+   occurred with the CI infrastructure or one of SSAPy's dependencies put out a
+   new release that is causing problems. If this is the case, please file an issue.
+
+
+We currently test against Python 3.10 on both macOS and Linux and
+perform 3 types of tests:
+
+.. _cmd-ssapy-unit-test:
+
+^^^^^^^^^^
+Unit Tests
+^^^^^^^^^^
+
[…]
