+++ _sources/api/ssapy.correlate_tracks.make_param_guess.rst.txt
@@ -1,0 +1,6 @@
+make_param_guess
+================
+
+.. currentmodule:: ssapy.correlate_tracks
+
+.. autofunction:: make_param_guess
