+++ _sources/api/ssapy.rvsampler.sample_ball.rst.txt
@@ -1,0 +1,6 @@
+sample_ball
+===========
+
+.. currentmodule:: ssapy.rvsampler
+
+.. autofunction:: sample_ball
