+++ _sources/api/ssapy.correlate_tracks.fit_arc_with_gaussian_prior.rst.txt
@@ -1,0 +1,6 @@
+fit_arc_with_gaussian_prior
+===========================
+
+.. currentmodule:: ssapy.correlate_tracks
+
+.. autofunction:: fit_arc_with_gaussian_prior
