+++ _sources/api/ssapy.io.merge_dicts.rst.txt
@@ -1,0 +1,6 @@
+merge_dicts
+===========
+
+.. currentmodule:: ssapy.io
+
+.. autofunction:: merge_dicts
