+++ tests/test_simple.py
@@ -1,81 +1,91 @@
-from unittest.mock import MagicMock, patch
-from astropy.time import Time
-import numpy as np
 import pytest
+import numpy as np
+from astropy.time import Time
+from unittest.mock import patch, MagicMock
 
-# Patch imports from simple.py module
-@pytest.fixture(autouse=True)
-def patch_dependencies():
-    with patch("simple.AccelKepler", return_value="kepler"), \
-         patch("simple.AccelThirdBody", side_effect=lambda body: f"thirdbody-{body}"), \
-         patch("simple.AccelHarmonic", side_effect=lambda body, a, b: f"harmonic-{body}-{a}-{b}"), \
-         patch("simple.AccelSolRad", return_value="solrad"), \
-         patch("simple.AccelEarthRad", return_value="earthrad"), \
-         patch("simple.AccelDrag", return_value="drag"), \
-         patch("simple.get_body", side_effect=lambda name, **kwargs: f"{name}-body"), \
-         patch("simple.RK78Propagator", side_effect=lambda accel, h: f"prop-{accel}-{h}"), \
-         patch("simple.Orbit.fromKeplerianElements", return_value="orbit-from-kepler"), \
-         patch("simple.Orbit", side_effect=lambda r, v, t0: f"orbit-from-rv-{r}-{v}-{t0}"), \
-         patch("simple.rv", return_value=(["r-out"], ["v-out"])), \
-         patch("simple.get_times", return_value=["t1", "t2"]), \
-         patch("simple.points_on_circle", return_value=[np.array([1, 2, 3])] * 4):
-        yield
-
-from ssapy import simple
+from ssapy.simple import (
+    keplerian_prop,
+    threebody_prop,
+    fourbody_prop,
+    best_prop,
+    ssapy_kwargs,
+    ssapy_prop,
+    ssapy_orbit,
+    get_similar_orbits,
+)
 
 def test_keplerian_prop():
-    result = simple.keplerian_prop(20)
-    assert result == "prop-kepler-20"
+    prop = keplerian_prop(20)
+    assert hasattr(prop, 'step')  # Should have RK78Propagator-like properties
+
 
 def test_threebody_prop():
-    result1 = simple.threebody_prop(10)
-    result2 = simple.threebody_prop(15)  # Check cached
-    assert result1 == result2
-    assert "thirdbody-moon" in result1
+    prop = threebody_prop(30)
+    assert hasattr(prop, 'step')
+
 
 def test_fourbody_prop():
-    result1 = simple.fourbody_prop(10)
-    result2 = simple.fourbody_prop(15)
-    assert result1 == result2
-    assert "thirdbody-moon" in result1
-    assert "thirdbody-Sun" in result1
+    prop = fourbody_prop(40)
+    assert hasattr(prop, 'step')
+
 
 def test_best_prop():
-    result1 = simple.best_prop(5)
-    result2 = simple.best_prop(10)
-    assert result1 == result2
-    assert "harmonic-Earth-body-140-140" in result1
-    assert "solrad" in result1
-    assert "drag" in result1
+    kwargs = dict(mass=200, area=0.02, CD=2.5, CR=1.1)
+    prop = best_prop(50, kwargs)
+    assert hasattr(prop, 'step')
+
 
 def test_ssapy_kwargs():
-    kw = simple.ssapy_kwargs()
-    assert kw["mass"] == 250
-    assert kw["area"] == 0.022
+    kwargs = ssapy_kwargs(100, 0.01, 2.1, 1.2)
[…]
