+++ tests/test_accel.py
@@ -1,11 +1,19 @@
 import numpy as np
 from astropy.time import Time
 import astropy.units as u
+import sys
+import importlib
+import types
+from types import SimpleNamespace
+import pytest
+
 
 import ssapy
 from ssapy.utils import norm, iers_interp
 from .ssapy_test_helpers import timer, sample_LEO_orbit, sample_GEO_orbit
-from ssapy.accel import AccelProd, Accel
+from ssapy.accel import (AccelProd, Accel, AccelKepler, AccelSolRad, 
+                         AccelEarthRad, AccelDrag, AccelConstNTW)
+from ssapy.constants import EARTH_MU, EARTH_RADIUS
 
 iers_interp(0.0)  # Prime the IERS interpolant cache
 earth = ssapy.get_body("earth")
@@ -1041,7 +1049,6 @@ def test_bielliptic_transfer():
     np.testing.assert_allclose(orbf.i, 0, atol=1e-6)
 
 class MockAccel(Accel):
-    """Mock implementation of Accel for testing purposes."""
     def __init__(self, time_breakpoints=None):
         super().__init__()
         self.time_breakpoints = time_breakpoints if time_breakpoints else []
@@ -1120,24 +1127,291 @@ def test_accelprod_equality():
     # Verify inequality
     assert accel_prod1 != accel_prod4
 
+MODULE_NAME = "ssapy.accel"
 
-if __name__ == '__main__':
-    test_MG_3_1()
-    test_MG_3_2()
-    test_accel_point_mass()
-    test_angles()
-    test_MG_3_4_accel()
-    test_MG_3_4()
-    test_MG_3_4_scipy()
-    test_MG_3_4_rk78()
-    test_RK4_vs_analytic()
-    test_scipy_propagator()
-    test_RK8()
-    test_RK78()
-    test_reverse()
-    test_Hohmann_transfer()
-    test_inclination_change()
-    test_bielliptic_transfer()
-    test_accelprod_equality()
-    test_accelprod_hash()
-    test_accelprod_initialization()
+@pytest.fixture(autouse=True)
+def cleanup_module_cache():
+    """Ensure a clean import state before each test."""
+    if MODULE_NAME in sys.modules:
+        del sys.modules[MODULE_NAME]
+    yield
+    if MODULE_NAME in sys.modules:
+        del sys.modules[MODULE_NAME]
+
+def test_import_erfa_present():
+    # Create a fake `erfa` module
+    fake_erfa = types.ModuleType("erfa")
+    sys.modules["erfa"] = fake_erfa
+
+    # Remove fallback if it was accidentally cached
+    sys.modules.pop("astropy._erfa", None)
+
+    # Import and test
+    import ssapy.accel
+    importlib.reload(ssapy.accel)
+
+    assert ssapy.accel.erfa is fake_erfa
+
+def test_import_fallback_astropy_erfa():
+    # Remove `erfa` to simulate ImportError
+    sys.modules.pop("erfa", None)
+
+    # Create a fake `astropy._erfa` module
+    fake_astropy_erfa = types.ModuleType("astropy._erfa")
+    sys.modules["astropy._erfa"] = fake_astropy_erfa
+
+    # Import and test
+    import ssapy.accel
+    importlib.reload(ssapy.accel)
+
+    assert ssapy.accel.erfa is fake_astropy_erfa
+
+
+# Dummy classes to verify structure
+class DummyAccelProd:
[…]

[…]
