### 2025-05-19 74b1f09
adding test loc to toml
