+++ pyproject.toml
@@ -56,6 +56,9 @@ packages = [
 [tool.poetry.include]
 files = ["ssapy/_ssapy.cpython-38-x86_64-linux-gnu.so"]
 
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+
 [project.urls]
 Homepage = "https://github.com/LLNL/SSAPy"
 Issues = "https://github.com/LLNL/SSAPy/issues"
