### 2025-05-20 0c62697
trying to speed up pytest
