+++ tests/test_accel.py
@@ -42,6 +42,7 @@ earth_MG = ssapy.Body(
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_MG_3_1():
     """Exercise 3.1 from Montenbruck and Gill
     Tests implementation of harmonic acceleration
@@ -77,6 +78,7 @@ def test_MG_3_1():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_MG_3_2():
     """Exercise 3.2 from Montenbruck and Gill
     Tests implementation of position of the Moon
@@ -112,6 +114,7 @@ def test_MG_3_2():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_accel_point_mass():
     t0 = Time("2006-03-14", scale='tt')
     tt = t0 + np.linspace(0, 4, 5)*u.d
@@ -148,6 +151,7 @@ def test_accel_point_mass():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_angles():
     # Using erfa instead of rolling our own...
     try:
@@ -183,6 +187,7 @@ def test_angles():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_MG_3_4_accel():
     """Exercise 3.4 from Montenbruck and Gill
     Tests implementation of harmonic, lunar, solar radiation, and other accelerations
@@ -429,6 +434,7 @@ def test_MG_3_4():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_MG_3_4_scipy():
     t = Time("1999-03-01", scale='utc')
     # LEO sat
@@ -572,6 +578,7 @@ def test_MG_3_4_scipy():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_MG_3_4_rk78():
     """Exercise 3.4 from Montenbruck and Gill.  Tests SSAPy orbit propagation
     with perturbations.
@@ -718,6 +725,7 @@ def test_MG_3_4_rk78():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_RK4_vs_analytic():
     # Test that analytic Keplerian propagation matches RK4 propagator when
     # acceleration is purely Keplerian
@@ -754,6 +762,7 @@ def test_RK4_vs_analytic():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_scipy_propagator():
     # Test that analytic Keplerian propagation matches SciPy ODE propagator when
     # acceleration is purely Keplerian
@@ -790,6 +799,7 @@ def test_scipy_propagator():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_RK8():
     # Test that analytic Keplerian propagation matches RK8 propagator when
     # acceleration is purely Keplerian
@@ -833,6 +843,7 @@ def test_RK8():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_RK78():
     # Test that analytic Keplerian propagation matches RK78 propagator when
     # acceleration is purely Keplerian
@@ -878,6 +889,7 @@ def test_RK78():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_reverse():
     t = Time("1999-03-01", scale='utc')
     # LEO sat
@@ -941,6 +953,7 @@ def test_reverse():
 
 
 @timer
+@pytest.mark.timeout(30)
 def test_Hohmann_transfer():
     """
     Test Hohmann transfer between Low Earth Orbit (LEO) and Geostationary 
@@ -975,6 +988,7 @@ def test_Hohmann_transfer():
     np.testing.assert_allclose(orbfinal.e, 0, atol=1e-5)
     np.testing.assert_allclose(orbfinal.i, 0, atol=1e-5)
 
[…]
