+++ .github/workflows/ci.yml
@@ -138,7 +138,7 @@ jobs:
     - name: Test with pytest
       run: |
         # Run pytest with coverage, using all available CPUs and verbose output
-        pytest -v -n auto --cov-report=xml --cov=ssapy
+        pytest -v -n auto --cov-report=xml --cov=ssapy --durations=20
     - name: Upload coverage reports to Codecov
       if: matrix.os == 'ubuntu-latest' && matrix.python-version == '3.12'
       uses: codecov/codecov-action@v4
