### 2025-05-20 505cff9
lagging ellipsoide tests
