+++ tests/test_accel.py
@@ -42,7 +42,6 @@ earth_MG = ssapy.Body(
 
 
 @timer
-@pytest.mark.timeout(30)
 def test_MG_3_1():
     """Exercise 3.1 from Montenbruck and Gill
     Tests implementation of harmonic acceleration
@@ -78,7 +77,7 @@ def test_MG_3_1():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_MG_3_2():
     """Exercise 3.2 from Montenbruck and Gill
     Tests implementation of position of the Moon
@@ -114,7 +113,7 @@ def test_MG_3_2():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_accel_point_mass():
     t0 = Time("2006-03-14", scale='tt')
     tt = t0 + np.linspace(0, 4, 5)*u.d
@@ -151,7 +150,7 @@ def test_accel_point_mass():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_angles():
     # Using erfa instead of rolling our own...
     try:
@@ -187,7 +186,7 @@ def test_angles():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_MG_3_4_accel():
     """Exercise 3.4 from Montenbruck and Gill
     Tests implementation of harmonic, lunar, solar radiation, and other accelerations
@@ -434,7 +433,7 @@ def test_MG_3_4():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_MG_3_4_scipy():
     t = Time("1999-03-01", scale='utc')
     # LEO sat
@@ -578,7 +577,7 @@ def test_MG_3_4_scipy():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_MG_3_4_rk78():
     """Exercise 3.4 from Montenbruck and Gill.  Tests SSAPy orbit propagation
     with perturbations.
@@ -725,7 +724,7 @@ def test_MG_3_4_rk78():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_RK4_vs_analytic():
     # Test that analytic Keplerian propagation matches RK4 propagator when
     # acceleration is purely Keplerian
@@ -762,7 +761,7 @@ def test_RK4_vs_analytic():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_scipy_propagator():
     # Test that analytic Keplerian propagation matches SciPy ODE propagator when
     # acceleration is purely Keplerian
@@ -799,7 +798,7 @@ def test_scipy_propagator():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_RK8():
     # Test that analytic Keplerian propagation matches RK8 propagator when
     # acceleration is purely Keplerian
@@ -843,7 +842,7 @@ def test_RK8():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_RK78():
     # Test that analytic Keplerian propagation matches RK78 propagator when
     # acceleration is purely Keplerian
@@ -889,7 +888,7 @@ def test_RK78():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_reverse():
     t = Time("1999-03-01", scale='utc')
     # LEO sat
@@ -953,7 +952,7 @@ def test_reverse():
 
 
 @timer
-@pytest.mark.timeout(30)
+ 
 def test_Hohmann_transfer():
     """
     Test Hohmann transfer between Low Earth Orbit (LEO) and Geostationary 
@@ -988,7 +987,7 @@ def test_Hohmann_transfer():
[…]
