### 2025-05-21 ca33704
more invalid tests
