### 2025-05-21 ff4fa41
added examples to documentation
