### 2025-05-21 e1dbbc8
added examples to documentation (#77)
* added examples to documentation

* added examples to documentation
