+++ .github/workflows/ci.yml
@@ -133,9 +133,11 @@ jobs:
     #    flake8 --version
     #    flake8 ssapy/ tests/ devel/
     - name: Install pytest plugins
+      if: matrix.os == 'ubuntu-latest'
       run: |
         pip install pytest pytest-cov pytest-xdist
     - name: Test with pytest
+      if: matrix.os == 'ubuntu-latest'
       run: |
         # Run pytest with coverage, using all available CPUs and verbose output
         pytest -v -n auto --cov-report=xml --cov=ssapy --durations=20
