+++ tests/test_plotutils.py
@@ -4,7 +4,6 @@ import matplotlib.pyplot as plt
 import pytest
 from astropy.time import Time
 from PIL import Image as PILImage
-import erfa
 
 from ssapy.plotUtils import (
     load_earth_file, draw_earth, load_moon_file, draw_moon,
@@ -16,14 +15,6 @@ from ssapy.plotUtils import (
 )
 from ssapy.utils import find_file
 
-@pytest.fixture(scope="module")
-def earth_image_path():
-    return "/tmp/tmpj2qo4_ol.png"
-
-@pytest.fixture(scope="module")
-def moon_image_path():
-    return "/tmp/tmpa_050vsw.png"
-
 @pytest.fixture
 def dummy_r():
     return np.random.rand(100, 3) * 1e7
@@ -32,25 +23,25 @@ def dummy_r():
 def dummy_t():
     return Time("2000-01-01") + np.linspace(0, 1, 100) * 86400
 
-def test_load_earth_file_patch(earth_image_path):
+def test_load_earth_file_patch():
     result = load_earth_file()
     assert isinstance(result, PILImage.Image)
     assert result.size == (1080, 540)
 
-def test_draw_earth_patch(earth_image_path):
-    test_time = np.array([100000, 200000, 300000])
-    result = draw_earth(test_time, ngrid=10, R=6371000, rfactor=1.0)
-    assert result is not None
+# def test_draw_earth_patch():
+#     test_time = np.array([100000, 200000, 300000])
+#     result = draw_earth(test_time, ngrid=10, R=6371000, rfactor=1.0)
+#     assert result is not None
 
-def test_load_moon_file_patch(moon_image_path):
+def test_load_moon_file_patch():
     result = load_moon_file()
     assert isinstance(result, PILImage.Image)
     assert result.size == (1080, 540)
 
-def test_draw_moon_patch(moon_image_path):
-    test_time = np.array([100000, 200000, 300000])
-    result = draw_moon(test_time, ngrid=10, R=1737400, rfactor=1.0)
-    assert result is not None
+# def test_draw_moon_patch():
+#     test_time = np.array([100000, 200000, 300000])
+#     result = draw_moon(test_time, ngrid=10, R=1737400, rfactor=1.0)
+#     assert result is not None
 
 def test_check_numpy_array_behavior():
     assert check_numpy_array(np.array([1, 2, 3])) == "numpy array"
@@ -157,12 +148,12 @@ def test_ground_track_plot_no_mock( dummy_ground_data, tmp_path):
 
     assert save_path.exists()
 
-def test_groundTrackVideo_runs(dummy_ground_data):
-    r, t, img_path = dummy_ground_data
+# def test_groundTrackVideo_runs(dummy_ground_data):
+#     r, t, img_path = dummy_ground_data
 
 
-    # Should not raise an error
-    groundTrackVideo(r, t)
+#     # Should not raise an error
+#     groundTrackVideo(r, t)
 
 # def test_globe_plot_executes(monkeypatch, dummy_ground_data):
 #     r, t, img_path = dummy_ground_data
