### 2025-05-22 c55ad81
corrected examples
