+++ tests/test_accel.py
@@ -12,7 +12,7 @@ import ssapy
 from ssapy.utils import norm, iers_interp
 from .ssapy_test_helpers import timer, sample_LEO_orbit, sample_GEO_orbit
 from ssapy.accel import (AccelProd, Accel, AccelKepler, AccelSolRad, 
-                         AccelEarthRad, AccelDrag, AccelConstNTW)
+                         AccelEarthRad, AccelDrag, AccelConstNTW, AccelSum)
 from ssapy.constants import EARTH_MU, EARTH_RADIUS
 
 iers_interp(0.0)  # Prime the IERS interpolant cache
@@ -42,111 +42,111 @@ earth_MG = ssapy.Body(
 
 
 @timer
-# def test_MG_3_1():
-#     """Exercise 3.1 from Montenbruck and Gill
-#     Tests implementation of harmonic acceleration
-#     """
+def test_MG_3_1():
+    """Exercise 3.1 from Montenbruck and Gill
+    Tests implementation of harmonic acceleration
+    """
     
-#     r = np.array([6525.919e3, 1710.416e3, 2508.886e3])
+    r = np.array([6525.919e3, 1710.416e3, 2508.886e3])
 
-#     # M&G results
-#     mg = {}
-#     mg[2] =  (-6.97922756436043556500,-1.82928105379987315790,-2.69001658551662448100)
-#     mg[4] =  (-6.97931189286893438606,-1.82931487068582887545,-2.68999140120129842657)
-#     mg[6] =  (-6.97922066700256138461,-1.82927878079804373535,-2.68997263886697091095)
-#     mg[8] =  (-6.97927699746898344557,-1.82928186346070154045,-2.68998582282075027194)
-#     mg[10] = (-6.97924447942545889134,-1.82928331385817810606,-2.68997524437058332936)
-#     mg[12] = (-6.97924725687663993767,-1.82928130662047960797,-2.68998625958353532184)
-#     mg[14] = (-6.97924919386132991406,-1.82928546814451764568,-2.68999164568811455212)
-#     mg[16] = (-6.97926211023274856160,-1.82928438361093181896,-2.68999719587274421784)
-#     mg[18] = (-6.97926208121088365033,-1.82928491800309545035,-2.68999523790429417858)
-#     mg[20] = (-6.97926186199732700999,-1.82928315091240034640,-2.68999053339306737342)
+    # M&G results
+    mg = {}
+    mg[2] =  (-6.97922756436043556500,-1.82928105379987315790,-2.69001658551662448100)
+    mg[4] =  (-6.97931189286893438606,-1.82931487068582887545,-2.68999140120129842657)
+    mg[6] =  (-6.97922066700256138461,-1.82927878079804373535,-2.68997263886697091095)
+    mg[8] =  (-6.97927699746898344557,-1.82928186346070154045,-2.68998582282075027194)
+    mg[10] = (-6.97924447942545889134,-1.82928331385817810606,-2.68997524437058332936)
+    mg[12] = (-6.97924725687663993767,-1.82928130662047960797,-2.68998625958353532184)
+    mg[14] = (-6.97924919386132991406,-1.82928546814451764568,-2.68999164568811455212)
+    mg[16] = (-6.97926211023274856160,-1.82928438361093181896,-2.68999719587274421784)
+    mg[18] = (-6.97926208121088365033,-1.82928491800309545035,-2.68999523790429417858)
+    mg[20] = (-6.97926186199732700999,-1.82928315091240034640,-2.68999053339306737342)
 
[…]
