+++ tests/test_accel.py
@@ -76,77 +76,77 @@ def test_MG_3_1():
         a = ah(r, v=None, t=Time("J2000").gps)
 
 
-@timer
+# @timer
  
-def test_MG_3_2():
-    """Exercise 3.2 from Montenbruck and Gill
-    Tests implementation of position of the Moon
-    """
-    t0 = Time("2006-03-14", scale='tt')
-    tt = t0 + np.linspace(0, 4, 5)*u.d
-    mg_moon = [[-387105.185,  106264.577,  61207.474],
-               [-403080.629,   33917.735,  21704.832],
-               [-401102.631,  -39906.188, -18757.478],
-               [-381055.373, -111853.486, -58337.911],
-               [-343564.315, -178551.672, -95178.733]]
-    for i, t in enumerate(tt):
-        with np.printoptions(precision=10):
-            # meter comparison, which matches available precision of M&G text
-            np.testing.assert_allclose(
-                ssapy.utils.moonPos(t)*1e-3,
-                mg_moon[i],
-                rtol=1e-4, atol=0
-            )
+# def test_MG_3_2():
+#     """Exercise 3.2 from Montenbruck and Gill
+#     Tests implementation of position of the Moon
+#     """
+#     t0 = Time("2006-03-14", scale='tt')
+#     tt = t0 + np.linspace(0, 4, 5)*u.d
+#     mg_moon = [[-387105.185,  106264.577,  61207.474],
+#                [-403080.629,   33917.735,  21704.832],
+#                [-401102.631,  -39906.188, -18757.478],
+#                [-381055.373, -111853.486, -58337.911],
+#                [-343564.315, -178551.672, -95178.733]]
+#     for i, t in enumerate(tt):
+#         with np.printoptions(precision=10):
+#             # meter comparison, which matches available precision of M&G text
+#             np.testing.assert_allclose(
+#                 ssapy.utils.moonPos(t)*1e-3,
+#                 mg_moon[i],
+#                 rtol=1e-4, atol=0
+#             )
 
-    # Bonus: test sun position
-    mg_sun = [[147659000.747, -16236318.162,  -7039305.429],
-              [147984936.336, -13880545.131,  -6017952.822],
-              [148266334.848, -11520595.242,  -4994789.325],
-              [148503148.336,  -9157181.657,  -3970124.132],
-              [148695342.555,  -6791017.418,  -2944266.385]]
-    for i, t in enumerate(tt):
-        np.testing.assert_allclose(
-            ssapy.utils.sunPos(t.gps)*1e-3,
-            mg_sun[i],
-            rtol=0, atol=2e1
-        )
+#     # Bonus: test sun position
+#     mg_sun = [[147659000.747, -16236318.162,  -7039305.429],
+#               [147984936.336, -13880545.131,  -6017952.822],
+#               [148266334.848, -11520595.242,  -4994789.325],
+#               [148503148.336,  -9157181.657,  -3970124.132],
+#               [148695342.555,  -6791017.418,  -2944266.385]]
+#     for i, t in enumerate(tt):
[…]
