+++ tests/test_compute.py
@@ -34,34 +34,6 @@ def mock_dependencies(monkeypatch):
     monkeypatch.setattr("ssapy.rv", mock_rv)
     monkeypatch.setattr("ssapy.Ellipsoid.cartToSphere", mock_ellipsoid_cartToSphere)
 
-# def test_groundTrack_geodetic():
-#     # Use sample_GEO_orbit and create time
-#     orbit = sample_GEO_orbit
-#     time = Time([0, 100, 200], format='gps')
-
-#     # Run groundTrack function
-#     lon, lat, height = groundTrack(orbit, time, format='geodetic')
-
-#     # Verify outputs
-#     assert lon.shape == (len(time),)
-#     assert lat.shape == (len(time),)
-#     assert height.shape == (len(time),)
-#     assert np.all(height >= 0)  # Height should be non-negative
-
-# def test_groundTrack_cartesian():
-#     # Use sample_GEO_orbit and create time
-#     orbit = sample_GEO_orbit
-#     time = Time([0, 100, 200], format='gps')
-
-#     # Run groundTrack function
-#     x, y, z = groundTrack(orbit, time, format='cartesian')
-
-#     # Verify outputs
-#     assert x.shape == (len(time),)
-#     assert y.shape == (len(time),)
-#     assert z.shape == (len(time),)
-#     assert np.allclose(x, orbit.r[0])  # Mock position should match input
-
 def test_groundTrack_invalid_format():
     # Use sample_GEO_orbit and create time
     orbit = sample_GEO_orbit
@@ -70,45 +42,6 @@ def test_groundTrack_invalid_format():
     # Run groundTrack with invalid format
     with pytest.raises(ValueError, match="Format must be either 'cartesian' or 'geodetic'"):
         groundTrack(orbit, time, format='invalid')
-
-# def test_groundTrack_scalar_time():
-#     # Use sample_GEO_orbit and scalar time
-#     orbit = sample_GEO_orbit
-#     time = Time(100, format='gps')  # Scalar time
-
-#     # Run groundTrack function
-#     lon, lat, height = groundTrack(orbit, time, format='geodetic')
-
-#     # Verify outputs
-#     assert np.isscalar(lon)
-#     assert np.isscalar(lat)
-#     assert np.isscalar(height)
-
-# def test_groundTrack_scalar_orbit():
-#     # Use sample_GEO_orbit and create time
-#     orbit = sample_GEO_orbit
-#     time = Time([0, 100, 200], format='gps')
-
-#     # Run groundTrack function
-#     lon, lat, height = groundTrack(orbit, time, format='geodetic')
-
-#     # Verify outputs
-#     assert lon.shape == (len(time),)
-#     assert lat.shape == (len(time),)
-#     assert height.shape == (len(time),)
-
-# def test_groundTrack_vector_orbit():
-#     # Create a list of sample GEO orbits and time
-#     orbit = [sample_GEO_orbit, sample_GEO_orbit]
-#     time = Time([0, 100, 200], format='gps')
-
-#     # Run groundTrack function
-#     lon, lat, height = groundTrack(orbit, time, format='geodetic')
-
-#     # Verify outputs
-#     assert lon.shape == (len(orbit), len(time))
-#     assert lat.shape == (len(orbit), len(time))
-#     assert height.shape == (len(orbit), len(time))
 
 # Mock implementation of `radec` function
 def mock_radec(
@@ -130,95 +63,6 @@ def mock_radec(
 def mock_dependencies(monkeypatch):
     monkeypatch.setattr("ssapy.radec", mock_radec)
 
-# def test_radecRate_with_EarthObserver():
[…]
