+++ tests/test_linker.py
@@ -51,72 +51,3 @@ def test_model_selector_params_normalize():
     p.normalize()
     for i in range(3):
         assert np.isclose(np.sum(p[i]), 1.0)
-
-
-# def test_binary_selector_linking():
-#     p = ssapy.BinarySelectorParams(4, 4)
-#     p.params[0, 0] = 1
-#     p.params[1, 1] = 1
-#     p.params[2, 1] = 1
-#     linked = p.get_linked_track_indices(1)
-#     assert set(linked) == {1, 2}
-#     unlinked = p.get_unlinked_track_indices()
-#     assert 3 in unlinked
-
-
-# def test_sample_p_orbit():
-#     iods = _create_iods_small()
-#     linker = ssapy.Linker(iods)
-#     p = linker.sample_Porbit_conditional_dist(1)
-#     assert np.isclose(np.sum(p), 1.0)
-
-
-# def test_update_orbit_parameters_unlinked():
-#     iods = _create_iods_small()
-#     linker = ssapy.Linker(iods)
-#     linker.orbit_selectors.params[:, :] = 0  # Unlink all
-#     linker.update_orbit_parameters()
-#     for iod in linker.iods:
-#         assert hasattr(iod, 'reset_to_pseudo_prior')
-
-
-# def test_sample_orbit_selectors():
-#     iods = _create_iods_small()
-#     linker = ssapy.Linker(iods)
-#     result = linker.sample_orbit_selectors_from_data_conditional(1, verbose=False)
-#     assert isinstance(result, np.ndarray)
-#     assert np.sum(result) == 1
-
-
-# def test_update_params_using_carlin_chib():
-#     iods = _create_iods_small()
-#     linker = ssapy.Linker(iods)
-#     linker.update_params_using_carlin_chib(verbose=False)
-#     assert isinstance(linker.orbit_selectors.params, np.ndarray)
-
-
-# def test_linker_repr():
-#     iods = _create_iods_small()
-#     linker = ssapy.Linker(iods)
-#     rep = repr(linker)
-#     assert "Linker" in rep
-#     assert "orbit_selectors" in rep
-#     assert "p_orbit" in rep
-
-
-# def test_linker_save_step(tmp_path):
-#     iods = _create_iods_small()
-#     linker = ssapy.Linker(iods)
-#     outfile_head = tmp_path / "linker_test"
-#     linker.save_step(str(outfile_head))
-#     assert (tmp_path / "linker_test_orbit_selectors.txt").exists()
-#     assert (tmp_path / "linker_test_p_orbit.txt").exists()
-#     for i in range(len(iods)):
-#         assert (tmp_path / f"linker_test_particles_{i}.txt").exists()
-
-
-# def test_linker_sample_loop():
-#     iods = _create_iods_small()
-#     linker = ssapy.Linker(iods)
-#     linker.sample(nStep=2)  # Just 2 steps for speed
-#     assert isinstance(linker.orbit_selectors.params, np.ndarray)

[…]
