+++ _sources/api/ssapy.body.PlanetPosition.rst.txt
@@ -1,0 +1,17 @@
+PlanetPosition
+==============
+
+.. currentmodule:: ssapy.body
+
+.. autoclass:: PlanetPosition
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~PlanetPosition.__call__
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: __call__
