+++ _sources/api/ssapy.io.read_tle_catalog.rst.txt
@@ -1,0 +1,6 @@
+read_tle_catalog
+================
+
+.. currentmodule:: ssapy.io
+
+.. autofunction:: read_tle_catalog
