+++ _sources/api/ssapy.rvsampler.RVSigmaProposal.rst.txt
@@ -1,0 +1,7 @@
+RVSigmaProposal
+===============
+
+.. currentmodule:: ssapy.rvsampler
+
+.. autoclass:: RVSigmaProposal
+   :show-inheritance:
