### 2025-05-27 c371912
fixed docs typos (#78)
* added examples to documentation

* added examples to documentation

* corrected examples
