### 2025-05-28 #8147 @athulpg007
Once the two issues I opened are addressed, I will be able to finish the review soon. I encourage the authors to provide at least a few examples of how to use the software (ideally to solve real-world analysis problems) in the docs.
