+++ tests/test_sampler.py
@@ -1,9 +1,9 @@
 import numpy as np
-
 from astropy.time import Time
 import astropy.units as u
 from astropy.coordinates import Longitude, Latitude
 from astropy.table import QTable
+import pytest
 
 import ssapy
 from ssapy.utils import cluster_emcee_walkers, norm
