+++ .github/workflows/python-publish.yml
@@ -1,94 +1,86 @@
-# This workflow will upload a Python Package to PyPI when a release is created
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python#publishing-to-package-registries
-
-# This workflow uses actions that are not certified by GitHub.
-# They are provided by a third-party and are governed by
-# separate terms of service, privacy policy, and support
-# documentation.
-
 name: Upload Python Package
 
 on:
-  release:
-    types: [published]
+  workflow_run:
+    workflows: ["CI/CD"]
+    types:
+      - completed
 
 permissions:
   contents: read
 
 jobs:
   release-build:
+    if: ${{ github.event.workflow_run.conclusion == 'success' }}
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
         os: [ubuntu-latest, macOS-latest]
         python-version: ["3.8", "3.12"]
+
     steps:
       - uses: actions/checkout@v4
+        with:
+          submodules: 'recursive'
+
       - uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
+          python-version: ${{ matrix.python-version }}
 
-      - name: Install dependencies (Linux)
+      - name: Install and build (Linux)
         if: matrix.os == 'ubuntu-latest'
         run: |
-          # Install dependencies
           sudo apt-get update
-          sudo apt-get install build-essential git git-lfs python3 python3-setuptools python3-venv graphviz cmake
-
-          # Initialize and update submodules
-          git submodule update --init --recursive
-
-          # Install Git LFS and pull LFS objects
+          sudo apt-get install build-essential git git-lfs python3-setuptools python3-venv graphviz cmake
           git lfs install
           git lfs pull
-          
-          # Upgrade pip and install Python dependencies
-          python -m pip install --upgrade pip setuptools flake8
+          python -m pip install --upgrade pip setuptools flake8 build
           python -m pip install -r requirements.txt
-          python -m pip list
-      
-      - name: Install dependencies (MacOS)
+          python3 -m build
+          mkdir -p dist
+          cp dist/*.whl dist/
+
+      - name: Install and build (macOS)
         if: matrix.os == 'macOS-latest'
         run: |
-          # Update Homebrew and install dependencies
           brew update
-          brew install git git-lfs python3 graphviz cmake
-      
-          # Initialize and update submodules
-          git submodule update --init --recursive
-      
-          # Install Git LFS and pull LFS objects
+          brew install git git-lfs python3 cmake
           git lfs install
           git lfs pull
-      
-          # Upgrade pip and install Python dependencies
-          python3 -m pip install --upgrade pip setuptools flake8
+          python3 -m pip install --upgrade pip setuptools flake8 cibuildwheel delocate
           python3 -m pip install -r requirements.txt
-      
-          # List installed Python packages
-          python3 -m pip list
-  pypi-publish:
-    runs-on: ubuntu-latest
-    needs:
-      - release-build
-    permissions:
-      # IMPORTANT: this permission is mandatory for trusted publishing
-      id-token: write
+          PYTHON_VERSION=$(python3 -c "import sys; print(f'cp{sys.version_info.major}{sys.version_info.minor}')")
+          export ARCHFLAGS="-arch arm64"
+          export CMAKE_OSX_ARCHITECTURES=arm64
+          export CIBW_BUILD="${PYTHON_VERSION}-macosx_*"
+          python3 -m cibuildwheel --platform macos
+          mkdir -p dist
+          cp wheelhouse/*.whl dist/
 
-    # Dedicated environments with protections for publishing are strongly recommended.
[…]
