+++ _sources/api/ssapy.utils.unit_to_tp.rst.txt
@@ -1,0 +1,6 @@
+unit_to_tp
+==========
+
+.. currentmodule:: ssapy.utils
+
+.. autofunction:: unit_to_tp
