+++ _sources/api/ssapy.io.parse_overpunched.rst.txt
@@ -1,0 +1,6 @@
+parse_overpunched
+=================
+
+.. currentmodule:: ssapy.io
+
+.. autofunction:: parse_overpunched
