+++ _sources/api/ssapy.io.pdstr_to_arrays.rst.txt
@@ -1,0 +1,6 @@
+pdstr_to_arrays
+===============
+
+.. currentmodule:: ssapy.io
+
+.. autofunction:: pdstr_to_arrays
