### 2025-05-31 LLNL/SSAPy#81 @andrea-sesta
A bug was found in SSAPy.compute
**Describe the bug**
When running the exact example code from the documentation, some ssapy.compute functions return outputs that contradict both the illustrated examples and basic physical expectations:
- `moon_normal_vector(t)` is not normalized (magnitude ≫ 1).
- `sun_shine(...)` yields near‐zero flux for a sunlit configuration.

**To Reproduce**
Follow the example code shown in the documentation of each function cited above, in particular the used code is the following:
```python
import numpy as np
from ssapy import *

t = Time("2024-1-1")
normal_vector = compute.moon_normal_vector(t)
print(f'{normal_vector=}')

r_sat = np.array([[1e7, 1e7, 1e7]])
r_earth = np.array([[0, 0, 0]])
r_sun = np.array([[1e11, 0, 0]])
sun_shine_dict = compute.sun_shine(r_sat, r_earth, r_sun, radius=0.4, albedo=0.20, albedo_front=0.05, area_panels=100)
print(f'{sun_shine_dict=}')
```
The observed output:
```
normal_vector=array([ 1.18111582e+07, -1.79752782e+08,  3.35900748e+08])
sun_shine_dict={'sun_bus': array([5.99960907e-18]), 'sun_panels': array([0.])}
```

**Expected behavior**
- `moon_normal_vector(t)` should return a unit‐length vector (‖normal_vector‖ ≈ 1).
- `sun_shine(...)` should return positive, finite solar flux values for a satellite at (1e7, 1e7, 1e7) m with the Sun at (1e11, 0, 0), not near zero.

**Output**
See above sections.

 - OS: macOS
 - Version: 15.5 (24F74)
