### 2025-05-31 #8147 @andrea-sesta
I agree with @athulpg007: some usage examples should be added to the documentation. Moreover, I encountered the following issue when running the example code in the API:

- https://github.com/LLNL/SSAPy/issues/81
