+++ _sources/api/ssapy.io.get_tel_pos_itrf_to_gcrs.rst.txt
@@ -1,0 +1,6 @@
+get_tel_pos_itrf_to_gcrs
+========================
+
+.. currentmodule:: ssapy.io
+
+.. autofunction:: get_tel_pos_itrf_to_gcrs
