+++ _sources/api/ssapy.utils.xyz_to_tp.rst.txt
@@ -1,0 +1,6 @@
+xyz_to_tp
+=========
+
+.. currentmodule:: ssapy.utils
+
+.. autofunction:: xyz_to_tp
