+++ _sources/api/ssapy.io.guess_csv_delimiter.rst.txt
@@ -1,0 +1,6 @@
+guess_csv_delimiter
+===================
+
+.. currentmodule:: ssapy.io
+
+.. autofunction:: guess_csv_delimiter
