### 2025-06-09 6a59261
normalized moon_normal_vector (#84)
Co-authored-by: yeager7 <yeager7@llnl.gov>
