### 2025-06-09 LLNL/SSAPy#81 @SuperdoerTrav
moon_normal_vector has been updated, looking into sun_shine
