### 2025-06-09 LLNL/SSAPy#81 @SuperdoerTrav
import numpy as np
import matplotlib.pyplot as plt
from ssapy import *
import yeager_utils as ut

# Time array over 2 months (1-day spacing)
t_arr = Time("2024-01-01") + np.arange(0, 60) * u.day

# Positions of Moon, Earth (origin), and Sun
r_sat = get_body("moon").position(t_arr).T  # shape (n, 3)
r_earth = np.zeros_like(r_sat)              # Earth at origin
r_sun = get_body("sun").position(t_arr).T   # shape (n, 3)

# Compute reflected flux
sun_shine_dict = compute.sun_shine(
    r_sat, r_earth, r_sun,
    radius=ut.MOON_RADIUS,
    albedo=0.07,                # Use geometric albedo instead of 'albedo'
    albedo_front=0.0,          # Moon has no panel front
    area_panels=0              # No panel area
)

# Apparent magnitude calculation from reflected flux
sun_Mag = 4.80  # absolute magnitude of Sun
r_sun_sat = np.linalg.norm(r_sat - r_sun, axis=-1)
d_pc = u.Unit("parsec").to("m")  # parsec in meters

# Compute Moon's apparent magnitude from reflected flux
M_v = sun_Mag + 5 * np.log10(r_sun_sat / (10 * d_pc)) - 2.5 * np.log10(sun_shine_dict['sun_bus'])

# Plot
plt.figure(figsize=(10, 5))
plt.plot(t_arr.datetime, M_v, marker='o', linestyle='-', label='Moon apparent magnitude')
plt.gca().invert_yaxis()
plt.xlabel("Date")
plt.ylabel("Apparent Magnitude (V)")
plt.title("Moon Apparent Magnitude from Reflected Sunlight (Over 2 Months)")
plt.grid(True)
plt.legend()
plt.tight_layout()
plt.show()
