### 2025-06-09 78c3f5c
updated _shine doc strings to reference fractonal solar fluxes instead of absolute fluxes.
