+++ ssapy/compute.py
@@ -1401,7 +1401,8 @@ def moon_shine(r_moon, r_sat, r_earth, r_sun, radius, albedo, albedo_moon, albed
     """
     Calculate the fractional flux of sunlight reflected from the Moon to the satellite.
 
-    This function computes the flux of sunlight reflected from the Moon to the satellite, including contributions from both the front and back surfaces of the satellite's solar panels.
+    This function estimates the *fraction* of the Sun’s total flux that is first reflected by the Moon
+    and then reflected again toward the observer by the satellite’s body and solar panels (both front and back surfaces).
 
     Parameters:
     ----------
@@ -1414,29 +1415,30 @@ def moon_shine(r_moon, r_sat, r_earth, r_sun, radius, albedo, albedo_moon, albed
     r_sun : (n, 3) numpy.ndarray
         Array of coordinates representing the position of the Sun.
     radius : float
-        Radius of the satellite in meters.
+        Radius of the satellite’s body in meters.
     albedo : float
-        Albedo of the satellite's surface.
+        Albedo of the satellite’s surface.
     albedo_moon : float
         Albedo of the Moon.
     albedo_back : float
-        Albedo of the back surface of the satellite's solar panels.
+        Albedo of the back surface of the satellite’s solar panels.
     albedo_front : float
-        Albedo of the front surface of the satellite's solar panels.
+        Albedo of the front surface of the satellite’s solar panels.
     area_panels : float
-        Area of the satellite's solar panels in square meters.
+        Area of the satellite’s solar panels in square meters.
 
     Returns:
     -------
     dict
-        Dictionary containing the flux contributions from the Moon to the satellite:
-        - 'moon_bus': Fraction of light reflected off the satellite's bus from the Moon.
-        - 'moon_panels': Fraction of light reflected off the satellite's panels from the Moon.
+        Dictionary containing *fractional* flux contributions from Moon-reflected sunlight:
+        - 'moon_bus': Fraction of light reflected off the satellite’s bus via Moon.
+        - 'moon_panels': Fraction of light reflected off the satellite’s panels via Moon.
 
     Notes:
     ------
-    - The function assumes that the solar panels are always facing the Sun and calculates flux based on the phase angles.
-    - Flux contributions from both the front and back surfaces of the solar panels are computed.
+    - Returns fractional (unitless) flux values — not W/m².
+    - Assumes the front of the panels face the Sun and the back receives secondary reflections.
+    - Both front and back surfaces may contribute to observed brightness depending on geometry.
 
     Example:
     --------
@@ -1444,7 +1446,8 @@ def moon_shine(r_moon, r_sat, r_earth, r_sun, radius, albedo, albedo_moon, albed
     >>> r_sat = np.array([[1e7, 1e7, 1e7]])
     >>> r_earth = np.array([[0, 0, 0]])
     >>> r_sun = np.array([[1e11, 0, 0]])
-    >>> moon_shine(r_moon, r_sat, r_earth, r_sun, radius=0.4, albedo=0.20, albedo_moon=0.12, albedo_back=0.50, albedo_front=0.05, area_panels=100)
+    >>> moon_shine(r_moon, r_sat, r_earth, r_sun, radius=0.4, albedo=0.20, albedo_moon=0.12,
+    ...            albedo_back=0.50, albedo_front=0.05, area_panels=100)
     {'moon_bus': array([...]), 'moon_panels': array([...])}
     """
     # https://amostech.com/TechnicalPapers/2013/POSTER/COGNION.pdf
@@ -1468,22 +1471,23 @@ def earth_shine(r_sat, r_earth, r_sun, radius, albedo, albedo_earth, albedo_back
     """
     Calculate the fractional flux of sunlight reflected from the Earth to the satellite.
 
[…]
