### 2025-06-09 0c86000
updated _shine doc strings
