+++ ssapy/compute.py
@@ -1399,10 +1399,9 @@ def calculate_orbital_elements(r_, v_, mu_barycenter=EARTH_MU):
 
 def moon_shine(r_moon, r_sat, r_earth, r_sun, radius, albedo, albedo_moon, albedo_back, albedo_front, area_panels):  # In SI units, takes single values or arrays returns a fractional flux
     """
-    Calculate the fractional flux of sunlight reflected from the Moon to the satellite.
+    Calculate the fractional solar flux reflected from the Moon to the satellite.
 
-    This function estimates the *fraction* of the Sun’s total flux that is first reflected by the Moon
-    and then reflected again toward the observer by the satellite’s body and solar panels (both front and back surfaces).
+    This function computes the fraction of solar flux reflected from the Moon to the satellite, including contributions from both the front and back surfaces of the satellite's solar panels.
 
     Parameters:
     ----------
@@ -1415,30 +1414,29 @@ def moon_shine(r_moon, r_sat, r_earth, r_sun, radius, albedo, albedo_moon, albed
     r_sun : (n, 3) numpy.ndarray
         Array of coordinates representing the position of the Sun.
     radius : float
-        Radius of the satellite’s body in meters.
+        Radius of the satellite in meters.
     albedo : float
-        Albedo of the satellite’s surface.
+        Albedo of the satellite's surface.
     albedo_moon : float
         Albedo of the Moon.
     albedo_back : float
-        Albedo of the back surface of the satellite’s solar panels.
+        Albedo of the back surface of the satellite's solar panels.
     albedo_front : float
-        Albedo of the front surface of the satellite’s solar panels.
+        Albedo of the front surface of the satellite's solar panels.
     area_panels : float
-        Area of the satellite’s solar panels in square meters.
+        Area of the satellite's solar panels in square meters.
 
     Returns:
     -------
     dict
-        Dictionary containing *fractional* flux contributions from Moon-reflected sunlight:
-        - 'moon_bus': Fraction of light reflected off the satellite’s bus via Moon.
-        - 'moon_panels': Fraction of light reflected off the satellite’s panels via Moon.
+        Dictionary containing the flux contributions from the Moon to the satellite:
+        - 'moon_bus': Fraction of solar flux reflected off the satellite's bus from the Moon.
+        - 'moon_panels': Fraction of solar flux reflected off the satellite's panels from the Moon.
 
     Notes:
     ------
-    - Returns fractional (unitless) flux values — not W/m².
-    - Assumes the front of the panels face the Sun and the back receives secondary reflections.
-    - Both front and back surfaces may contribute to observed brightness depending on geometry.
+    - The function assumes that the solar panels are always facing the Sun and calculates flux based on the phase angles.
+    - Fraction of solar flux contributions from both the front and back surfaces of the solar panels are computed.
 
     Example:
     --------
@@ -1446,8 +1444,7 @@ def moon_shine(r_moon, r_sat, r_earth, r_sun, radius, albedo, albedo_moon, albed
     >>> r_sat = np.array([[1e7, 1e7, 1e7]])
     >>> r_earth = np.array([[0, 0, 0]])
     >>> r_sun = np.array([[1e11, 0, 0]])
-    >>> moon_shine(r_moon, r_sat, r_earth, r_sun, radius=0.4, albedo=0.20, albedo_moon=0.12,
-    ...            albedo_back=0.50, albedo_front=0.05, area_panels=100)
+    >>> moon_shine(r_moon, r_sat, r_earth, r_sun, radius=0.4, albedo=0.20, albedo_moon=0.12, albedo_back=0.50, albedo_front=0.05, area_panels=100)
     {'moon_bus': array([...]), 'moon_panels': array([...])}
     """
[…]
