### 2025-06-09 52211e5
normalize moon_normal_vector
