### 2025-06-09 a67d51d
Fix/sun shine (#85)
* updated _shine doc strings to reference fractonal solar fluxes instead of absolute fluxes.

---------

Co-authored-by: yeager7 <yeager7@llnl.gov>
