+++ _sources/api/ssapy.compute.radecRate.rst.txt
@@ -1,0 +1,6 @@
+radecRate
+=========
+
+.. currentmodule:: ssapy.compute
+
+.. autofunction:: radecRate
