+++ _sources/api/ssapy.simple.best_prop.rst.txt
@@ -1,0 +1,6 @@
+best_prop
+=========
+
+.. currentmodule:: ssapy.simple
+
+.. autofunction:: best_prop
