+++ _sources/api/ssapy.io.b3obs2pos.rst.txt
@@ -1,0 +1,6 @@
+b3obs2pos
+=========
+
+.. currentmodule:: ssapy.io
+
+.. autofunction:: b3obs2pos
