+++ _sources/api/ssapy.compute.refine_pass.rst.txt
@@ -1,0 +1,6 @@
+refine_pass
+===========
+
+.. currentmodule:: ssapy.compute
+
+.. autofunction:: refine_pass
