+++ _sources/api/ssapy.plotUtils.koe_plot.rst.txt
@@ -1,0 +1,6 @@
+koe_plot
+========
+
+.. currentmodule:: ssapy.plotUtils
+
+.. autofunction:: koe_plot
