+++ _sources/api/ssapy.utils.get_angle.rst.txt
@@ -1,0 +1,6 @@
+get_angle
+=========
+
+.. currentmodule:: ssapy.utils
+
+.. autofunction:: get_angle
