+++ _sources/api/ssapy.io.save_h5.rst.txt
@@ -1,0 +1,6 @@
+save_h5
+=======
+
+.. currentmodule:: ssapy.io
+
+.. autofunction:: save_h5
