+++ _sources/api/ssapy.io.mkdir.rst.txt
@@ -1,0 +1,6 @@
+mkdir
+=====
+
+.. currentmodule:: ssapy.io
+
+.. autofunction:: mkdir
