+++ _sources/api/ssapy.utils.ecliptic_xyz_to_equatorial.rst.txt
@@ -1,0 +1,6 @@
+ecliptic_xyz_to_equatorial
+==========================
+
+.. currentmodule:: ssapy.utils
+
+.. autofunction:: ecliptic_xyz_to_equatorial
