+++ _sources/api/ssapy.compute.lunar_lagrange_points.rst.txt
@@ -1,0 +1,6 @@
+lunar_lagrange_points
+=====================
+
+.. currentmodule:: ssapy.compute
+
+.. autofunction:: lunar_lagrange_points
