### 2025-06-23 #8147 @astrophpeter
Hi @fruzsinaagocs,

Yes, we are making progress.
1. Issue 81, raised by @andrea-sesta, has been addressed in series of PRs mentioned in  https://github.com/LLNL/SSAPy/issues/81. @andrea-sesta, please confirm that the fixes resolves your issue.
2. Issue 80, raised by @andrea-sesta, has been addressed in https://github.com/LLNL/SSAPy/pull/88. @andrea-sesta, please confirm that the fix resolves your issue. If so, I will close it and the merge the PR soon.
3. Issue 79, raised by @andrea-sesta, has been addressed in https://github.com/LLNL/SSAPy/pull/83. @andrea-sesta, please confirm that the fix resolves your issue. If so, I will close it and the merge the PR soon.
4. Regarding the documentation issue, we have now added some examples. Please see https://software.llnl.gov/SSAPy/examples.html. @athulpg007, can you confirm if this example is sufficient, or suggest additional examples if needed?
