### 2025-06-27 #8147 @athulpg007
re 4: the added example in documentation, yes this serves as a minimum working example.
[optional] If able to, please consider adding 1-2 more illustrative examples to highlight the use of SSAPy for potential users.
