+++ JOSS/paper.bib
@@ -402,3 +402,57 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2023amos.conf..208Y},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
+@software{OREKIT_2024,
+  author       = {maisonobe and
+                  Bryan Cazabonne and
+                  Evan Ward and
+                  Romain Serra and
+                  Maxime Journot and
+                  Sébastien Dinot and
+                  Guilhem Bonnefille and
+                  Thomas Neidhart and
+                  Luc Maisonobe and
+                  Clément Jonglez and
+                  Mark Rutten and
+                  yjeand and
+                  Julio Hernanz and
+                  lirw1984 and
+                  Vincent Cucchietti and
+                  Andrew Goetz and
+                  Guiuux and
+                  gaetanpierre0 and
+                  Lars Næsbye Christensen and
+                  Alberto Fossà and
+                  jvalet and
+                  Alberto Ferrero and
+                  gabb5 and
+                  liscju and
+                  Christopher Schank and
+                  Tanner Mills and
+                  plan3d and
+                  Vyom Yadav and
+                  Shiva Iyer and
+                  Petrus Hyvönen},
+  title        = {CS-SI/Orekit: 12.2},
+  month        = oct,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {12.2},
+  doi          = {10.5281/zenodo.13950582},
+  url          = {https://doi.org/10.5281/zenodo.13950582},
+}
+
+@inproceedings{TUDAT,
+title = "TUDAT: a modular and robust astrodynamics toolbox",
+keywords = "Conf.proc. > 3 pag",
+author = "K Kumar and {van Barneveld}, PWL and D Dirkx and JCP Melman and E Mooij and R Noomen",
+year = "2012",
+language = "English",
+publisher = "ESA",
+pages = "1--8",
+editor = "A Benoit",
+booktitle = "Conference proceeding of the 5th ICATT Conference",
+note = "5th ICATT Conference ; Conference date: 29-05-2012 Through 01-06-2012",
+}
