### 2025-07-02 #8147 @astrophpeter
Thanks both. @athulpg007 we just added some more examples. @andrea-sesta we're working on updating the JOSS paper text again. Could you please let me know of any more outstanding items we need to take care of for the review?
