+++ _sources/api/ssapy.compute.earth_shine.rst.txt
@@ -1,0 +1,6 @@
+earth_shine
+===========
+
+.. currentmodule:: ssapy.compute
+
+.. autofunction:: earth_shine
