+++ _sources/api/ssapy.utils.horizontal_to_equatorial.rst.txt
@@ -1,0 +1,6 @@
+horizontal_to_equatorial
+========================
+
+.. currentmodule:: ssapy.utils
+
+.. autofunction:: horizontal_to_equatorial
