+++ _sources/api/ssapy.utils.get_times.rst.txt
@@ -1,0 +1,6 @@
+get_times
+=========
+
+.. currentmodule:: ssapy.utils
+
+.. autofunction:: get_times
