### 2025-07-07 #8147 @andrea-sesta
@astrophpeter — Thanks for addressing my comments.

@fruzsinaagocs — Once the JOSS paper has been updated, my review will be complete.
