+++ JOSS/paper.bib
@@ -456,3 +456,26 @@ editor = "A Benoit",
 booktitle = "Conference proceeding of the 5th ICATT Conference",
 note = "5th ICATT Conference ; Conference date: 29-05-2012 Through 01-06-2012",
 }
+
+@article{earthmodel, 
+author = {Pavlis, Nikolaos K. and Holmes, Simon A. and Kenyon, Steve C. and Factor, John K.}, 
+title = {The development and evaluation of the Earth Gravitational Model 2008 (EGM2008)}, 
+journal = {Journal of Geophysical Research: Solid Earth}, volume = {117}, number = {B4}, pages = {}, 
+keywords = {gravitational, harmonics, model, spherical}, 
+doi = {10.1029/2011JB008916}, 
+url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2011JB008916}, 
+eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2011JB008916}, 
+year = {2012} 
+} 
+ 
+@article{lunarmodel, 
+author = {Lemoine, Frank G. and Goossens, Sander and Sabaka, Terence J. and Nicholas, Joseph B. and Mazarico, Erwan and Rowlands, David D. and Loomis, Bryant D. and Chinn, Douglas S. and Caprette, Douglas S. and Neumann, Gregory A. and Smith, David E. and Zuber, Maria T.}, 
+title = {High‒degree gravity models from GRAIL primary mission data}, 
+journal = {Journal of Geophysical Research: Planets}, volume = {118}, number = {8}, pages = {1676-1698}, 
+keywords = {lunar gravity, orbit determination, KBRR data, planetary geodesy}, 
+doi = {10.1002/jgre.20118}, 
+url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1002/jgre.20118}, 
+eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1002/jgre.20118}, 
+year = {2013} 
+}
+ 
