### 2025-07-08 87880af
minor grammatical fixes
