### 2025-07-08 75b04a3
fix minor grammatical issues in the JOSS paper (#91)
* updated paper

* minor grammatical fixes
