+++ JOSS/paper.bib
@@ -405,7 +405,7 @@ archivePrefix = {arXiv},
 
 
 @software{OREKIT_2024,
-  author       = {maisonobe and
+  author       = {Maisonobe and
                   Bryan Cazabonne and
                   Evan Ward and
                   Romain Serra and
