+++ JOSS/paper.md
@@ -117,7 +117,6 @@ cislunar dynamics such as Earth (EGM2008;@earthmodel) and Lunar (GRGM1200A;@luna
 SSAPy has utilities for determining from any location on Earth, on-sky brightness, proper motion, right ascension and declination
 and provides conversions between on sky coordinates, TLEs, the Geocentric Celestial Reference Frame and other commonly used coordinates. 
 There are also built in observation linking tools and orbit refinement. SSAPy, with its full-featured modeling framework and scalable, parallelizable
-
 functionality, fills the gap in the orbital software landscape.
 
 
