+++ JOSS/paper.md
@@ -81,26 +81,26 @@ SSAPy is a fast and flexible orbit modeling and analysis tool for orbits spannin
 low-Earth into the cislunar regime. Orbits can be flexibly specified from common
 input formats such as Keplerian elements or two-line
 element (TLE) data files. SSAPy allows users to model satellites and specify parameters such
-as satellite area, mass, and drag coefficients. SSAPy includes a customizable force propagation
+as satellite area, mass, and drag coefficients. SSAPy includes a customizable force-propagation
 with a range of Earth, Lunar, radiation, atmospheric, and maneuvering models. SSAPy makes
 use of various community integration methods and can calculate
 time-evolved orbital quantities, including satellite magnitudes and state vectors.
 Users can specify various space- and ground-based observation models with support for
 multiple coordinate and reference frames. SSAPy also supports orbit analysis and
 propagation methods such as multiple hypothesis tracking and has built-in uncertainty quantification.
-The majority of SSAPy's methods are vectorized and parallelizable, allowing effective use of
-high performance computer (HPC) systems. Finally, SSAPy has plotting functionality, allowing users to
-visualize orbits and trajectories, an example of which is shown in Figures 1 and 2.
+The majority of SSAPy's methods are vectorized and parallelizable, allowing for effective use of
+high-performance computer (HPC) systems. Finally, SSAPy has plotting functionality, allowing users to
+visualize orbits and trajectories. An example is shown in Figures 1 and 2.
 
 SSAPy has been used for the
-classification of cislunar [@Higgins2024] and closely-spaced [@Pruett2024], orbits as
+classification of cislunar [@Higgins2024] and closely-spaced [@Pruett2024] orbits as
 well as for studying the long-term stability of orbits in cislunar space [@Yeager2023]. SSAPy
 has also been used to build a case study for rare events analysis in the context of satellites
 passing close to each other in space [@Miller2022;@Bernstein2021].
 
 # Statement of need
 
-Cislunar space is a region between earth out to beyond the Moon's orbit that includes the
+Cislunar space is a region between Earth out to beyond the Moon's orbit that includes the
 Lagrange points. This region of space is of growing importance to scientific and other space exploration endeavors [e.g., @Duggan2019].
 Understanding, mapping, and modeling orbits through cislunar space is
 critical to all of these endeavors. The challenge for cislunar orbits is that n-body dynamics (e.g., gravitational forces
@@ -108,19 +108,19 @@ from the Sun, Earth, Moon and other planets) are significant, leading to unpredi
 In this chaotic regime, orbits cannot be reduced to simple parametric descriptions making scalable orbit
 simulation and modeling a critical analysis tool [@Yeager2023]. Current orbit modeling software tools
 are predominantly used via graphical user interfaces (e.g., The General Mission Analysis Tool; @Hughes2014 or the Systems Tool Kit)
-and are not optimized for large scale simulation on HPC systems. Orbital modeling codes that
+and are not optimized for large-scale simulation on HPC systems. Orbital modeling codes that
 can be run on HPC systems (e.g., REBOUND; @Rein2012) lack full observable generation and modeling capabilities
-with uncertainty quantification. Existing space dynamics libraries such as Orekit (@OREKIT_2024) and Tudat (@TUDAT) have 
-comprehensive and many overlapping features with SSAPy. However, one point of difference is that they rely on spherical 
+with uncertainty quantification. Existing space dynamics libraries such as Orekit (@OREKIT_2024) and Tudat (@TUDAT) share many 
+features with SSAPy. However, one point of difference is that they rely on spherical 
[…]
