### 2025-07-10 2b75d6b
fixed more typos in the JOSS paper (#92)
* updated paper

* minor grammatical fixes

* fixed formatting issue

* minor typo correction

* corrected more typos
