### 2025-07-10 #8147 @fruzsinaagocs
@astrophpeter, can you please add the missing DOIs? The editorialbot suggestions seem to match what you're looking for. I checked the "skipped" DOIs, and they seem to be mostly conference posters without DOIs, which is fine. The only exception is your citation for TUDAT, which has more recent release(s) (e.g. [this](https://meetingorganizer.copernicus.org/EPSC2022/EPSC2022-253.html)) with DOIs. It's up to you whether you'd like to cite that instead of or in addition to the original conference proceeding, which has neither text nor a DOI, so imo is a bit of a dead-end for the reader.
