### 2025-07-10 13adad0
update bib DOI for JOSS
