+++ JOSS/paper.bib
@@ -47,7 +47,8 @@ Administration and the United States Space Force",
   author={Vallado, David and Crawford, Paul and Hujsak, Ricahrd and Kelso, TS},
   booktitle={AIAA/AAS Astrodynamics Specialist Conference and Exhibit},
   pages={6753},
-  year={2006}
+  year={2006},
+  doi={10.2514/6.2006-6753}
 }
 
 
@@ -183,7 +184,8 @@ archivePrefix = {arXiv},
   author={Hughes, Steven P and Qureshi, Rizwan H and Cooley, Steven D and Parker, Joel J},
   booktitle={AIAA/AAS astrodynamics specialist conference},
   pages={4151},
-  year={2014}
+  year={2014},
+  doi={10.2514/6.2014-4151}
 }
 
 
@@ -444,17 +446,18 @@ archivePrefix = {arXiv},
   url          = {https://doi.org/10.5281/zenodo.13950582},
 }
 
-@inproceedings{TUDAT,
-title = "TUDAT: a modular and robust astrodynamics toolbox",
-keywords = "Conf.proc. > 3 pag",
-author = "K Kumar and {van Barneveld}, PWL and D Dirkx and JCP Melman and E Mooij and R Noomen",
-year = "2012",
-language = "English",
-publisher = "ESA",
-pages = "1--8",
-editor = "A Benoit",
-booktitle = "Conference proceeding of the 5th ICATT Conference",
-note = "5th ICATT Conference ; Conference date: 29-05-2012 Through 01-06-2012",
+
+@INPROCEEDINGS{TUDAT,
+       author = {{Dirkx}, Dominic and {Fayolle}, Marie and {Garrett}, Geoffrey and {Avillez}, Miguel and {Cowan}, Kevin and {Cowan}, Sean and {Encarnacao}, Joao and {Fortuny Lombrana}, Carlos and {Gaffarel}, J{\'e}r{\'e}mie and {Hener}, Jonas and {Hu}, Xuanyu and {van Nistelrooij}, Maarten and {Oggionni}, Filippo and {Plumaris}, Michael},
+        title = "{The open-source astrodynamics Tudatpy software - overview for planetary mission design and science analysis}",
+    booktitle = {European Planetary Science Congress},
+         year = 2022,
+        month = sep,
+          eid = {EPSC2022-253},
+        pages = {EPSC2022-253},
+          doi = {10.5194/epsc2022-253},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022EPSC...16..253D},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @article{earthmodel, 
