### 2025-07-11 #8147 @astrophpeter
Thanks @fruzsinaagocs I have added the missing DOIs and replaced the Tudat reference. https://github.com/LLNL/SSAPy/pull/93 . Please let me know if you would like me make any further changes.
