+++ _sources/api/ssapy.rvsampler.circular_guess.rst.txt
@@ -1,0 +1,6 @@
+circular_guess
+==============
+
+.. currentmodule:: ssapy.rvsampler
+
+.. autofunction:: circular_guess
