+++ _sources/api/ssapy.utils.unit_vector.rst.txt
@@ -1,0 +1,6 @@
+unit_vector
+===========
+
+.. currentmodule:: ssapy.utils
+
+.. autofunction:: unit_vector
