+++ _sources/api/ssapy.plotUtils.draw_dashed_circle.rst.txt
@@ -1,0 +1,6 @@
+draw_dashed_circle
+==================
+
+.. currentmodule:: ssapy.plotUtils
+
+.. autofunction:: draw_dashed_circle
