+++ _sources/api/ssapy.gravity.AccelThirdBody.rst.txt
@@ -1,0 +1,17 @@
+AccelThirdBody
+==============
+
+.. currentmodule:: ssapy.gravity
+
+.. autoclass:: AccelThirdBody
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~AccelThirdBody.__call__
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: __call__
