+++ _sources/api/ssapy.orbit_solver.TwoPosOrbitSolver.rst.txt
@@ -1,0 +1,17 @@
+TwoPosOrbitSolver
+=================
+
+.. currentmodule:: ssapy.orbit_solver
+
+.. autoclass:: TwoPosOrbitSolver
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~TwoPosOrbitSolver.solve
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: solve
