+++ _sources/api/ssapy.utils.lazy_property.rst.txt
@@ -1,0 +1,7 @@
+lazy_property
+=============
+
+.. currentmodule:: ssapy.utils
+
+.. autoclass:: lazy_property
+   :show-inheritance:
