+++ _sources/api/ssapy.simple.ssapy_prop.rst.txt
@@ -1,0 +1,6 @@
+ssapy_prop
+==========
+
+.. currentmodule:: ssapy.simple
+
+.. autofunction:: ssapy_prop
