+++ _sources/api/ssapy.utils.isAttributeSafeToTransfer.rst.txt
@@ -1,0 +1,6 @@
+isAttributeSafeToTransfer
+=========================
+
+.. currentmodule:: ssapy.utils
+
+.. autofunction:: isAttributeSafeToTransfer
