### 2025-07-17 #8147 @astrophpeter
Thanks @fruzsinaagocs

the new tagged release of the software is v1.1
this version is archived with zenodo here: https://zenodo.org/records/16048969
the zenodo DOI is 10.5281/zenodo.16048969
The title and authors match the JOSS submission
The zenodo license matches the software license (MIT)
