+++ _sources/api/ssapy.utils.sun_ra_dec.rst.txt
@@ -1,0 +1,6 @@
+sun_ra_dec
+==========
+
+.. currentmodule:: ssapy.utils
+
+.. autofunction:: sun_ra_dec
