+++ _sources/api/ssapy.propagator.RK4Propagator.rst.txt
@@ -1,0 +1,7 @@
+RK4Propagator
+=============
+
+.. currentmodule:: ssapy.propagator
+
+.. autoclass:: RK4Propagator
+   :show-inheritance:
