+++ _sources/api/ssapy.compute.calculate_orbital_elements.rst.txt
@@ -1,0 +1,6 @@
+calculate_orbital_elements
+==========================
+
+.. currentmodule:: ssapy.compute
+
+.. autofunction:: calculate_orbital_elements
