+++ _sources/modules/ssapy.compute.rst.txt
@@ -1,0 +1,5 @@
+ssapy.compute
+=============
+
+.. automodapi:: ssapy.compute
+    :no-heading:
