+++ _sources/api/ssapy.rvsampler.APrior.rst.txt
@@ -1,0 +1,17 @@
+APrior
+======
+
+.. currentmodule:: ssapy.rvsampler
+
+.. autoclass:: APrior
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~APrior.__call__
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: __call__
