+++ _sources/api/ssapy.propagator.default_numerical.rst.txt
@@ -1,0 +1,6 @@
+default_numerical
+=================
+
+.. currentmodule:: ssapy.propagator
+
+.. autofunction:: default_numerical
