+++ _sources/api/ssapy.utils.catalog_to_apparent.rst.txt
@@ -1,0 +1,6 @@
+catalog_to_apparent
+===================
+
+.. currentmodule:: ssapy.utils
+
+.. autofunction:: catalog_to_apparent
