+++ _sources/api/ssapy.correlate_tracks.iterate_mht.rst.txt
@@ -1,0 +1,6 @@
+iterate_mht
+===========
+
+.. currentmodule:: ssapy.correlate_tracks
+
+.. autofunction:: iterate_mht
