### 2025-07-27 892bcb6
Initial editorial tweaks for JOSS publication
