+++ JOSS/paper.bib
@@ -1,13 +1,14 @@
 @techreport{Bernstein2021,
   author       = {Bernstein, Jason and Filippov, Andrey and Schneider, Michael and Miller, Caleb},
-  title        = {Quantifying Uncertainty in All-to-All Estimates of Space Object Conjunction Probabilities using U-Statistics},
+  title        = {Quantifying Uncertainty in All-to-All Estimates of Space Object Conjunction Probabilities using {U}-Statistics},
   institution  = {Lawrence Livermore National Lab. (LLNL), Livermore, CA (United States)},
   annote       = {Predicting space object conjunctions is inherently probabilistic due to initial state and orbit model uncertainty. A commonly considered Monte Carlo estimator of the conjunction probability is the ’all-to-all’ estimator. Given independent random samples of the trajectories of both objects, the estimator is the percentage of all pairs of trajectories that result in a conjunction. Intuitively, the all-to-all estimator is the best possible estimator of the conjunction probability since it considers all pairs of Monte Carlo samples. However, its distribution is not available in closed-form, which limits its use in practice and makes this intuition difficult to make rigorous. In this paper, the all-to-all estimator is identified as a U-statistic, which implies that it has several favorable properties. Specifically, the estimator is the minimum variance unbiased estimator of the conjunction probability and is asymptotically Gaussian distributed. An approximate confidence interval for the conjunction probability is obtained from an estimate of the asymptotic Gaussian distribution. We show how to efficiently compute the confidence interval and demonstrate that the interval has the nominal coverage level. The confidence intervals are also seen to be narrower than those based on the commonly-used each-to-each estimator. Furthermore, the all-to-all estimator is shown to allow different Monte Carlo sample sizes, whereas the each-to-each estimator requires equal sample sizes.},
   doi          = {10.2172/1825370},
   url          = {https://www.osti.gov/biblio/1825370},
   place        = {United States},
   year         = {2021},
-  month        = {10}}
+  month        = {10}
+}
 
 @misc{MOU,
   author       = "",
@@ -21,36 +22,25 @@ Administration and the United States Space Force",
 }
 
 @INPROCEEDINGS{Duggan2019,
-
   author={Duggan, Matthew and Simon, Xavier and Moseman, Travis},
-
-  booktitle={2019 IEEE Aerospace Conference}, 
-
-  title={Lander and Cislunar Gateway Architecture Concepts for Lunar Exploration}, 
-
+  booktitle={2019 IEEE Aerospace Conference},
+  title={Lander and Cislunar Gateway Architecture Concepts for Lunar Exploration},
   year={2019},
-
   volume={},
-
   number={},
-
   pages={1-9},
-
   keywords={Logic gates;Moon;Space vehicles;Deep-space communications;Orbits;Propulsion;Robots},
-
-  doi={10.1109/AERO.2019.8741766}}
-
+  doi={10.1109/AERO.2019.8741766}
 }
 
 @inproceedings{Vallado2006,
[…]
