### 2025-07-28 8200c9e
Initial editorial tweaks for JOSS publication (#94)
