### 2025-07-28 #8147 @astrophpeter
Thanks @warrickball we have merged your changes.
