+++ .github/workflows/python-publish.yml
@@ -37,8 +37,6 @@ jobs:
           python -m pip install --upgrade pip setuptools flake8 build
           python -m pip install -r requirements.txt
           python3 -m build
-          mkdir -p dist
-          cp dist/*.whl dist/
 
       - name: Install and build (macOS)
         if: matrix.os == 'macOS-latest'
@@ -55,7 +53,7 @@ jobs:
           export CIBW_BUILD="${PYTHON_VERSION}-macosx_*"
           python3 -m cibuildwheel --platform macos
           mkdir -p dist
-          cp wheelhouse/*.whl dist/
+          cp -n wheelhouse/*.whl dist/
 
       - name: Upload dist artifacts
         uses: actions/upload-artifact@v4
