+++ .github/workflows/ci.yml
@@ -1,5 +1,7 @@
 name: CI/CD
 
+permissions:
+  contents: write
 on:
   # This Workflow can be triggered manually
   workflow_dispatch:
