### 2025-07-28 #8147 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
